### 2021-05-29 #3319 @whedon
[PRE REVIEW]: oce: an R package for Oceanographic Analysis
**Submitting author:** @dankelley (<a href="http://orcid.org/0000-0001-7808-5911">Dan E. Kelley</a>)
**Repository:** <a href="https://github.com/dankelley/oce" target ="_blank">https://github.com/dankelley/oce</a>
**Version:** 1.5-0
