### 2021-05-29 48450f6
README.md gets a badge for JOSS submission
