### 2021-05-29 #3319 @dankelley
Dear editor,

As a response to the first comment in this issue, we are putting below a list of three possible reviewers, with GitHub username followed by full name.  We can supply more names, if needed.  In selecting these names, we have tried to span a range of researchers at different stages of their careers, and with different backgrounds.

Please note that we examined the spreadsheet of those who have agreed to review for JOSS, but found none with both R experience and oceanographic experience.  This is why we widened our search.  With any luck, our suggestions for reviewers might lead to more entries in your list of potential reviewers.

1. erinann Erin LaBrecque

2. SeascapeScience Nick Record

3. gonzalobravoargentina Gonzalo Bravo
