### 2021-05-29 9fabe82
cite Hornik 2012 for CRAN (as JOSS bot suggests)
