+++ inst/joss/paper.bib
@@ -1,3 +1,19 @@
+@article{hornik_comprehensive_2012,
+	title = {The {Comprehensive} {R} {Archive} {Network}},
+	volume = {4},
+	issn = {1939-5108, 1939-0068},
+	url = {https://onlinelibrary.wiley.com/doi/10.1002/wics.1212},
+	doi = {10.1002/wics.1212},
+	language = {en},
+	number = {4},
+	urldate = {2021-05-29},
+	journal = {WIREs Computational Statistics},
+	author = {Hornik, Kurt},
+	month = jul,
+	year = {2012},
+	pages = {394--398}
+}
+
 @misc{noauthor_comprehensive_2021,
 	title = {The {Comprehensive} {R} {Archive} {Network}},
 	url = {https://cran.r-project.org/},
