### 2021-05-29 #3319 @dankelley
Dear editor, a citation for Hornik's paper on CRAN has been added, as suggested by the JOSS bot. This is my first experience with that bot, and so I am writing this note instead of just the next line, to update the submission. I hope that's within protocol.  Dan.
