### 2021-05-29 #3319 @arfon
☝️ @dankelley – the commands to Whedon need to be the first line of a message but it's encouraged for authors to interact with the bot too!
