### 2021-06-02 775717c
remove context() from testthat tests
