+++ tests/testthat/test_local_gps.R
@@ -1,5 +1,4 @@
 library(oce)
-context("gps")
 test_that("gps", {
           if (1 == length(list.files(path=".", pattern="local_data"))) {
               g <- read.oce("local_data/test_trk2.gpx") # test file from rgdal package
