### 2021-06-11 6413f04
vectorShow() gets new arg showNA
