### 2021-06-12 d08fce1
make subset(adp,pressure...) work for sentinel-V
More testing is required, on a sentinel-V dataset.  See
https://github.com/dankelley/oce/issues/1837

ut we still nee
