### 2021-06-16 f0970bd
minor fixes
