+++ R/adp.R
@@ -446,11 +446,13 @@ setMethod(f="summary",
                             "\n", sep=''))
               }
               v.dim <- dim(object[["v"]])
-              if (!is.ad2cp(object)) {
-                  cat("* Number of profiles:", v.dim[1], "\n")
-                  cat("* Number of cells:   ", v.dim[2], "\n")
-                  cat("* Number of beams:   ", v.dim[3], "\n")
-                  cat("* Cell size:         ", object[["cellSize"]], "m\n")
+              isAD2CP <- is.ad2cp(object)
+              if (!isAD2CP) {
+                  cat("* Number of beams:    ", v.dim[3], "\n", sep="")
+                  cat("* Number of profiles: ", v.dim[1], "\n", sep="")
+                  cat("* Number of cells:    ", v.dim[2], "\n", sep="")
+                  cat("* Number of beams:    ", v.dim[3], "\n", sep="")
+                  cat("* Cell size:          ", object[["cellSize"]], "m\n", sep="")
               }
               if ("time" %in% names(object@data)) {
                   cat("* Summary of times between profiles:\n")
@@ -518,7 +520,6 @@ setMethod(f="summary",
               if ("ensembleNumber" %in% names(object@metadata)) {
                   cat(paste("* Ensemble Numbers:  ", vectorShow(object@metadata$ensembleNumber, msg="")))
               }
-              isAD2CP <- is.ad2cp(object)
               if (!isAD2CP) {
                   if ("numberOfCells" %in% metadataNames) {
                       dist <- object[["distance"]]
@@ -548,7 +549,11 @@ setMethod(f="summary",
                   }
                   beamUnspreaded <- object[["oceBeamUnspreaded"]]
                   cat("* Beams::\n")
-                  cat("    Number:          ", if (is.null(numberOfBeams)) "?" else numberOfBeams, "\n")
+                  if ("vv" %in% names(object@data))
+                      cat("    Number:          ", if (is.null(numberOfBeams)) "?" else numberOfBeams, "slanted, plus 1 central\n")
+                  else
+                      cat("    Number:          ", if (is.null(numberOfBeams)) "?" else numberOfBeams, "slanted\n")
+
                   cat("    Slantwise Angle: ", if (is.null(beamAngle)) "?" else beamAngle , "\n")
                   if (numberOfBeams > 0)
                       cat("    Orientation:     ", if (is.null(orientation)) "?" else orientation, "\n")
@@ -1057,23 +1062,22 @@ setMethod(f="subset",
                   }
               } else if (length(grep("distance", subsetString))) {
                   oceDebug(debug, "subsetting an adp by distance\n")
-                  if (length(grep("time", subsetString)))
-                      stop("cannot subset by both time and distance; split into multiple calls")
+                  if (grepl("time|pressure", subsetString))
+                      stop("cannot subset by both bin (e.g. distance) and profile (i.e. time or pressure)")
                   ## keep <- eval(substitute(subset), x@data, parent.frame(2))
                   keep <- eval(expr=substitute(expr=subset, env=environment()), envir=x@data, enclos=parent.frame(2))
-                  oceDebug(debug, vectorShow(keep, "keeping ordinary bins:"), "\n")
+                  oceDebug(debug, "subset() will retain", sum(keep), "of", length(keep), "bins in slant beams\n")
                   haveVerticalBeam <- "vv" %in% names(x@data) # assume, later, that va, vg, vq and vdistance exist
                   if (haveVerticalBeam) {
                       oceDebug(debug, "have a vertical beam\n")
                       vkeep <- eval(expr=substitute(expr=subset,env=environment()),envir=list(distance=x@data$vdistance),enclos=parent.frame(2))
-                      oceDebug(debug, vectorShow(vkeep, "keeping vertical-beam bins:"), "\n")
[…]
