### 2021-06-16 e929a11
test local RDI-teledyne sentinel-V adp file
