### 2021-06-24 #3319 @kthyng
Hi @dankelley and welcome to JOSS! I am happy to see your submission! Sorry for the delay, I am handling a lot of papers right now and had some summer travel too. Let's see if we can get some reviewers. Thanks for the list — I'll start from there (esp since I don't know so many R people).
