### 2021-06-24 #3319 @kthyng
@erinann and @gonzalobravoargentina — Are you interested in reviewing this submission for JOSS, the Journal of Open Source Software? This is an open review which takes place on github with issues. You can learn more about the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. I would hope you could start the review in the next few weeks with the intention of being through the typically iterative process in 4-6 weeks. Thanks for your consideration.
