### 2021-06-26 a93dd9d
improve pwelch() for issue 1839
See
https://github.com/dankelley/oce/issues/1839
and
https://github.com/dankelley/oce-issues/tree/main/18xx/1839
