+++ .github/workflows/auto-assign-to-project.yml
@@ -18,5 +18,5 @@ jobs:
       if: github.event.action == 'opened'
       with:
         project: 'https://github.com/dankelley/oce/projects/7'
-        column_name: 'To do'
+        column_name: 'Needs triage'
 
