### 2021-06-29 #3319 @gonzalobravoargentina
Hi @kthyng,
I'm very honoured with the invitation.
Although I am intermediate R used, I have used the Oce package rather well and think I can give some humble contributions. However if others reviewers with more programing experience are available I got not problem in leaving my place.
Thanks
