+++ tests/testthat.R
@@ -1,3 +1,4 @@
+library(testthat)
 library(oce)
 
 test_check("oce")
