+++ tests/testthat/test_curl.R
@@ -6,9 +6,9 @@ test_that("curl 1", {
           taux <- matrix(c(1,6,9,3,7,11,14,10,4,13,2,12,16,8,15,5), nrow=4, ncol=4, byrow=TRUE)
           tauy <- matrix(c(3,20,1,14,1,19,2,6,21,13,28,16,24,4,15,17), nrow=4, ncol=4, byrow=TRUE)
           C <- curl(u=taux, v=tauy, x=lon, y=lat, geographical=TRUE, method=2)
-          expect_equal(C$curl[1,1], -3.338849e-05, tolerance=0.000001e-5, scale=1)
-          expect_equal(C$x[1], 161.25, tolerance=0.00001, scale=1)
-          expect_equal(C$y[1], 83.75, tolerance=0.00001, scale=1)
+          expect_equal(C$curl[1,1], -3.338849e-05, tolerance=0.000001e-5)
+          expect_equal(C$x[1], 161.25, tolerance=0.00001)
+          expect_equal(C$y[1], 83.75, tolerance=0.00001)
 })
 
 test_that("curl 2", {
