+++ R/astronomy.R
@@ -233,9 +233,10 @@ siderealTime <- function(t)
 #' Richmond, Virginia, USA: Willmann-Bell, 1982.
 #'
 #' @examples
+#' library(oce)
 #' ## example from Meeus
 #' t <- ISOdatetime(1977, 4, 26, hour=0, min=0, sec=0, tz="UTC")+0.4*86400
-#' expect_equal(julianDay(t), 2443259.9)
+#' stopifnot(all.equal(julianDay(t), 2443259.9))
 #'
 #' @family things related to astronomy
 #' @family things related to time
