+++ tests/testthat/test_sw.R
@@ -64,12 +64,12 @@ test_that("rho and sigma", {
           sigma <- swSigma(SP, t, p, longitude, latitude, "gsw")
           expect_equal(rhoGSW-1000, sigma)
           ## The following was hard-coded using values from GSW3.03, and it failed with GSW3.05.
-          ##expect_equal(30.823971,swSigma(35,13,1000,eos="gsw"),scale=1,tolerance=0.000001)
-          expect_equal(30.818302,swSigma(35,T90fromT68(13),1000,eos="unesco"),scale=1,tolerance=0.000001)
+          ##expect_equal(30.823971,swSigma(35,13,1000,eos="gsw"),tolerance=0.000001)
+          expect_equal(30.818302,swSigma(35,T90fromT68(13),1000,eos="unesco"),tolerance=0.000001)
           ## The sigmaT tests are not from definititive test values, and so are just
           ## checks against future changes.
-          ##expect_equal(swSigmaT(35, 13, 1000, eos="gsw"), 26.396816, scale=1, tolerance=0.000001)
-          expect_equal(swSigmaT(35, T90fromT68(13), 1000, eos="unesco"), 26.393538, scale=1, tolerance=0.000001)
+          ##expect_equal(swSigmaT(35, 13, 1000, eos="gsw"), 26.396816,tolerance=0.000001)
+          expect_equal(swSigmaT(35, T90fromT68(13), 1000, eos="unesco"), 26.393538,tolerance=0.000001)
 })
 
 test_that("potential_temperature (UNESCO)", {
@@ -79,9 +79,9 @@ test_that("potential_temperature (UNESCO)", {
           ## second with a ctd object as an arg. Note the need to convert to the 1968
           ## temperature scale, which was used in the UNESCO formulation.
           expect_equal(swTheta(40, T90fromT68(40), 10000, eos="unesco"),
-                       T90fromT68(36.89073), scale=1, tolerance=0.00002)
+                       T90fromT68(36.89073),tolerance=0.00002)
           expect_equal(swTheta(as.ctd(40, T90fromT68(40), 10000), eos="unesco"),
-                       T90fromT68(36.89073), scale=1, tolerance=0.00002)
+                       T90fromT68(36.89073),tolerance=0.00002)
           ## Test self-consistency at the surface (also a test of vector reframing)
           T <- 10 + rnorm(50)
           expect_equal(0, sum(abs(T - swTheta(rep(35, 50), T, 0, eos="unesco"))))
@@ -106,8 +106,8 @@ test_that("SA and CT, sound speed (GSW)", {
           expect_equal(SA, swAbsoluteSalinity(ctd))
           expect_equal(CT, swConservativeTemperature(salinity=SP, temperature=t, pressure=p, longitude=lon, latitude=lat))
           expect_equal(CT, swConservativeTemperature(ctd))
-          expect_equal(1731.995, swSoundSpeed(40,T90fromT68(40),1e4,eos="unesco"),scale=1,tolerance=0.001)
-          expect_equal(1731.995, swSoundSpeed(as.ctd(40,T90fromT68(40),1e4),eos="unesco"),scale=1,tolerance=0.001)
+          expect_equal(1731.995, swSoundSpeed(40,T90fromT68(40),1e4,eos="unesco"),tolerance=0.001)
+          expect_equal(1731.995, swSoundSpeed(as.ctd(40,T90fromT68(40),1e4),eos="unesco"),tolerance=0.001)
           SA <- gsw::gsw_SA_from_SP(SP=40, p=1e4, longitude=300, latitude=30)
           CT <- gsw::gsw_CT_from_t(SA, 40, 1e4)
           speedGSW <- gsw::gsw_sound_speed(SA, CT, 1e4)
[…]
