+++ man/swSigmaTheta.Rd
@@ -60,7 +60,7 @@ extracted from it, and used for the calculation instead of any values
 provided in the other arguments.
 }
 \examples{
-expect_equal(26.4212790994, swSigmaTheta(35, 13, 1000, eos="unesco"))
+stopifnot(abs(26.4212790994 - swSigmaTheta(35, 13, 1000, eos="unesco")) < 1e-7)
 
 }
 \references{

[…]
