### 2021-06-30 9c47067
catch msg/output in local_adp test
