+++ tests/testthat/test_local_adp.R
@@ -2,165 +2,187 @@ library(oce)
 
 f <- "~/Dropbox/data/archive/sleiwex/2008/moorings/m07/adp/sontek_h53/raw/adp_sontek_h53.adp"
 if (file.exists(f)) {
-    test_that("Sontek adp", {
-              n <- 5000
-              adp <- read.oce(f, from=n+1, to=n+3)
-              ## numeric time comparison is easier for these whacky times
-              expect_equal(as.numeric(adp[["time"]])-1214438000,
-                           c(0.5599999428, 10.6500000954, 20.6400001049))
+    test_that("Sontek adp",
+              {
+                  n <- 5000
+                  adp <- read.oce(f, from=n+1, to=n+3)
+                  ## numeric time comparison is easier for these whacky times
+                  expect_equal(as.numeric(adp[["time"]])-1214438000,
+                               c(0.5599999428, 10.6500000954, 20.6400001049))
 
-              v <- c(0.51, 0.51, 0.51, 0.756, 0.756, 0.756, 0.17, 0.17, 0.17,
-                     0.1, 0.1, 0.1, 1.7, 1.7, 1.7, 0.227, 0.227, 0.227, 0.05,
-                     0.05, 0.05, 1.5, 1.5, 1.5, 0.123, 0.123, 0.123, 0.123,
-                     0.123, 0.123, 0.123, 0.123, 0.123, 0, 0, 0, 1.114, 1.435,
-                     1.174, 0.477, 0.622, 0.029, -0.052, 0.011, -0.092, 0, 0,
-                     0, 20.554, 19.021, 20.302, 0.074, 0.079, 0.08, -0.978,
-                     -0.686, -1.154, -1.054, -0.764, -1.215, -1.173, -0.872,
-                     1.218, 1.223, -0.939, 1.172, 1.186, -1.005, 1.131, 1.141,
-                     -1.103, 1.107, 1.136, -1.171, 1.063, 1.12, -1.216, 1.024,
-                     1.074, 1.232, 1.041, 1.048, 1.207, 1.061, 1.013, 1.191,
-                     1.047, 0.971, 1.133, 1.052, 0.98, 1.102, 1.023, 0.976,
-                     1.084, 1.026, 0.977, 1.06, 1.017, 0.975, 1.03, 1.029,
-                     0.998, 0.996, 1.001, 1.002, 0.974, 0.975, 0.989, 0.884,
-                     0.947, 0.951, 0.866, 0.887, 0.97, 0.863, 0.855, 0.952,
-                     0.607, 0.831, 0.982, 0.656, 0.803, 0.986, 0.565, 0.749,
-                     0.969, 0.499, 0.713, 0.972, 0.462, 0.714, 0.959, 0.466,
-                     0.683, 0.882, 0.446, 0.607, 0.787, 0.44, 0.562, 0.696,
-                     0.455, 0.514, 0.609, 0.534, 0.46, 0.324, 0.269, 0.229,
-                     0.928, 1.068, 0.335, 0.716, 0.877, 0.222, 0.592, 0.698,
-                     0.194, 0.486, 0.546, 0.126, 0.428, 0.482, 0.068, 0.401,
-                     0.469, 0.024, 0.365, 0.394, 0.043, 0.346, 0.381, 0.002,
-                     0.314, 0.303, -0.055, 0.299, 0.294, -0.075, 0.285, 0.263,
-                     -0.057, 0.271, 0.312, -0.049, 0.237, 0.33, -0.015, 0.201,
-                     0.328, 0.015, 0.224, 0.348, 0.057, 0.234, 0.343, 0.041,
[…]
