### 2021-06-30 896fda4
adapt to new testhat
Some tests such as

  accyBurstMatlab <- c(-0.3144531, ETC
  expect_equal(d1[["accelerometery", "burst"]][1:10], accyBurstMatlab, scale=1, tolerance=1e-6)

started to fail with testthat-3.0.3, and the pattern seems to be that
the last digit is off by a value or two.  But I also notice that
expect_equal() seems to treat the 'scale' differently, and I don't
actually see it in the docs anymore, so I dropped that.

I went through all tests that started making errors and fixed them all
up, usually by changing a tolerance of, say 1e-6 to 3e-6 or so.  That
seems to work in pretty
[…]
