### 2021-06-30 e72c1de
lower non-MacOS testing (hopefully, temporarily)
