### 2021-07-02 50e0550
read.adp.rdi() skips to first 7f7f byte pair
See https://github.com/dankelley/oce/issues/1845
