+++ src/RcppExports.cpp
@@ -292,8 +292,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // do_ldc_rdi_in_file
-List do_ldc_rdi_in_file(StringVector filename, IntegerVector from, IntegerVector to, IntegerVector by, IntegerVector mode, IntegerVector debug);
-RcppExport SEXP _oce_do_ldc_rdi_in_file(SEXP filenameSEXP, SEXP fromSEXP, SEXP toSEXP, SEXP bySEXP, SEXP modeSEXP, SEXP debugSEXP) {
+List do_ldc_rdi_in_file(StringVector filename, IntegerVector from, IntegerVector to, IntegerVector by, IntegerVector startIndex, IntegerVector mode, IntegerVector debug);
+RcppExport SEXP _oce_do_ldc_rdi_in_file(SEXP filenameSEXP, SEXP fromSEXP, SEXP toSEXP, SEXP bySEXP, SEXP startIndexSEXP, SEXP modeSEXP, SEXP debugSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -301,9 +301,10 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< IntegerVector >::type from(fromSEXP);
     Rcpp::traits::input_parameter< IntegerVector >::type to(toSEXP);
     Rcpp::traits::input_parameter< IntegerVector >::type by(bySEXP);
+    Rcpp::traits::input_parameter< IntegerVector >::type startIndex(startIndexSEXP);
     Rcpp::traits::input_parameter< IntegerVector >::type mode(modeSEXP);
     Rcpp::traits::input_parameter< IntegerVector >::type debug(debugSEXP);
-    rcpp_result_gen = Rcpp::wrap(do_ldc_rdi_in_file(filename, from, to, by, mode, debug));
+    rcpp_result_gen = Rcpp::wrap(do_ldc_rdi_in_file(filename, from, to, by, startIndex, mode, debug));
     return rcpp_result_gen;
 END_RCPP
 }
