+++ R/RcppExports.R
@@ -92,8 +92,8 @@ do_ldc_ad2cp_in_file <- function(filename, from, to, by, DEBUG) {
     .Call(`_oce_do_ldc_ad2cp_in_file`, filename, from, to, by, DEBUG)
 }
 
-do_ldc_rdi_in_file <- function(filename, from, to, by, mode, debug) {
-    .Call(`_oce_do_ldc_rdi_in_file`, filename, from, to, by, mode, debug)
+do_ldc_rdi_in_file <- function(filename, from, to, by, startIndex, mode, debug) {
+    .Call(`_oce_do_ldc_rdi_in_file`, filename, from, to, by, startIndex, mode, debug)
 }
 
 do_matrix_smooth <- function(mat) {
