### 2021-07-03 d375c91
remove 4-y old commented-out bisectAdpRdiBuf code
