+++ R/adp.rdi.R
@@ -703,88 +703,6 @@ read.adp.rdi <- function(file, from, to, by, tz=getOption("oceTz"),
     if (!toGiven)
         to <- 0
     profileStart <- NULL # prevent scope warning from rstudio; defined later anyway
-    ## bisectAdpRdiBuf <- function(buf, t.find, add=0, debug=0) {
-    ##     oceDebug(debug, "bisectAdpRdiBuffer(t.find=", format(t.find), ", add=", add, ") {\n", unindent=1)
-    ##     len <- length(profileStart)
-    ##     lower <- 1
-    ##     upper <- len
-    ##     passes <- floor(10 + log(len, 2)) # won't need this many; only do this to catch coding errors
-    ##     for (pass in 1:passes) {
-    ##         middle <- floor( (upper + lower) / 2 )
-    ##         year   <- unabbreviateYear(readBin(buf[profileStart[middle] +  4], what="integer", n=1, size=1, signed=FALSE))
-    ##         month  <- readBin(buf[profileStart[middle] +  5], what="integer", n=1, size=1, signed=FALSE)
-    ##         day    <- readBin(buf[profileStart[middle] +  6], what="integer", n=1, size=1, signed=FALSE)
-    ##         hour   <- readBin(buf[profileStart[middle] +  7], what="integer", n=1, size=1, signed=FALSE)
-    ##         minute <- readBin(buf[profileStart[middle] +  8], what="integer", n=1, size=1, signed=FALSE)
-    ##         second <- readBin(buf[profileStart[middle] +  9], what="integer", n=1, size=1, signed=FALSE)
-    ##         sec100 <- readBin(buf[profileStart[middle] + 10], what="integer", n=1, size=1, signed=FALSE)
-    ##         t <- ISOdatetime(year, month, day, hour, minute, second + sec100/100, tz=tz)
-    ##         oceDebug(debug, "t=", format(t), "| y=", year, " m=", month, " d=",
-    ##                  format(day, width=2), " h=", format(hour, width=2), " m=",
-    ##                  format(minute, width=2), "s=", format(second, width=2), "sec100=", sec100, "| pass",
-    ##                  format(pass, width=2), "/", passes, "| middle=", middle, "(",
-    ##                  format(middle/upper*100, digits=4), "%)\n")
-    ##         if (t.find < t)
-    ##             upper <- middle
-    ##         else
-    ##             lower <- middle
-    ##         if (upper - lower < 2)
-    ##             break
-    ##     }
-    ##     middle <- middle + add          # may use add to extend before and after window
-    ##     if (middle < 1)
-    ##         middle <- 1
-    ##     if (middle > len)
-    ##         middle <- len
-    ##     t <- ISOdatetime(unabbreviateYear(readBin(buf[profileStart[middle]+4], "integer", size=1, signed=FALSE, endian="little")),
-    ##                      as.integer(buf[profileStart[middle]+5]), # month
-    ##                      as.integer(buf[profileStart[middle]+6]), # day
-    ##                      as.integer(buf[profileStart[middle]+7]), # hour
-    ##                      as.integer(buf[profileStart[middle]+8]), # min
-    ##                      as.integer(buf[profileStart[middle]+9])+0.01*as.integer(buf[profileStart[middle]+10]), # decimal second
-    ##                      tz=tz)
-    ##     oceDebug(debug, "result: t=", format(t), " at vsdStart[", middle, "]=", profileStart[middle], "\n")
-    ##     oceDebug(debug, "} # bisectAdpRdiBuffer()\n", unindent=1)
-    ##     return(list(index=middle, time=t))
-    ## }
-    ##
[…]
