### 2021-07-05 30c3998
try to move sf to suggests (not working)
See https://github.com/dankelley/oce/issues/1846
