### 2021-07-08 3569ba9
move 'sf' from Depends to Suggests
