### 2021-07-15 e2908f6
testing EOW (edge-of-world) code in sandbox/dk/eow
