+++ sandbox/dk/eow/Makefile
@@ -1,0 +1,8 @@
+all: $(patsubst %.R,%.out,$(wildcard *.R))
+%.out: %.R
+	R --no-save < $< &> $@
+clean:
+	-rm *~ *.out
+view:
+	open *.png
+

[…]
