+++ sandbox/dk/eow/eow02.R
@@ -1,8 +1,17 @@
-doplot <- FALSE
+library(oce)
+source("~/git/oce/R/map.R")
+doplot <- !FALSE
+gridTest <- !FALSE
+
 debug <- 0L
-nlowroot <- 14 # 0.4 km (= 6371e3/2^14)
-nlowroot <- 16 # 0.1 km (= 6371e3/2^16)
-#nlowroot <- 15 # 0.2 km (= 6371e3/2^15)
+nlowroot <- 16                         #  0.09 km (= 6371/2^16)
+nlowroot <- 15                         #  0.19 km (= 6371/2^15)
+nlowroot <- 14                         #  0.39 km (= 6371/2^14)
+nlowroot <- 13                         #  0.79 km (= 6371/2^13)
+nlowroot <- 12                         #  1.5  km (= 6371/2^12)
+nlowroot <- 11                         #  3.1  km (= 6371/2^11)
+nlowroot <- 10                         #  6.2  km (= 6371/2^10)
+nlowroot <- 9                          # 12    km (= 6371/2^9)
 #nlowroot <- 12 # 1.5 km # eow03_20210714T074152.pdf
 #nlowroot <- 10 # 6.2 km
 ntheta <- 360 # 4.3  s with nlowroot=15
@@ -53,11 +62,10 @@ lowroot <- function(f, xlow, xhigh, n=15L, ...)
 }
 
 FAC <- 0.99
-library(oce)
 data(coastlineWorld)
-projs <- c("+proj=ortho",
-           "+proj=robin",
-           "+proj=moll")
+projs <- c("ortho +lon_0=-30 +lat_0=-20",
+           "robin",
+           "moll")
 if (interactive())
     projs <- projs[1]
 nprojs <- length(projs)
@@ -73,8 +81,9 @@ func <- function(r, theta)         # +1 if (r,theta) is visible, -1 otherwise
 }
 
 for (proj in projs) {
+    proj <- paste0("+proj=", proj)
     message(proj)
-    mapPlot(coastlineWorld, proj=proj, col="lightgray", type=if(doplot) "l" else "n")
+    mapPlot(coastlineWorld, proj=proj, col="black", type=if(doplot) "l" else "n")
     mtext(proj)
     usr <- par("usr")
     x0 <- mean(usr[1:2])
@@ -113,12 +122,45 @@ for (proj in projs) {
     XX[length(XX)] <- XX[1]
     YY[length(YY)] <- YY[1]
     eow <- sf::st_polygon(list(outer=cbind(XX, YY)))
-    if (doplot)
-        plot(eow, add=TRUE, border="red", lwd=3)
-    save(eow, file=paste0("eow", "_", gsub(".*=", "", proj), ".rda"))
+    if (doplot) {
+        #points(XX, YY)
+        #plot(eow, add=TRUE, border="red", lwd=3)
+    }
+    projBase <- gsub(".*=", "", gsub(" .*$", "", proj))
+    projBase <- paste0("eow_", projBase , ".rda")
+    save(eow, file=projBase)
+    message("saved to ", projBase)
     clon <- coastlineWorld[["longitude"]]
     clat <- coastlineWorld[["latitude"]]
 }
+
+if (gridTest) {
+    usr <- par("usr")
+    nxx <- nyy <- 50
+    xx <- seq(usr[1], usr[2], length.out=nxx)
+    yy <- seq(usr[3], usr[4], length.out=nyy)
+    xxx <- yyy <- NULL
+    for (ix in seq_len(nxx)[20]) {
+        for (iy in seq_len(nyy)[5:11]) {
[…]
