### 2021-07-16 7570b11
notes
