+++ sandbox/dk/eow/eow02rgdal.R
@@ -1,0 +1,175 @@
+library(oce)
+library(rgdal)
+source("~/git/oce/R/map.R")
+doplot <- !FALSE
+gridTest <- !FALSE
+
+debug <- 0L
+nlowroot <- 16                         #  0.09 km (= 6371/2^16)
+nlowroot <- 15                         #  0.19 km (= 6371/2^15)
+nlowroot <- 14                         #  0.39 km (= 6371/2^14)
+nlowroot <- 13                         #  0.79 km (= 6371/2^13)
+nlowroot <- 12                         #  1.5  km (= 6371/2^12)
+nlowroot <- 11                         #  3.1  km (= 6371/2^11)
+nlowroot <- 10                         #  6.2  km (= 6371/2^10)
+nlowroot <- 9                          # 12    km (= 6371/2^9)
+#nlowroot <- 12 # 1.5 km # eow03_20210714T074152.pdf
+#nlowroot <- 10 # 6.2 km
+ntheta <- 360 # 4.3  s with nlowroot=15
+ntheta <- 180 # 2.5  s with nlowroot=15
+ntheta <- 180 # 1.8  s with nlowroot=12
+ntheta <- 90  # 1.0  s with nlowroot=12
+ntheta <- 90  # 0.83 s with nlowroot=10
+# Time ntheta nlowroot Notes
+#  4.3    360       15
+#  2.5    180       15
+#  1.8    180       12
+#  1.1     90       15
+#  0.91    90       12
+#  2.7     90       16
+#  1.4     90       14 eow03.pdf looks good, acceptably slow
+
+if (!interactive()) pdf("eow02.pdf")
+
+#' assume decreasing function.
+#' @param f a function that decreases as x increases
+#' @param xlow an x value that yields f>0
+#' @param xhigh a larger x value
+#' @param n number of iterations. Note that 3*6371e3/2^15=583.2825 so dx typically under 1km
+#' @param ... extra parameters, passed to f()
+#' @return highest x that has f(x) > 0
+lowroot <- function(f, xlow, xhigh, n=15L, ...)
+{
+    flow <- f(xlow, ...)
+    if (flow <= 0.0)
+        stop("f(xlow) must be positive, but it is ", flow)
+    fhigh <- f(xhigh, ...)
+    if (fhigh > 0.0)
+        return(fhigh) # +ve at xhigh
+    while (n > 0)
+    {
+        xmid <- 0.5 * (xlow + xhigh)
+        fmid <- f(xmid, ...)
+        if (fmid < 0.0) {
+            xhigh <- xmid
+        } else if (fmid > 0.0) {
+            xlow <- xmid
+        } else {
+            break
+        }
+        n <- n - 1L
+    }
+    list(x=xlow, dx=xhigh-xlow)
+}
+
+FAC <- 0.99
+data(coastlineWorld)
+projs <- c("ortho +lon_0=-30 +lat_0=-20",
+           "robin",
+           "moll")
+if (interactive())
+    projs <- projs[1]
+nprojs <- length(projs)
+par(mar=c(1,1,2,1))
+if (interactive())
+    par(mfrow=c(1,nprojs))
+
+func <- function(r, theta)         # +1 if (r,theta) is visible, -1 otherwise
+{
+    x <- x0 + r*cos(theta)
+    y <- y0 + r*sin(theta)
+    ifelse(is.na(map2lonlat(x, y)$longitude), -1, 1)
+}
+
+for (proj in projs) {
+    proj <- paste0("+proj=", proj)
[…]
