+++ sandbox/dk/eow/eow02rgdal.R
@@ -12,22 +12,15 @@ nlowroot <- 13                         #  0.79 km (= 6371/2^13)
 nlowroot <- 12                         #  1.5  km (= 6371/2^12)
 nlowroot <- 11                         #  3.1  km (= 6371/2^11)
 nlowroot <- 10                         #  6.2  km (= 6371/2^10)
-nlowroot <- 9                          # 12    km (= 6371/2^9)
+#nlowroot <- 9                          # 12    km (= 6371/2^9)
+#nlowroot <- 8                          # 25    km (= 6371/2^8)
+#nlowroot <- 7                          # 50    km (= 6371/2^7)
 #nlowroot <- 12 # 1.5 km # eow03_20210714T074152.pdf
 #nlowroot <- 10 # 6.2 km
-ntheta <- 360 # 4.3  s with nlowroot=15
-ntheta <- 180 # 2.5  s with nlowroot=15
-ntheta <- 180 # 1.8  s with nlowroot=12
-ntheta <- 90  # 1.0  s with nlowroot=12
-ntheta <- 90  # 0.83 s with nlowroot=10
-# Time ntheta nlowroot Notes
-#  4.3    360       15
-#  2.5    180       15
-#  1.8    180       12
-#  1.1     90       15
-#  0.91    90       12
-#  2.7     90       16
-#  1.4     90       14 eow03.pdf looks good, acceptably slow
+ntheta <- 360
+ntheta <- 180
+ntheta <- 180
+ntheta <- 90
 
 if (!interactive()) pdf("eow02.pdf")
 
@@ -78,8 +71,10 @@ func <- function(r, theta)         # +1 if (r,theta) is visible, -1 otherwise
 {
     x <- x0 + r*cos(theta)
     y <- y0 + r*sin(theta)
-    ifelse(is.na(map2lonlat(x, y)$longitude), -1, 1)
+    ll <- try(rgdal::project(cbind(x,y), proj, inv=TRUE), silent=TRUE)
+    if (inherits(ll, "try-error") || !is.finite(ll[[1,1]])) -1 else 1
 }
+
 
 for (proj in projs) {
     proj <- paste0("+proj=", proj)
@@ -89,87 +84,47 @@ for (proj in projs) {
     usr <- par("usr")
     x0 <- mean(usr[1:2])
     y0 <- mean(usr[3:4])
-    R <- X <- Y <- vector("numeric", ntheta) # FIXME: for testing
+    R <- vector("numeric", ntheta) # FIXME: for testing
     Rmax <- (1/2) * sqrt((usr[2]-usr[1])^2 + (usr[4]-usr[3])^2) # hypotenuese
     thetas <- seq(0.0, 2*pi, length.out=ntheta)
+    owarn <- options()$warn
+    options(warn=-1)
+    # TIMING.  I think nlowroot=10, ntheta=90 is a reasonable choice.
+    # user-time nlowroot ntheta Notes
+    #      1.09       10     90   6 km resolution
+    #      1.05        9     90  12 km resolution (insignificant speedup)
+    #      2.08       10    180   6 km resolution (time scales as ntheta)
+    print(system.time({
     for (i in seq_along(thetas)) {
-        lr <- lowroot(func, xlow=0, xhigh=Rmax, n=nlowroot, theta=thetas[i])
+        capture.output({
+            lr <- lowroot(func, xlow=0, xhigh=Rmax, n=nlowroot, theta=thetas[i])
+        })
         R[i] <- lr$x
-        if (debug>1L)
[…]
