+++ sandbox/dk/eow/eow00.R
@@ -1,0 +1,6 @@
+if (FALSE) {
+    download.file("http://www.naturalearthdata.com/http//www.naturalearthdata.com/download/10m/physical/ne_10m_coastline.zip",
+                  destfile = "CL.zip")
+    unzip("CL.zip", exdir="CL")
+    CL <- rgdal::readOGR("CL/ne_10m_coastline.shp")
+}
