### 2021-07-18 bb61d31
Some promising results in eow02sf_02.R +
But more to be done:

1. what if *both* poles are in view?
2. are there problems at other lon_0 values?

I think I'll make a shiny app where you can click on a spot, and that
becomes the centre.  The LH panel could be the world in present
mapPlot() form, which responds to clicks. I could add the spokes again.
Maybe also a slider to set the number, etc.

With a shiny app, I can get more testers trying to break the algorithm.
The request would be simple: play with this and email me snapshots of
problem cases.  (The snapshot gives me lon_0 and lat_0, so I 
[…]
