+++ sandbox/dk/eow/eow0fix_proj.R
@@ -1,0 +1,11 @@
+projFix <- function(crs) # FIXME: check sf version (?)
+{
+    if (grepl("+proj=ortho", crs)) {
+        if (!grepl("+R=", crs))
+            crs <- paste(crs, "+R=6378137")
+        if (!grepl("+f=", crs))
+            crs <- paste(crs, "+f=0")
+    }
+    crs
+}
+
