+++ sandbox/dk/eow/Makefile
@@ -2,7 +2,7 @@ all: $(patsubst %.R,%.out,$(wildcard *.R))
 %.out: %.R
 	R --no-save < $< &> $@
 clean:
-	-rm *~ *.out
+	-rm *~ *.out *pdf *png
 view:
 	open *.png
 

[…]
