### 2021-07-18 c3d9d14
put NA at end so we catch Zimbabwe
