+++ sandbox/dk/eow/eow00coastline.R
@@ -1,8 +1,8 @@
 # eow08coastline.R: create polygon coastline (takes 0.02s if oce loaded, 1.1s otherwise)
 library(oce)
 data(coastlineWorld, package="oce")
-lon <- coastlineWorld[["longitude"]]
-lat <- coastlineWorld[["latitude"]]
+lon <- c(coastlineWorld[["longitude"]], NA)
+lat <- c(coastlineWorld[["latitude"]], NA)
 NAs <- which(is.na(lon))
 if (any(1L == diff(NAs)))
     stop("malformed coastline file (adjacent NA values)")
