+++ sandbox/dk/eow/ortho_app.R
@@ -1,9 +1,13 @@
-rm(list=ls())
+# ortho_app.R patterned on eow02sf_02.R
 library(oce)
-doplot <- !FALSE
-gridTest <- !FALSE
-lon0 <- -20
-lat0 <- -30
+library(shiny)
+data(coastlineWorld)
+load("coastlineWorldPolygon.rda")      # polygonzed coastlineWorld, made with eow00coastline.R
+
+drawRadii <- FALSE
+drawCircumferance <- TRUE
+plotHeight <- 300
+mapWidth <- 200
 
 projFix <- function(crs) # FIXME: check sf version (?)
 {
@@ -31,18 +35,17 @@ ntheta <- 360 # 4.3  s with nlowroot=15
 ntheta <- 180 # 2.5  s with nlowroot=15
 ntheta <- 180 # 1.8  s with nlowroot=12
 ntheta <- 90  # 1.0  s with nlowroot=12
-ntheta <- 90  # 0.83 s with nlowroot=10
 ntheta <- 45
-# Time ntheta nlowroot Notes
-#  4.3    360       15
-#  2.5    180       15
-#  1.8    180       12
-#  1.1     90       15
-#  0.91    90       12
-#  2.7     90       16
-#  1.4     90       14 eow03.pdf looks good, acceptably slow
+ntheta <- 20
 
-if (!interactive()) pdf("eow02sf.pdf")
+func <- function(r, theta, proj, proj0)
+{
+    #message("in func, proj=", proj)
+    x <- x0 + r*cos(theta)
+    y <- y0 + r*sin(theta)
+    t <- try(sf::sf_project(proj, proj0, cbind(x, y), warn=FALSE), silent=TRUE)
+    if (inherits(t, "try-error")) -1 else 1
+}
 
 #' assume decreasing function.
 #' @param f a function that decreases as x increases
@@ -75,67 +78,153 @@ lowroot <- function(f, xlow, xhigh, n=15L, ...)
     list(x=xlow, dx=xhigh-xlow)
 }
 
-FAC <- 0.99
-data(coastlineWorld)
-projs <- c("+proj=ortho +lon_0=-30 +lat_0=-20",
-           "+proj=robin",
-           "+proj=moll")
-if (interactive())
-    projs <- projs[1]
-nprojs <- length(projs)
-par(mar=c(1,1,2,1))
-if (interactive())
-    par(mfrow=c(1,nprojs))
+projs <- projFix("+proj=ortho +lon_0=-30 +lat_0=-20")
+proj0 <- projFix("+proj=longlat +datum=WGS84 +no_defs +R=6378137 +f=0")
 
-func <- function(r, theta)         # +1 if (r,theta) is visible, -1 otherwise
-{
-    x <- x0 + r*cos(theta)
-    y <- y0 + r*sin(theta)
-    ifelse(is.na(map2lonlat(x, y)$longitude), -1, 1)
+ui <- fluidPage(fluidRow(span(HTML("<center>Click in left panel to adjust view</center>"))),
+                fluidRow(column(5, plotOutput("plotL", click="click")),
+                         column(6, plotOutput("plotMR"))))
+
+server <- function(input, output, session) {
+    state <- reactiveValues(x=0.0, y=0.0)
+    observeEvent(input$click,
+                 {
+                     state$x <- input$click$x
+                     state$y <- input$click$y
+                 }
+    )
+
[…]
