### 2021-07-19 ff9796d
intersect in LL space
