+++ sandbox/dk/eow/ortho_app_xy.R
@@ -1,0 +1,214 @@
+# ortho_app.R patterned on eow02sf_02.R
+library(oce)
+library(shiny)
+data(coastlineWorld)
+# polygonzed coastlineWorld, made with eow00coastline.R
+load("coastlineWorldPolygon.rda")
+
+proj0 <- "+proj=longlat +datum=WGS84 +no_defs +R=6378137 +f=0"
+PROJ <- "ortho" # perhaps others will come later
+drawRadii <- FALSE
+drawCircumferance <- TRUE
+pch <- 20
+col <- 2
+cex <- 2
+colRegion <- rgb(1, 0, 0, 0.07)
+colLand <- "lightgray"
+
+projFix <- function(crs) # FIXME: check sf version (?)
+{
+    if (grepl("+proj=ortho", crs)) {
+        if (!grepl("+R=", crs))
+            crs <- paste(crs, "+R=6378137")
+        if (!grepl("+f=", crs))
+            crs <- paste(crs, "+f=0")
+    }
+    crs
+}
+
+eowToLongLat <- function(eow, proj, proj0)
+{
+    message("eowToLongLat with proj=", proj)
+    message("             and proj0=", proj0)
+    DANeow1 <<- eow
+    eowLL <- sf::sf_project(proj, proj0, eow, keep=TRUE, warn=!FALSE)
+    DANeow2 <<- eowLL
+    if (grepl("=ortho", proj)) { # ortho is tricky: add lines to whichever pole is visible
+        NPoleVisible <- !anyNA(sf::sf_project(proj0, proj, cbind(0,90), keep=TRUE, warn=FALSE))
+        SPoleVisible <- !anyNA(sf::sf_project(proj0, proj, cbind(0,-90), keep=TRUE, warn=FALSE))
+        #lonOrder <- order(eowLL[,1])
+        #eowLL <- eowLL[lonOrder,]
+        #DANeow3 <<- eowLL
+        if (SPoleVisible) {
+            message("S pole visible")
+            nearSPole <- eowLL[,2] < (-60)
+            eowLL[nearSPole,2] <- -90
+        } else if (NPoleVisible) {
+            message("N pole visible")
+            nearSPole <- eowLL[,2] > 70
+            eowLL[nearSPole,2] <- 90
+        }
+        DANeow3 <<- eowLL
+        # close the polygon
+        eowLL <- rbind(eowLL, c(eowLL[1,]))
+        DANeow4 <<- eowLL
+    }
+    eowLL
+}
+
+debug <- 0L
+nlowroot <- 16                         #   0.09 km (= 6371/2^16)
+nlowroot <- 15                         #   0.19 km (= 6371/2^15)
+nlowroot <- 14                         #   0.39 km (= 6371/2^14)
+nlowroot <- 13                         #   0.79 km (= 6371/2^13)
+nlowroot <- 12                         #   1.5  km (= 6371/2^12)
+nlowroot <- 11                         #   3.1  km (= 6371/2^11)
+nlowroot <- 10                         #   6.2  km (= 6371/2^10)
+#nlowroot <- 12 # 1.5 km # eow03_20210714T074152.pdf
+#nlowroot <- 10 # 6.2 km
+ntheta <- 360 # 4.3  s with nlowroot=15
+ntheta <- 180 # 2.5  s with nlowroot=15
+ntheta <- 180 # 1.8  s with nlowroot=12
+ntheta <- 90  # 1.0  s with nlowroot=12
+ntheta <- 45
+
+func <- function(r, theta, x0, y0, proj, proj0)
+{
+    #message("in func, proj=", proj)
+    x <- x0 + r*cos(theta)
+    y <- y0 + r*sin(theta)
[…]
