### 2021-07-19 6028792
ortho_app_xy improving but
Still, if AA is a sliver, it comes out wrong.  I suspect this is because
we are collapsing NA points after the projection.
