+++ sandbox/dk/eow/ortho_app_xy.R
@@ -113,15 +113,16 @@ lowroot <- function(f, xlow, xhigh, n=15L, ...)
 }
 
 
-ui <- fluidPage(fluidRow(span(HTML("<center>Click in any panel to adjust view</center>"))),
-                fluidRow(column(9, plotOutput("plot1", click="click1"))),
+ui <- fluidPage(fluidRow(span(HTML("<center>Click to recenter view</center>"))),
+                #fluidRow(column(9, plotOutput("plot1", click="click1"))),
                 fluidRow(column(6, plotOutput("plot2", click="click2")),
                          column(6, plotOutput("plot3", click="click2"))))
 
 server <- function(input, output, session) {
     state <- reactiveValues(x=0.0,
                             y=0.0,
-                            eow=NULL,
+                            eowXY=NULL,
+                            eowLL=NULL,
                             proj=projFix("+proj=ortho +lon_0=0 +lat_0=0"))
 
     observeEvent(input$click1,
@@ -136,10 +137,12 @@ server <- function(input, output, session) {
                  {
                      LL <- sf::sf_project(state$proj, proj0,
                                           cbind(input$click2$x, input$click2$y), keep=TRUE, warn=!FALSE)
-                     state$x <- max(min(LL[1,1], 180), -180)
-                     state$y <- max(min(LL[1,2], 90), -90)
-                     state$proj <- projFix(sprintf("+proj=%s +lon_0=%.2f +lat_0=%.2f",
-                                                   PROJ, state$x, state$y))
+                     if (is.finite(LL[1,1]) && is.finite(LL[1,2])) {
+                         state$x <- max(min(LL[1,1], 180), -180)
+                         state$y <- max(min(LL[1,2], 90), -90)
+                         state$proj <- projFix(sprintf("+proj=%s +lon_0=%.2f +lat_0=%.2f",
+                                                       PROJ, state$x, state$y))
+                     }
                  }
     )
 
@@ -147,16 +150,17 @@ server <- function(input, output, session) {
         #par(mar=c(1.5, 10.5, 1.5, 1.5))
         plot(coastlineWorld, mar=c(1,15,1,1.5))
         points(state$x, state$y, pch=pch, col=col, cex=cex)
-        if (!is.null(state$eow))
-            polygon(state$eow[,1], state$eow[,2], col=colRegion)
+        if (!is.null(state$eowLL))
+            polygon(state$eowLL[,1], state$eowLL[,2], col=colRegion)
         mtext(state$proj)
     }, height=300)
 
     output$plot2 <- renderPlot({ # left panel (not clickable)
-        par(mar=c(0.2,0.2,1,0.2))
+        par(mar=rep(1,4))#c(0.2,0.2,1,0.2))
         mapPlot(coastlineWorld, proj=state$proj, col=colLand, axes=FALSE)
         mapPoints(state$x, state$y, pch=pch, col=col, cex=cex)
-        mtext("Existing mapPlot() Method (blue=edge of earth)")
+        mtext("Existing", adj=0, cex=par("cex"))
+        mtext(state$proj, adj=1, cex=par("cex"))
         usr <- par("usr")
         x0 <- mean(usr[1:2])
         y0 <- mean(usr[3:4])
@@ -169,42 +173,67 @@ server <- function(input, output, session) {
         x <- x0 + R * cos(thetas)
         y <- y0 + R * sin(thetas)
         eow <- cbind(x=x, y=y)
-        message("next is eos (x y)")
-        print(file=stderr(), eow)
-        state$eow <- eowToLongLat(eow=eow, proj=state$proj, proj0=proj0)
-        message("next is state$eos (long lat)")
-        print(file=stderr(), state$eow)
-        lines(x, y, col=2, lwd=3)
[…]
