### 2021-07-20 ca05fae
compute polygon in-situ
