+++ sandbox/dk/eow/ortho_app_xy.R
@@ -1,10 +1,33 @@
 # ortho_app.R patterned on eow02sf_02.R
 library(oce)
 library(shiny)
+createCoastlinePolygon <- function(coastline)
+{
+    lon <- c(coastline[["longitude"]], NA)
+    lat <- c(coastline[["latitude"]], NA)
+    NAs <- which(is.na(lon))
+    if (any(1L == diff(NAs)))
+        stop("malformed coastline file (adjacent NA values)")
+    npoly <- length(NAs) - 1L
+    p <- vector("list", npoly)
+    for (i in seq_len(npoly)) {
+        look <- seq(NAs[i] + 1L, NAs[i+1]-1L)
+        LON <- lon[look]
+        LAT <- lat[look]
+        len <- length(look)
+        # Guess how to close those polygons that are not closed
+        if (LON[1] != LON[len] || LAT[1] != LAT[len]) {
+            LON <- c(LON, LON[1])
+            LAT <- c(LAT, LAT[1])
+        }
+        p[[i]] <- list(cbind(LON, LAT))
+    }
+    sf::st_multipolygon(p)
+}
 data(coastlineWorld)
-# polygonzed coastlineWorld, made with eow00coastline.R
-load("coastlineWorldPolygon.rda")
+coastlineWorldPolygon <- createCoastlinePolygon(coastlineWorld)
 
+plotHeight <- 400
 proj0 <- "+proj=longlat +datum=WGS84 +no_defs +R=6378137 +f=0"
 PROJ <- "ortho" # perhaps others will come later
 drawRadii <- FALSE
@@ -30,15 +53,10 @@ eowToLongLat <- function(eow, proj, proj0)
 {
     message("eowToLongLat with proj=", proj)
     message("             and proj0=", proj0)
-    DANeow1 <<- eow
     eowLL <- sf::sf_project(proj, proj0, eow, keep=TRUE, warn=!FALSE)
-    DANeow2 <<- eowLL
     if (grepl("=ortho", proj)) { # ortho is tricky: add lines to whichever pole is visible
         NPoleVisible <- !anyNA(sf::sf_project(proj0, proj, cbind(0,90), keep=TRUE, warn=FALSE))
         SPoleVisible <- !anyNA(sf::sf_project(proj0, proj, cbind(0,-90), keep=TRUE, warn=FALSE))
-        #lonOrder <- order(eowLL[,1])
-        #eowLL <- eowLL[lonOrder,]
-        #DANeow3 <<- eowLL
         if (SPoleVisible) {
             message("S pole visible")
             nearSPole <- eowLL[,2] < (-60)
@@ -48,33 +66,29 @@ eowToLongLat <- function(eow, proj, proj0)
             nearSPole <- eowLL[,2] > 70
             eowLL[nearSPole,2] <- 90
         }
-        DANeow3 <<- eowLL
         # close the polygon
         eowLL <- rbind(eowLL, c(eowLL[1,]))
-        DANeow4 <<- eowLL
     }
     eowLL
 }
 
 debug <- 0L
-nlowroot <- 16                         #   0.09 km (= 6371/2^16)
-nlowroot <- 15                         #   0.19 km (= 6371/2^15)
-nlowroot <- 14                         #   0.39 km (= 6371/2^14)
+#nlowroot <- 16                         #   0.09 km (= 6371/2^16)
+#nlowroot <- 15                         #   0.19 km (= 6371/2^15)
+#nlowroot <- 14                         #   0.39 km (= 6371/2^14)
 nlowroot <- 13                         #   0.79 km (= 6371/2^13)
-nlowroot <- 12                         #   1.5  km (= 6371/2^12)
+#nlowroot <- 12                         #   1.5  km (= 6371/2^12)
[…]
