### 2021-07-21 3dc6d01
remove outdated 'history' note dated May 2018
