+++ .github/workflows/test-coverage.yaml
@@ -22,7 +22,7 @@ jobs:
 
       - name: Query dependencies
         run: |
-          install.packages(c('remotes','sf')
+          install.packages(c('remotes','sf'))
           saveRDS(remotes::dev_package_deps(dependencies = TRUE), ".github/depends.Rds", version = 2)
           writeLines(sprintf("R-%i.%i", getRversion()$major, getRversion()$minor), ".github/R-version")
         shell: Rscript {0}
