### 2021-07-26 dc7deaa
adp flags: permit specifying any names
Before, we had an error message if the name was anything other than 'v'.
See https://github.com/dankelley/oce/issues/1852
