### 2021-07-26 9abd8e6
issue1852 - setFlags now permits any names
