### 2021-07-26 d30d1e4
test that data(adp) retains dim/type as before
This is to prevent accidentally changing the dataset, if we make moves
to optionally store a, g and d as numeric, instead of the present raw
type.  See #1854
