### 2021-07-29 #3319 @diodon
Hi @kthyng, thank you for the invitation. I'm quite busy the next 2-3 weeks but after I can review the paper. I'm quite familiar with the oce package so I will be glad to look at the submission. If the proposed time frame is good for you, then I can do it.
Cheers, EKS
