### 2021-08-05 d7f8913
adp[[i]]: better error if length(i)>1 (#1855)
