+++ R/adp.rdi.R
@@ -29,10 +29,9 @@
 
 decodeHeaderRDI <- function(buf, debug=getOption("oceDebug"), tz=getOption("oceTz"), ...)
 {
-
-    ##
-    ## header, of length 6 + 2 * numberOfDataTypes bytes
-    ##
+    #
+    # header length 6+2*numberOfDataTypes bytes (see e.g. Figure 44, page 160 of Surveyor docs)
+    #
     oceDebug(debug, "decodeHeaderRDI(buf, debug=", debug, ") {\n", unindent=1)
     if (buf[1] != 0x7f || buf[2] != 0x7f)
         stop("first two bytes in file must be 0x7f 0x7f, but they are 0x", buf[1], " 0x", buf[2])
@@ -247,6 +246,10 @@ decodeHeaderRDI <- function(buf, debug=getOption("oceDebug"), tz=getOption("oceT
     ##       HEADER  (6+2*num.types) bytes
     ##       FLD 50 bytes
     ##       VLD 60 bytes
+    # Ocean Surveyor Technical Manual_Dec20.pdf figure 44 page 160
+    #       HEADER (6+2*num.types) bytes
+    #       FLD 50 bytes
+    #       VLD 60 bytes
     ## dataOffset[1] = within-ensemble byte offset for FLD (e.g. Table D-1 of Surveyor manual)
     ## dataOffset[2] = within-ensemble byte offset for VLD (e.g. Table D-1 of Surveyor manual)
     ## thus, length of FLD is dataOffset[2]-dataOffset[1]
@@ -367,6 +370,7 @@ decodeHeaderRDI <- function(buf, debug=getOption("oceDebug"), tz=getOption("oceT
                 ##salinity=salinity
                 ##headingAlignment,
                 ##headingBias,
+                haveBinaryFixedAttitudeHeader=any(codes[, 1] == 0x00 & codes[, 2] == 0x30),
                 haveActualData=haveActualData)
     oceDebug(debug, "} # decodeHeaderRDI()\n", unindent=1)
     res
@@ -586,7 +590,7 @@ decodeHeaderRDI <- function(buf, debug=getOption("oceDebug"), tz=getOption("oceT
 #'   \tab 0x00 \tab 0x0c \tab Sentinel vertical beam amplitude\cr
 #'   \tab 0x00 \tab 0x0d \tab Sentinel vertical beam percent good\cr
 #'   \tab 0x00 \tab 0x20 \tab VMDASS\cr
-#' \*\tab 0x00 \tab 0x30 \tab binary fixed attitude (see p169 of reference 3 for format)\cr
+#'   \tab 0x00 \tab 0x30 \tab Binary Fixed Attitude header\cr
 #'   \tab 0x00 \tab 0x32 \tab Sentinel transformation matrix\cr
 #'   \tab 0x00 \tab 0x0a \tab Sentinel data\cr
 #'   \tab 0x00 \tab 0x0b \tab Sentinel correlation\cr
@@ -742,6 +746,7 @@ read.adp.rdi <- function(file, from, to, by, tz=getOption("oceTz"),
         buf <- buf[seq(startIndex, length(buf))]
     }
     header <- decodeHeaderRDI(buf, debug=debug-1)
+    oceDebug(debug, "have Binary Fixed Attitude Header: ", header$haveBinaryFixedAttitudeHeader, "\n")
     if (header$haveActualData) {
         numberOfBeams <- header$numberOfBeams
         numberOfCells <- header$numberOfCells
@@ -858,7 +863,9 @@ read.adp.rdi <- function(file, from, to, by, tz=getOption("oceTz"),
             codes <- header$codes
[…]
