### 2021-08-07 69e3546
cleanup re Sentinel Binary Fixed Attitude header
