+++ R/adp.rdi.R
@@ -50,22 +50,17 @@ decodeHeaderRDI <- function(buf, debug=getOption("oceDebug"), tz=getOption("oceT
         warning("dataOffset and numberOfDataTypes are inconsistent -- this dataset seems damaged")
     oceDebug(debug, "dataOffset=", paste(dataOffset, sep=" "), "\n")
     oceDebug(debug, "sort(diff(dataOffset))=", paste(sort(diff(dataOffset)), sep=" "), "\n")
-    ##
-    ## See if this is a sentinel file by looking for dataType ID bytes
-    ## of 0x00 and 0x70 (V series system configuration)
+    # Set up codes
     codes <- cbind(buf[1 + c(0, dataOffset)], buf[1+c(0, dataOffset) + 1])
-    oceDebug(debug, "buf[1:10] near line 95: ", paste("0x", paste(buf[1:10], sep=" "), sep=""), "\n")
-    oceDebug(debug, "codes[,1]=", paste("0x", paste(codes[,1], sep=""), sep=""), "\n")
-    oceDebug(debug, "codes[,2]=", paste("0x", paste(codes[,2], sep=""), sep=""), "\n")
-    if (any(codes[, 1] == 0x00 & codes[, 2] == 0x70)) {
-        oceDebug(debug, "Detected dataType 0x00 0x70 for Sentinel V series configuration")
-        isSentinel <- TRUE
-    } else {
-        isSentinel <- FALSE
-    }
-    oceDebug(debug, "isSentinel=", isSentinel, " as inferred from whether we have 0x00 0x70 ID\n")
-    ##
-    ## Fixed Leader Data, abbreviated FLD, pointed to by the dataOffset
+    oceDebug(debug, "set up codes starting at buf[1:10]=", paste("0x", paste(buf[1:10], sep=", "), sep=""), "\n")
+    oceDebug(debug, " codes[,1]=", paste("0x", paste(codes[,1], sep=""), sep=""), "\n")
+    oceDebug(debug, " codes[,2]=", paste("0x", paste(codes[,2], sep=""), sep=""), "\n")
+    # Determine whether this is a Sentinel V file.
+    isSentinel <- FALSE
+    isSentinel <- any(codes[, 1] == 0x00 & codes[, 2] == 0x70)
+    oceDebug(debug, "isSentinel =", isSentinel, "(inferred from whether we have 0x00 0x70 ID)\n")
+    #
+    # Fixed Leader Data, abbreviated FLD, pointed to by the dataOffset
     FLD <- buf[dataOffset[1]+1:(dataOffset[2] - dataOffset[1])]
     oceDebug(debug, "Fixed Leader Data:", paste(FLD, collapse=" "), "\n")
     if (FLD[1] != 0x00 && FLD[1] != 0x01)
@@ -79,6 +74,7 @@ decodeHeaderRDI <- function(buf, debug=getOption("oceDebug"), tz=getOption("oceT
     oceDebug(debug, "firmwareVersion=", firmwareVersion, "(numerically, it is", firmwareVersionNumeric, ")\n")
     ##if (firmwareVersion < 16.28) warning("firmwareVersion ", firmwareVersion, " is less than 16.28, and so read.adp.rdi() may not work properly")
 
+    # If no actual data, return something minimal
     if (!haveActualData)
         return(list(instrumentType="adcp",
                     firmwareVersionMajor=firmwareVersionMajor,
@@ -89,16 +85,16 @@ decodeHeaderRDI <- function(buf, debug=getOption("oceDebug"), tz=getOption("oceT
[…]
