### 2021-08-07 fb32116
demote a URL that failed this morning
New policy: demote \url{} to \code{} upon failure by
devtools::check_win_release().  There's no point in keeping these and
hoping for the best, because failures might harm CRAN submission, and
it's trivially easy for a user to copy the info and paste it into a
browser.
