### 2021-08-07 acb505d
remove stray output re debug value
