### 2021-08-09 #3594 @whedon
[REVIEW]: oce: an R package for Oceanographic Analysis
**Submitting author:** @dankelley (<a href="http://orcid.org/0000-0001-7808-5911">Dan E. Kelley</a>)
**Repository:** https://github.com/dankelley/oce
**Branch with paper.md** (empty if default branch):
**Version:** 1.6-1

- [x] #3726
