### 2021-08-11 00ca4c5
improve interpBarnes doc and debug output (#1864)
See https://github.com/dankelley/oce/issues/1864
