### 2021-08-13 #3319 @SeascapeScience
Hi Kristen,

For some reason this didn’t come through in my github notifications, and I’m just seeing it now in my email.

I’m guessing I’m too late, but please consider me as a potential reviewer in the future.

Cheers,
Nick

> On Jul 12, 2021, at 11:28 AM, Kristen Thyng ***@***.***> wrote:
>
>
> Hi @SeascapeScience <https://github.com/SeascapeScience>! Are you interested in reviewing this submission for JOSS, the Journal of Open Source Software? This is an open review which takes place on github with issues. You can learn more about the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>. I would hope you could start the review in the next few weeks with the intention of being through the typically iterative process in 4-6 weeks. Thanks for your consideration.
>
