### 2021-08-17 #3319 @kthyng
Thanks for the response @SeascapeScience! If you're interested, you could sign up to be a reviewer here: https://joss.theoj.org/reviewer-signup.html
