### 2021-08-24 cdca207
improve plot(section) docs re col vs zcol
