+++ man/plot-section-method.Rd
@@ -19,6 +19,8 @@
   xlim = NULL,
   ylim = NULL,
   zlim = NULL,
+  zbreaks = NULL,
+  zcol = NULL,
   map.xlim = NULL,
   map.ylim = NULL,
   clongitude,
@@ -30,8 +32,6 @@
   ztype = "contour",
   longitude0,
   latitude0,
-  zbreaks = NULL,
-  zcol = NULL,
   legend.loc = "bottomright",
   showStations = FALSE,
   showStart = TRUE,
@@ -95,9 +95,20 @@ drawing a coastline.}
 
 \item{ylim}{Optional limit for y axis (only in sections, not map)}
 
-\item{zlim}{Optional two-element numerical vector specifying the
-limit on the plotted field. This is used only if \code{ztype="image"};
-see also \code{zbreaks} and \code{zcol}.}
+\item{zlim, zbreaks, zcol}{Elements that control colours for \code{image} and \code{points}
+plot types, i.e. if \code{ztype} is either \code{"points"} or \code{"image"}.
+\code{zlim} is a two-element numerical vector specifying the
+limit on the plotted field.  If not provided, it defaults to the data
+range.
+\code{zbreaks} controls the colour breaks, in a manner that is similar to
+the \code{\link[=image]{image()}} parameter named \code{breaks}.  If not provided, \code{zbreaks} is
+inferred from \code{zlim}.
+\code{zcol}, which controls the colour scheme, may be a vector of colours
+(of length 1 less than \code{zbreaks}), or a function that takes an
+integer as its sole argument and returns that number of colours.
+If not provided, \code{zcol} defaults to \code{\link[=oceColoursViridis]{oceColoursViridis()}}.
+These three parameters are used in Example 6, an illustration of
+Atlantic salinity along 36N.}
 
 \item{map.xlim, map.ylim}{Optional limits for station map; \code{map.ylim} is
 ignored if \code{map.xlim} is provided.}
@@ -130,13 +141,6 @@ each station.}
 
 \item{longitude0, latitude0}{Location of the point from which distance is measured.
 These values are ignored unless \code{xtype} is \code{"distance"}.}
-
-\item{zbreaks, zcol}{Indication of breaks and colors to be used if \code{ztype="points"} or
-\code{"image"}. If not provided, reasonable default are used. If \code{zlim}
-is given but \code{breaks} is not given, then \code{breaks} is computed to
-run from \code{zlim[1]} to \code{zlim[2]}. If \code{zcol} is a function,
-it will be invoked with an argument equal to
-\code{1+length(zbreaks)}.}
 
 \item{legend.loc}{Location of legend, as supplied to \code{\link[=legend]{legend()}}, or
 set to the empty string to avoid plotting a legend.}
@@ -176,8 +180,8 @@ ignored otherwise.}
 \item{mar}{Value to be used with \code{\link{par}}\code{("mar")}. If not provided,
 a default is set up.}
 
-\item{col}{Color, which defaults to \code{\link{par}}\code{("col")} for line types, but
-to \link{oceColorsViridis} for image types.}
+\item{col}{Color for line types.  If not provided, this defaults to
+\code{\link{par}}\code{("col")}.  See \code{zcol}, for \code{ztype="image"} and \code{ztype="points"}.}
 
 \item{cex}{Numerical character-expansion factor, which defaults to \code{\link{par}}\code{("cex")}.}
 
@@ -242,10 +246,10 @@ library(oce)
 data(section)
 sg <- sectionGrid(section)
 
[…]
