### 2021-08-25 443cadf
add adpConvertRawToNumeric and adpFlagPastBoundary
