+++ man/read.adp.nortek.Rd
@@ -112,7 +112,9 @@ Other things related to adp data:
 \code{\link{[[<-,adp-method}},
 \code{\link{ad2cpHeaderValue}()},
 \code{\link{adp-class}},
+\code{\link{adpConvertRawToNumeric}()},
 \code{\link{adpEnsembleAverage}()},
+\code{\link{adpFlagPastBoundary}()},
 \code{\link{adp_rdi.000}},
 \code{\link{adp}},
 \code{\link{as.adp}()},
