+++ man/adpConvertRawToNumeric.Rd
@@ -1,25 +1,20 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/adp.R
-\name{beamUnspreadAdp}
-\alias{beamUnspreadAdp}
-\title{Adjust ADP Signal for Spherical Spreading}
+\name{adpConvertRawToNumeric}
+\alias{adpConvertRawToNumeric}
+\title{Convert Raw to Numeric Values For adp Objects}
 \usage{
-beamUnspreadAdp(
-  x,
-  count2db = c(0.45, 0.45, 0.45, 0.45),
-  asMatrix = FALSE,
+adpConvertRawToNumeric(
+  object = NULL,
+  variables = NULL,
   debug = getOption("oceDebug")
 )
 }
 \arguments{
-\item{x}{an \linkS4class{adp} object.}
+\item{object}{an \linkS4class{adp} object.}
 
-\item{count2db}{a set of coefficients, one per beam, to convert from beam
-echo intensity to decibels.}
-
-\item{asMatrix}{a boolean that indicates whether to return a numeric matrix,
-as opposed to returning an updated object (in which the matrix is cast to a
-raw value).}
+\item{variables}{variables stored in an \linkS4class{adp} object that has
+the same dimensional as \code{v} and is stored in a raw format.}
 
 \item{debug}{an integer specifying whether debugging information is
 to be printed during the processing. This is a general parameter that
@@ -30,40 +25,18 @@ be printed. If one function calls another, it usually reduces the value of
 by specifying higher \code{debug} values.}
 }
 \value{
-An \linkS4class{adp} object.
+\code{adpConvertRawToNumeric} returns an \linkS4class{adp} object whose specified
+variables have been converted from raw to numerical format.
 }
 \description{
-Compensate ADP signal strength for spherical spreading.
-}
-\details{
-First, beam echo intensity is converted from counts to decibels, by
-multiplying by \code{count2db}.  Then, the signal decrease owing to
-spherical spreading is compensated for by adding the term
-\eqn{20\log10(r)}{20*log10(r)}, where \eqn{r}{r} is the distance from the
-sensor head to the water from which scattering is occurring.  \eqn{r}{r} is
-given by \code{x[["distance"]]}.
+Convert variables in an \linkS4class{adp} object from raw to numeric format.
 }
 \examples{
 library(oce)
 data(adp)
-plot(adp, which=5) # beam 1 echo intensity
-adp.att <- beamUnspreadAdp(adp)
-plot(adp.att, which=5) # beam 1 echo intensity
-## Profiles
-par(mar=c(4, 4, 1, 1))
-a <- adp[["a", "numeric"]]             # second arg yields matrix return value
-distance <- adp[["distance"]]
-plot(apply(a,2,mean), distance, type='l', xlim=c(0,256))
-lines(apply(a,2,median), distance, type='l',col='red')
-legend("topright",lwd=1,col=c("black","red"),legend=c("original","attenuated"))
-## Image
-plot(adp.att, which="amplitude",col=oce.colorsViridis(100))
-
-}
-\references{
-The coefficient to convert to decibels is a personal
-communication.  The logarithmic term is explained in textbooks on acoustics,
-optics, etc.
+adp[["a"]][,,1][,1]
+ADP <- adpConvertRawToNumeric(adp)
+ADP[["a"]][,,1][,1]
 }
 \seealso{
 Other things related to adp data: 
@@ -71,7 +44,6 @@ Other things related to adp data:
 \code{\link{[[<-,adp-method}},
[…]
