### 2021-08-25 2c0724d
add message for handleFlags
