+++ NEWS.md
@@ -3,15 +3,15 @@
 ## 1.5.0
 
 * Add `showNA` argument to `vectorShow()`.
+* Add `adpConvertRawToNumeric()`.
+* Add `adpFlagPastBoundary()`.
+* Change `read.odf()` to handle new CODEs and UNITs.
 * Change `oce.plot.ts()` by adding `simplify` argument.
 * Change `pwelch()`, improving low-frequency results.
 * Fix `[[` to yield ITS90 temperature for all classes, not just `ctd`.
 * Fix `as.ctd()` handling of temperature scale of first argument.
 * Fix `as.section()` handling of list of `argo` objects as first argument.
 * Fix `colormap()` handling of `name` argument.
-* `read.odf()` handles many new CODE and UNIT possibilities.
-* Add `adpConvertRawToNumeric`.
-* Add `adpFlagPastBoundary`.
 
 ## 1.4.0
 
