+++ DESCRIPTION
@@ -35,9 +35,8 @@ Description: Supports the analysis of Oceanographic data, including 'ADCP'
     Provides specialized functions for calculating seawater properties such as
     potential temperature in either the 'UNESCO' or 'TEOS-10' equation of state.
     Produces graphical displays that conform to the conventions of the
-    Oceanographic literature. This package is discussed extensively in
-    Dan Kelley's book Oceanographic Analysis with R, published
-    in 2018 by 'Springer-Verlag' with ISBN 978-1-4939-8842-6.
+    Oceanographic literature. This package is discussed extensively by
+    Kelley (2018) "Oceanographic Analysis with R" <doi:10.1007/978-1-4939-8844-0>.
 License: GPL (>= 2)
 Encoding: UTF-8
 URL: https://dankelley.github.io/oce/
