### 2021-09-09 78f8d52
few-word change in Rd from new roxygen2
