+++ DESCRIPTION
@@ -42,7 +42,7 @@ Encoding: UTF-8
 URL: https://dankelley.github.io/oce/
 LazyData: false
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.1.1
+RoxygenNote: 7.1.2
 BuildVignettes: true
 VignetteBuilder: knitr
 NeedsCompilation: yes
