+++ tests/testthat/local_data/how_to_subset_binary_data.R
@@ -1,0 +1,3 @@
+b <- readBin("adp_sentinel_v.pd0", what="raw", n=1e6)
+writeBin(b, "test.pd0")
+
