### 2021-10-24 44cb0da
read.ctd.woce() ignores missing final newline (close #1875)
https://github.com/dankelley/oce/issues/1875
