+++ R/ctd.woce.R
@@ -543,7 +543,7 @@ read.ctd.woce <- function(file, columns=NULL, station=NULL, missingValue, deploy
 
         ## Read the data into a buffer, since there will likely be
         ## a trailer line at the end, and read.table() cannot handle that.
-        lines <- readLines(file)
+        lines <- readLines(file, warn=FALSE)
         nlines <- length(lines)
         if (length(grep("^END", lines[nlines])))
             lines <- lines[-nlines]
