### 2021-10-24 b3a5898
fix testthat problem (close 1876)
