+++ tests/testthat/test_sw.R
@@ -1,4 +1,4 @@
-# vim:textwidth=120:expandtab:shiftwidth=2:softtabstop=2
+# vim:textwidth=140:expandtab:shiftwidth=4:softtabstop=4
 
 library(oce)
 # Table of contents.
@@ -37,398 +37,356 @@ library(oce)
 # [2] Trevor J. McDougall, 1987. Neutral Surfaces, Journal of Physical
 #     Oceanography, volume 17, pages 1950-1964.
 
-test_that("rho and sigma",
-          {
-            # 1. rho and sigma
-            # 1.1 UNESCO rho [1 p19]. Note that we must
-            # convert to the T68 temperature scale, which was in use at the time
-            # that [1] was written.
-            S <- c( 0,   0,   0,   0,  35,  35,  35,  35)
-            T <- T90fromT68(c( 5,   5,  25,  25,   5,   5,  25,  25))
-            p <- c( 0, 1e4,   0, 1e4,   0, 1e4,   0, 1e4)
-            rho <- c(999.96675, 1044.12802, 997.04796, 1037.90204, 1027.67547, 1069.48914, 1023.34306, 1062.53817)
-            expect_equal(swRho(S, T, p, eos="unesco"), rho)
-            # check sigma from this
-            expect_equal(swRho(S, T, p, eos="unesco")-1000, swSigma(S,T,p,eos="unesco"))
-            # 1.2 GSW
-            # Since gsw_ functions are tested in the gsw package, we just need a consistency check.
-            longitude <- 188
-            latitude <- 4
-            SP <- 35
-            t <- 10
-            p <- 1000
-            SA <- gsw::gsw_SA_from_SP(SP, p, longitude, latitude)
-            CT <- gsw::gsw_CT_from_t(SA, t, p)
-            # Test density.
-            rhoGSW <- gsw::gsw_rho(SA, CT, p)
-            rho <- swRho(SP, t, p, longitude, latitude, "gsw")
-            expect_equal(rhoGSW, rho)
-            # Now use density to test sigma (not provided by gsw).
-            sigma <- swSigma(SP, t, p, longitude, latitude, "gsw")
-            expect_equal(rhoGSW-1000, sigma)
-            # The following was hard-coded using values from GSW3.03, and it failed with GSW3.05.
-            expect_equal(30.818302,swSigma(35,T90fromT68(13),1000,eos="unesco"),tolerance=0.000001)
-            # The sigmaT tests are not from definititive test values, and so are just
-            # checks against future changes.
-            expect_equal(swSigmaT(35, T90fromT68(13), 1000, eos="unesco"), 26.393538,tolerance=0.000001)
-          }
-)
+test_that("rho and sigma", {
+    # 1. rho and sigma
+    # 1.1 UNESCO rho [1 p19]. Note that we must
+    # convert to the T68 temperature scale, which was in use at the time
+    # that [1] was written.
+    S <- c( 0,   0,   0,   0,  35,  35,  35,  35)
+    T <- T90fromT68(c( 5,   5,  25,  25,   5,   5,  25,  25))
+    p <- c( 0, 1e4,   0, 1e4,   0, 1e4,   0, 1e4)
+    rho <- c(999.96675, 1044.12802, 997.04796, 1037.90204, 1027.67547, 1069.48914, 1023.34306, 1062.53817)
+    expect_equal(swRho(S, T, p, eos="unesco"), rho)
+    # check sigma from this
+    expect_equal(swRho(S, T, p, eos="unesco")-1000, swSigma(S,T,p,eos="unesco"))
+    # 1.2 GSW
[…]
