### 2021-11-04 2467ca2
argo[["cycle"]] works as argo[["cycleNumber"]]
