### 2021-11-04 #3594 @gonzalobravoargentina
I have not many comments about the software because is continuously improving and all the basic tools for oceanographic analysis are cover. This paper will increase the visibility of OCE package and more users will contribute for upgrading the quality of the codes.

The list of available R packages for Oceanographic data is not so long but, in my opinion, Oce is the most complete and user friendly
See:
[https://hafezahmad.medium.com/popular-r-packages-used-in-oceanography-fisheries-and-environmental-science-e3fbf0805f1a](url)

I would like to add some comments regarding the statement of need of software paper and State of the field of the software paper.

_In the marine biology community, R is one of the most popular software and a great number of people are used to this language. In many cases marine biologists need to process oceanographic data for their studies and the OCE package is a great option to avoid learning new languages (e.g. MATLAB, Python).  The R oce package was an excellent tool during my studies and the support of the authors was excellent._

 @dankelley_Do you think that a short list of advantages and disadvantages of R vs MATLAB or others in the oceanographic field could be added?_
