### 2021-11-04 #3594 @dankelley
@gonzalobravoargentina that's a good idea about R vs Matlab.  I talk about R, Matlab and others in my textbook Oceanographic Analysis with R.  I think I will follow your advice, and write a paragraph about this for the README file on the landing page for github/dankelley/R.

And many thanks for mentioning biological oceanography.  I have also found, in my teaching of undergraduates, that R is very popular in that community.

Again, thanks for the review.

Dan
