### 2021-11-04 d0238b1
add material on why R is good for oceanography
I thank @gonzalobravoargentina for the suggestion to discuss this topic,
as part of a review for a JOSS paper about oce.  See
https://github.com/openjournals/joss-reviews/issues/3594 for details.
