### 2021-11-04 #3594 @dankelley
@gonzalobravoargentina I have added the following paragraph to the README.  I hope it gets close to what you were thinking of.  Thanks again.  Dan.

## Why use R for oceanographic analysis?

The R language is popular in many branches of science, and Oceanography is not
an exception. Perhaps because of its broad statistical support, R is especially
popular in Biological Oceanography, and its popularity seems to be growing also
among Chemical and Geological Oceanographers. In the remaining subdiscipline of
Physical Oceanography, many researchers remain attached to Matlab, which was
the strongest choice for interactive work in the late 1900s. However, the
commercial and closed-source nature of Matlab is a problem in the eyes of many
people, and so open-source alternatives have become increasingly popular.  Of
these, Python and R are the most popular languages at the moment, and many
Oceanographers are conversant with each, switching between the two according to
the task at hand. A strength of R in this regard is the availability of
well-vetted and well-documented packages for handling a wide spectrum of
specialized tasks. Of these, the oce package is particularly important.
