### 2021-11-07 70e0e6c
experimental in R, code speedup in src
