+++ src/interp_barnes.cpp
@@ -25,19 +25,17 @@ inline double exp_approx(double x)
 }
 #endif
 
-static time_t start;
-
 static double interpolate_barnes(double xx, double yy, double zz, /* interpolate to get zz value at (xx,yy) */
     int skip, /* value in (x,y,z) to skip, or -1 if no skipping */
     unsigned int nx, double *x, double *y, double *z, double *w, /* data num, locations, values, weights */
     double *z_last, /* last estimate of z at (x,y) */
-    double xr, double yr, int debug) /* influence radii */
+    double xr, double yr) /* influence radii */
 {
   double sum = 0.0, sum_w = 0.0;
+  double dx, dy, d, weight;
   for (unsigned int k = 0; k < nx; k++) {
     // R trims NA (x,y values so no need to check here
     if ((int)k != skip) {
-      double dx, dy, d, weight;
       dx = (xx - x[k]) / xr;
       dy = (yy - y[k]) / yr;
       d = dx*dx + dy*dy;
@@ -59,8 +57,7 @@ static double weight_barnes(double xx, double yy,
     unsigned int n, double *x, double *y, double *z, double *w,
     double xr, double yr)
 {
-  double sum_w, dx, dy, d, weight;
-  sum_w = 0.0;
+  double sum_w = 0.0, dx, dy, d, weight;
   for (unsigned int k = 0; k < n; k++) {
     if ((int)k != skip) {
       dx = (xx - x[k]) / xr;
@@ -79,42 +76,39 @@ static double weight_barnes(double xx, double yy,
 
 
 // [[Rcpp::export]]
-List do_interp_barnes(NumericVector x, NumericVector y, NumericVector z, NumericVector w, NumericVector xg, NumericVector yg, NumericVector xr, NumericVector yr, NumericVector gamma, NumericVector iterations)
+List do_interp_barnes(NumericVector x, NumericVector y, NumericVector z, NumericVector w,
+    NumericVector xg, NumericVector yg,
+    NumericVector xr, NumericVector yr,
+    NumericVector gamma, NumericVector iterations)
 {
-  start = time(NULL);
   int nx = x.size();
   int nxg = xg.size();
   int nyg = yg.size();
-  NumericMatrix zg(nxg, nyg), wg(nxg, nyg); // predictions on the grid
-  NumericVector zd(nx); // predictions at the data
-
-  //Rprintf("xg[0]=%f; size=%d\n", xg[0], xg.size());
-  //Rprintf("yg[0]=%f; size=%d\n", yg[0], yg.size());
-
   double rgamma = gamma[0]; // gamma
   if (rgamma < 0.0)
-    ::Rf_error("gamma=%f cannot be non-positive", rgamma);
+    ::Rf_error("cannot have gamma < 0, but got gamma=%f", rgamma);
   int niter = floor(0.5 + iterations[0]); // number of iterations
-  if (niter < 0)
-    ::Rf_error("cannot have a negative number of iterations.  Got %d ", niter);
+  if (niter < 1)
+    ::Rf_error("cannot have fewer than 1 iteration, but got niter=%d ", niter);
   if (niter > 20)
-    ::Rf_error("cannot have more than 20 iterations.  Got %d ", niter);
+    ::Rf_error("cannot have more than 20 iterations, but got niter=%d ", niter);
   if (xr[0] <= 0)
-    ::Rf_error("cannot have xr<=0 but it is %f", xr[0]);
+    ::Rf_error("cannot have xr<=0 but got xr=%f", xr[0]);
   if (yr[0] <= 0)
-    ::Rf_error("cannot have yr<=0 but it is %f", yr[0]);
[…]
