### 2021-11-08 e61d7ba
document some timing tests
