### 2021-11-18 3678b14
ctd[[ on 'Absolute Salinity' & 'absolute salinity'
See

https://github.com/dankelley/oce/issues/1885
