### 2021-11-19 1dec3b3
disallow ctd[["absolute salinity"]]
See https://github.com/dankelley/oce/issues/1885
