### 2021-11-19 612bc51
remove ctd[["absolute salinity"]] from tests
