### 2021-11-21 7e40434
reorder items in diagr
