+++ man/figures/ctd-object-diagram.R
@@ -34,9 +34,9 @@ lines(c(xl, xr), rep(0.79, 2), col=lcol)
 lines(rep(xl, 2), c(0.79, 0.75), col=lcol)
 lines(rep(xr, 2), c(0.79, 0.75), col=lcol)
 textInBox(xl, 0.9-2*l, "       data       ")
-textInBox(xl+o, 0.9-4*l, "pressure", 3)
+textInBox(xl+o, 0.9-4*l, "salinity", 2.5)
 textInBox(xl++o, 0.9-6*l, "temperature", 4.9)
-textInBox(xl+o, 0.9-8*l, "salinity", 2.5)
+textInBox(xl+o, 0.9-8*l, "pressure", 3)
 textInBox(xl+o, 0.9-10*l, " ... ", 1.1)
 lines(rep(xl-0.05, 2), c(0.66, 0.09), col=lcol)
 lines(c(xl-0.05, xl-0.05+0.06), rep(0.55, 2), col=lcol)
