+++ man/figures/ctd-object-diagram.R
@@ -33,7 +33,7 @@ lines(rep(xm,2), c(0.83, 0.745), col=lcol)
 lines(c(xl, xr), rep(0.79, 2), col=lcol)
 lines(rep(xl, 2), c(0.79, 0.75), col=lcol)
 lines(rep(xr, 2), c(0.79, 0.75), col=lcol)
-textInBox(xl, 0.9-2*l, "       data       ")
+textInBox(xl, 0.9-2*l, "      data      ")
 textInBox(xl+o, 0.9-4*l, "salinity", 2.5)
 textInBox(xl++o, 0.9-6*l, "temperature", 4.9)
 textInBox(xl+o, 0.9-8*l, "pressure", 3)
