+++ man/figures/ctd-object-diagram.R
@@ -36,7 +36,7 @@ lines(rep(xr, 2), c(0.79, 0.75), col=lcol)
 textInBox(xl, 0.9-2*l, "      data      ")
 textInBox(xl+o, 0.9-4*l, "salinity", 2.5)
 textInBox(xl++o, 0.9-6*l, "temperature", 4.9)
-textInBox(xl+o, 0.9-8*l, "pressure", 3)
+textInBox(xl+o, 0.9-8*l, "pressure", 3.4)
 textInBox(xl+o, 0.9-10*l, " ... ", 1.1)
 lines(rep(xl-0.05, 2), c(0.66, 0.09), col=lcol)
 lines(c(xl-0.05, xl-0.05+0.06), rep(0.55, 2), col=lcol)
@@ -46,8 +46,8 @@ lines(c(xl-0.05, xl-0.05+0.065), rep(0.09, 2), col=lcol)
 
 # Middle
 textInBox(xm, 0.9-2*l, "   metadata   ")
-textInBox(xm+o, 0.9-4*l, "header", 2.11)
-textInBox(xm+o, 0.9-6*l, "filename", 2.7)
+textInBox(xm+o, 0.9-4*l, "units", 1.5)
+textInBox(xm+o, 0.9-6*l, "filename", 2.9)
 textInBox(xm+o, 0.9-8*l, " ... ", 0.5)
 lines(rep(xm-0.05, 2), c(0.66, 0.25), col=lcol)
 lines(c(xm-0.05, xm-0.05+0.06), rep(0.55, 2), col=lcol)
