+++ man/figures/ctd-object-diagram.R
@@ -1,5 +1,5 @@
 # oce object diagram (a terribly brittle kludge)
-png("ctd-object-new.png", width=5, height=2.5, unit="in", res=150, pointsize=9)
+png("ctd-object.png", width=5, height=2.5, unit="in", res=150, pointsize=9)
 
 textInBox <- function(x, y, text, xoff=0, yoff=0,
     col="black", cex=1, font=1, lwd=1, pos=1)
@@ -16,7 +16,7 @@ textInBox <- function(x, y, text, xoff=0, yoff=0,
     yb <- y - 2*h - ex/2
     yt <- y
     lines(c(xl, xl, xr, xr, xl), c(yb, yt, yt, yb, yb), col=col, lwd=lwd)
-    list(xl=xl, xr=xr, yb=yb, yt=yt)
+    list(xl=xl, xr=xr, yb=yb, yt=yt, xm=0.5*(xl+xr), ym=0.5*(yb+yt))
 }
 
 par(mar=rep(0, 4))
@@ -27,6 +27,7 @@ xl <- 0.1
 xm <- 0.5
 xr <- 0.9
 lcol <- "darkgray"                     # line colour
+llwd <- 2
 
 # Middle
 ctdObject  <- textInBox(xm, 0.92, "ctd object")
@@ -47,27 +48,32 @@ processingLogTime <- textInBox(xr+o, 0.9-4*l, "time", 1)
 processingLogAction <- textInBox(xr+o, 0.9-6*l, "action", 2)
 
 # Join 'ctd object' with its three descendents
-lines(rep(xm,2), c(ctdObject$yb, metadata$yt), col=lcol)
+lines(rep(xm,2), c(ctdObject$yb, metadata$yt), col=lcol, lwd=llwd)
+lines(rep(xm-0.04,2), c(ctdObject$yb, 0.5*(ctdObject$yb+metadata$yt)), col=lcol, lwd=llwd)
+lines(c(data$xm, xm-0.04), rep(0.5*(ctdObject$yb+metadata$yt),2), col=lcol, lwd=llwd)
+lines(c(processingLog$xm, xm+0.04), rep(0.5*(ctdObject$yb+metadata$yt),2), col=lcol, lwd=llwd)
+lines(rep(xm+0.04,2), c(ctdObject$yb, 0.5*(ctdObject$yb+metadata$yt)), col=lcol, lwd=llwd)
+#lines(rep(xm,2), c(ctdObject$yb, metadata$yt), col=lcol, lwd=llwd)
 ymid <- 0.5 * (ctdObject$yb + metadata$yt)
-lines(c(xl, xr), rep(ymid, 2), col=lcol)
-lines(rep(xl, 2), c(ymid, data$yt), col=lcol)
-lines(rep(xr, 2), c(ymid, data$yt), col=lcol)
+#lines(c(xl, xr), rep(ymid, 2), col=lcol, lwd=llwd)
+lines(rep(xl, 2), c(ymid, data$yt), col=lcol, lwd=llwd)
+lines(rep(xr, 2), c(ymid, data$yt), col=lcol, lwd=llwd)
 # Join 'data' and its 4 descendents
-lines(rep(xl-0.05, 2), c(data$yb, with(dataDots, (yb+yt)/2)), col=lcol)
+lines(rep(xl-0.05, 2), c(data$yb, with(dataDots, (yb+yt)/2)), col=lcol, lwd=llwd)
 DX <- 0.05
-lines(c(xl-DX, dataSalinity$xl), rep(with(dataSalinity, (yb+yt)/2), 2), col=lcol)
-lines(c(xl-DX, dataSalinity$xl), rep(with(dataTemperature, (yb+yt)/2), 2), col=lcol)
[…]
