### 2021-11-29 63917e1
improve error if class(first arg) is not 'numeric'
