### 2021-12-02 f0b67d8
fix doc error for swSpecificHeat()
See https://github.com/dankelley/oce/issues/1887 for details.
Thanks, @j-harbin, for pointing this out.
