+++ R/argo.R
@@ -66,11 +66,21 @@ NULL
 #'
 #' @section Details of the specialized `argo` method:
 #'
-#' There are several possibilities, depending on the nature of `i`.
 #' Note that [argo-class] data may contain both unadjusted data and adjusted
 #' data.  By default, this extraction function refers to the former, but a
 #' preference for the latter may be set with [preferAdjusted()], the
 #' documentation of which explains (fairly complex) details.
+#'
+#' There are several possibilities for how argo[[i]] works,
+#' depending on the nature of `i`. The possibilities are as
+#' follows.
+#'
+#' * If `i` is `"?"`, then the return value is an
+#' alphabetically sorted character vector that holds
+#' the names of the available data items.  This is constructed
+#' from the names of items in the `data`
+#' and `metadata` slots, combined with certain computed entries
+#' (e.g. `"Absolute Salinity"`).
 #'
 #' * If `i` is `"profile"` and `j` is an integer vector,
 #' then an argo object is returned, as specified by `j`. For example,
@@ -139,7 +149,16 @@ setMethod(f="[[",
               dots <- list(...)
               debug <- if ("debug" %in% names(dots)) dots$debug else 0
               oceDebug(debug, "[[,argo-method(\"", i, "\") {\n", sep="", style="bold", unindent=1)
-              if (i == "profile") {
+              iKnown <- sort(unique(c(names(x@data), names(x@metadata),
+                          "profile",
+                          "CT", "N2", "SA", "sigmaTheta", "theta", "sigma0",
+                          "sigma1", "sigma2", "sigma3", "sigma4",
+                          "z", "depth", "ID", "id", "cycleNumber", "cycle",
+                          "latitude", "longitude",
+                          "*Flag", "*Unit")))
+              if (i == "?") {
+                  return(iKnown)
+              } else if (i == "profile") {
                   ## This assignment to profile is merely to prevent a NOTE from
                   ## the syntax checker. It is needed because of issues with non-standard
                   ## evaluation in subset() calls. This is a problem that many
