+++ man/sub-sub-argo-method.Rd
@@ -40,7 +40,7 @@ data.  By default, this extraction function refers to the former, but a
 preference for the latter may be set with \code{\link[=preferAdjusted]{preferAdjusted()}}, the
 documentation of which explains (fairly complex) details.
 
-There are several possibilities for how argo[\link{i}] works,
+There are several possibilities for how \code{argo[[i]]} works,
 depending on the nature of \code{i}. The possibilities are as
 follows.
 \itemize{
@@ -53,7 +53,7 @@ and \code{metadata} slots, combined with certain computed entries
 \item If \code{i} is \code{"profile"} and \code{j} is an integer vector,
 then an argo object is returned, as specified by \code{j}. For example,
 \code{argo[["profile", 2:5]]} is equivalent to
-\verb{subset(argo, profile \\\%in\\\% 2:5)}.
+\code{subset(argo, profile \%in\% 2:5)}.
 \item If \code{i} is \code{"CT"}, then
 Conservative Temperature is returned, as computed with
 \code{\link[gsw:gsw_CT_from_t]{gsw::gsw_CT_from_t}}\verb{(SA,t,p)}, where
