+++ man/sub-sub-argo-method.Rd
@@ -40,16 +40,19 @@ data.  By default, this extraction function refers to the former, but a
 preference for the latter may be set with \code{\link[=preferAdjusted]{preferAdjusted()}}, the
 documentation of which explains (fairly complex) details.
 
-The results from \code{argo[[i]]} depend on the nature of \code{i} and,
-in some cases, an additional argument named \code{j}.
+The results from \code{argo[[i]]} or \code{argo[[i,j]]} depend on the
+nature of \code{i} and (if provided) \code{j}.
 The possibilities are as follows.
 \itemize{
 \item If \code{i} is \code{"?"}, then the return value is an
 alphabetically sorted character vector that holds
-the names of the available data items.  This is constructed
+the names of the items that can be returned by \code{[[}.
+This vector is constructed
 from the names of items in the \code{data}
 and \code{metadata} slots, combined with certain computed entries
-(e.g. \code{"Absolute Salinity"}).
+(e.g. \code{"Absolute Salinity"}).  Flags and units can also be
+retrieved, as explained under the heading \dQuote{Details of the general
+method}.
 \item If \code{i} is \code{"profile"} and \code{j} is an integer vector,
 then an argo object is returned, as specified by \code{j}. For example,
 \code{argo[["profile", 2:5]]} is equivalent to
