### 2021-12-10 f2859fb
address issue 1894 (df for swN2)
