+++ R/sw.R
@@ -295,11 +295,11 @@ swN2 <- function(pressure, sigmaTheta=NULL, derivs, df,
         if (derivs == "simple") {
             sigmaThetaDeriv <- c(0, diff(sigmaTheta) / diff(pressure))
         } else if (derivs == "smoothing") {
-            depths <- sum(!is.na(pressure))
+            depths <- length(unique(pressure[ok]))
             if (missing(df)) {
-                df <- if (depths > 100) f <- floor(depths / 10) # at least 10
-                    else if (depths > 20) f <- floor(depths / 3) # at least 7
-                    else if (depths > 10) f <- floor(depths / 2) # at least 5
+                df <- if (depths > 100) floor(depths / 10) # at least 10
+                    else if (depths > 20) floor(depths / 3) # at least 7
+                    else if (depths > 10) floor(depths / 2) # at least 5
                     else depths
                     oceDebug(getOption("oceDebug"), "df not supplied, so set to ", df, "(note: #depths=", depths, ")\n")
             }
