### 2021-12-10 b631c6d
improve debug msgs
