+++ R/section.R
@@ -357,11 +357,38 @@ setMethod(f="[[",
                   }
               }
               nstation <- length(x@data$station)
-              ## some derived things (not all ... be sure to document when adding things!)
-              ##20160809 if (i %in% c("theta", "potential temperature", "sigmaTheta")) {
-              ##20160809     res <- unlist(lapply(x@data$station, function(ctd) ctd[[i]]))
-              ##20160809     return(res)
-              ##20160809 }
+              if (i == "station") {
+                  if (missing(j)) {    # all stations
+                      res <- x@data$station
+                  } else {             # specified stations
+                      if (is.character(j)) {
+                          nj <- length(j)
+                          stationNames <- unlist(lapply(x@data$station,
+                                  function(x) x@metadata$station))
+                          if (nj == 1) {
+                              w <- which(stationNames == j)
+                              res <- if (length(w)) x@data$station[[w[1]]] else NULL
+                          } else {
+                              res <- vector("list", nj)
+                              for (jj in j) {
+                                  w <- which(stationNames == j)
+                                  res[[jj]] <- if (length(w)) x@data$station[[w[1]]]
+                                      else NULL
+                              }
+                          }
+                      } else {
+                          nj <- length(j)
+                          if (nj == 1) {
+                              res <- x@data$station[[j]]
+                          } else {
+                              res <- vector("list", nj)
+                              for (jj in j)
+                                  res[[jj]] <- x@data$station[[jj]]
+                          }
+                      }
+                  }
+                  return(res)
+              }
               if (i %in% names(x@metadata)) {
                   if (i %in% c("longitude", "latitude")) {
                       if (!missing(j) && j == "byStation") {
@@ -375,12 +402,12 @@ setMethod(f="[[",
                   } else {
                       return(x@metadata[[i]])
                   }
-              } else if (i %in% c("absolute salinity", "CT", "conservative temperature",
+              }
+              if (i %in% c("absolute salinity", "CT", "conservative temperature",
                                   "density", "depth", "nitrite", "nitrate",
                                   "potential temperature", "SA", "sigmaTheta",
                                   "spice", "theta", "z",
                                   names(x@data$station[[1]]@data))) {
-                  ##message('section.R:311 section[["', i, '"]]')
                   if (!missing(j) && substr(j, 1, 4) == "grid") {
                       if (j == "grid:distance-pressure") {
                           numStations <- length(x@data$station)
@@ -424,41 +451,16 @@ setMethod(f="[[",
                       }
                       return(res)
                   }
-              } else if (i == "station") {
-                  if (missing(j)) {
-                      res <- x@data$station
-                  } else {
-                      if (is.character(j)) {
-                          nj <- length(j)
-                          stationNames <- unlist(lapply(x@data$station, function(x) x@metadata$station))
-                          if (nj == 1) {
-                              w <- which(stationNames == j)
-                              res <- if (length(w)) x@data$station[[w[1]]] else NULL
-                          } else {
-                              res <- vector("list", nj)
-                              for (jj in j) {
-                                  w <- which(stationNames == j)
-                                  res[[jj]] <- if (length(w)) x@data$station[[w[1]]] else NULL
-                              }
-                          }
-                      } else {
[…]
