+++ R/ctd.R
@@ -1,4 +1,4 @@
-## vim:textwidth=128:expandtab:shiftwidth=4:softtabstop=4
+## vim:textwidth=80:expandtab:shiftwidth=4:softtabstop=4
 
 #' Class to Store CTD (or general hydrographic) Data
 #'
@@ -98,6 +98,87 @@ setAs("list", "ctd", function(from) {
       as.ctd(from) #salinity=from$salinity, temperature=from$temperature, pressure=from$pressure)
 })
 
+#' Repair a malformed ctd object
+#'
+#' Make a [ctd-class] object adhere more closely with the expected form, e.g. by
+#' moving certain things from the `data` slot to the `metadata` slot, where
+#' other oce functions may assume they will be located.
+#' This can be handy for objects that were set up
+#' incorrectly, perhaps by inappropriate user insertions.
+#'
+#' The possible changes fall into the following categories.
+#'
+#' 1. If unit-length values for `latitude`, `longitude`, `time`, or `station`
+#' exist in the `data` slot, move them to the `metadata` slot.  However,
+#' leave them in `data`  if their length exceeds 1, because this can
+#' arise with towyo data.
+#'
+#' 2. If the `metadata` or `data` slot contains items named
+#' `time`, `recoveryTime`, `startTime`, or `systemUploadTime`,
+#' and if these are not in POSIXt format, then use [as.POSIXct()] with
+#'`tz="UTC"` to convert them to POSIXt format. If that conversion fails,
+#' owing to an unrecognizable format, then the original value
+#' is retained, unaltered.
+#'
+#' @param x a [ctd-class] object.
+#'
+#' @template debugTemplate
+#'
+#' @return A [ctd-class] object that is based on `x`, but possibly
+#' with some elements changed as described in the \dQuote{Details}
+#' section.
+#'
+#' @examples
+#' library(oce)
+#' data(ctd)
+#' # Insert location information into 'data', although it belongs in 'metadata'.
+#' ctd@data$latitude <- ctd@metadata$latitude   # do NOT do this!
+#' ctd@data$longitude <- ctd@metadata$longitude # do NOT do this!
+#' repaired <- repairCtd(ctd)
+#'
+#' @family things related to ctd data
+#'
+#' @author Dan Kelley
+repairCtd <- function(x, debug=getOption("oceDebug"))
+{
+    oceDebug(debug, "repairCtd(x) {\n", sep="", unindent=1, style="bold")
+    dnames <- names(x@data)
+    mnames <- names(x@metadata)
+    # 1: move things from data to metadata
+    for (item in c("longitude", "latitude", "time", "station")) {
+        if (item %in% dnames) {
+            if (length(x@data[[item]]) == 1L) {
+                warning("moving unit-length data$", item, " to metadata$", item, "\n", sep="")
+                x@metadata[[item]] <- x@data[[item]]
+                x@data[[item]] <- NULL
+            }
+        }
+    }
+    # 2: make as POSIXct: time recoveryTime startTime systemUploadTime
+    for (item in c("recoveryTime", "startTime", "systemUploadTime", "time")) {
+        if (item %in% names(x@metadata) && !inherits(x@metadata[[item]], "POSIXt")) {
+            trial <- try(as.POSIXct(x@metadata[[item]], tz="UTC"), silent=TRUE)
+            if (!inherits(trial, "try-error")) {
+                x@metadata[[item]] <- trial
+                warning(paste0("changed metadata$", item, " to a POSIXct value\n"))
+            } else {
+                warning(paste0("cannot change metadata$", item, " to a POSIXct value due to faulty format\n"))
+            }
+        }
[…]
