+++ man/cnvName2oceName.Rd
@@ -260,6 +260,7 @@ Other things related to ctd data:
 \code{\link{read.ctd.woce.other}()},
 \code{\link{read.ctd.woce}()},
 \code{\link{read.ctd}()},
+\code{\link{repairCtd}()},
 \code{\link{setFlags,ctd-method}},
 \code{\link{subset,ctd-method}},
 \code{\link{summary,ctd-method}},
