+++ man/plotScan.Rd
@@ -134,6 +134,7 @@ Other things related to ctd data:
 \code{\link{read.ctd.woce.other}()},
 \code{\link{read.ctd.woce}()},
 \code{\link{read.ctd}()},
+\code{\link{repairCtd}()},
 \code{\link{setFlags,ctd-method}},
 \code{\link{subset,ctd-method}},
 \code{\link{summary,ctd-method}},
