+++ man/read.ctd.woce.other.Rd
@@ -117,6 +117,7 @@ Other things related to ctd data:
 \code{\link{read.ctd.sbe}()},
 \code{\link{read.ctd.woce}()},
 \code{\link{read.ctd}()},
+\code{\link{repairCtd}()},
 \code{\link{setFlags,ctd-method}},
 \code{\link{subset,ctd-method}},
 \code{\link{summary,ctd-method}},
