### 2021-12-15 bd9f7b5
clean up some merge signals
