+++ R/ctd.R
@@ -637,78 +637,94 @@ setMethod(f="summary",
 #' will detect the setup, and subtract nitrite from the sum to yield
 #' nitrate.
 #'
-#' The list given below provides notes on some quantities that are,
-#' or may be, computed.
+#' The list given below provides notes on some quantities that are
+#' available using e.g. `ctd[[i]]`.
 #'
-#' * `conductivity` without a second argument (e.g. `a[["conductivity"]]`)
-#' returns the value stored in the object. However, if a second argument is given,
+#' * If `i` is `"?"`, then the return value is a list
+#' containing four items, each of which is a character vector
+#' holding the names of things that can be accessed with `[[`.
+#' The `data` and `metadata` items hold the names of
+#' entries in the object's data and metadata
+#' slots, respectively. The `dataDerived`
+#' and `metadataDerived` items hold the names of things
+#' that can be inferred from the object's contents, e.g.
+#' `"SA"` is named in `dataDerived`, indicating that
+#' `argo[["SA"]]` is permitted (to compute Absolute Salinity).
+#'
+#' * If `i` is `"conductivity"` without a second argument (e.g. `a[["conductivity"]]`)
+#' then the return value is the seawater electrical conductivity (if available
+#' or computable). However, if a second argument is given,
 #' and it is string specifying a unit, then conversion is made to that unit. The
 #' permitted units are: either `""` or `"ratio"` (for ratio),
 #' `"uS/cm"`, `"mS/cm"` and `"S/m"`. The calculations are based on
 #' the definition of conductivity ratio as the ratio between measured conductivity
 #' and the standard value 4.2914 S/m.
 #'
-#' * `CT` or `Conservative Temperature`: Conservative Temperature,
-#' computed with [gsw::gsw_CT_from_t()].
+#' * If `i` is `"CT"` or `"Conservative Temperature"` then Conservative
+#' Temperature, computed with [gsw::gsw_CT_from_t()], is returned.
 #'
-#' * `density`: seawater density, computed with [swRho]`(x)`.
-#' (Note that it may be better to call that function directly, to gain
+#' * If `i` is `"density"` then seawater density, computed with [swRho]`(x)`,
+#' is returned. (Note that it may be better to call that function directly, to gain
 #' control of the choice of equation of state, etc.)
 #'
-#' * `depth`: Depth in metres below the surface, computed
-#' with [swDepth]`(x)`.
+#' * If `i` is `"depth"` then the depth in metres below the surface, computed
+#' with [swDepth]`(x)`, is returned.
 #'
-#' * `N2`: Square of Brunt-Vaisala frequency, computed  with [swN2]`(x)`.
+#' * If `i` is `"N2"` then the square of Brunt-Vaisala frequency, computed  with
+#' [swN2]`(x)`, is returned.
 #'
-#' * `potential temperature`: Potential temperature in the
-#' UNESCO formulation, computed with [swTheta]`(x)`.
-#' This is a synonym for `theta`.
+#' * If `i` is `"potential temperature"` or `"theta"`, then potential temperature in the
+#' UNESCO formulation, computed with [swTheta]`(x)`, is returned.
 #'
-#' * `Rrho`: Density ratio, computed with [swRrho]`(x)`.
+#' * If `i` is `"Rrho"` then density ratio, computed with [swRrho]`(x)`, is
+#' returned.
 #'
-#' * `SA` or `Absolute Salinity`: Absolute Salinity,
-#' computed with [gsw::gsw_SA_from_SP()].
+#' * If `i` is `"SA"` or `"Absolute Salinity"` then Absolute Salinity,
[…]
