+++ R/argo.R
@@ -1,4 +1,4 @@
-# vim:textwidth=128:expandtab:shiftwidth=4:softtabstop=4:foldmethod=marker
+# vim:textwidth=80:expandtab:shiftwidth=4:softtabstop=4:foldmethod=marker
 #' Class to Store Argo Data
 #'
 #' This class stores data from Argo floats.
@@ -72,18 +72,18 @@ NULL
 #' documentation of which explains (fairly complex) details.
 #'
 #' The results from `argo[[i]]` or `argo[[i,j]]` depend on the
-#' nature of `i` and (if provided) `j`.
-#' The possibilities are as follows.
+#' nature of `i` and (if provided) `j`. The details are as follows.
 #'
-#' * If `i` is `"?"`, then the return value is an
-#' alphabetically sorted character vector that holds
-#' the names of the items that can be returned by `[[`.
-#' This vector is constructed
-#' from the names of items in the `data`
-#' and `metadata` slots, combined with certain computed entries
-#' (e.g. `"Absolute Salinity"`).  Flags and units can also be
-#' retrieved, as explained under the heading \dQuote{Details of the general
-#' method}.
+#' * If `i` is `"?"`, then the return value is a list
+#' containing four items, each of which is a character vector
+#' holding the names of things that can be accessed with `[[`.
+#' The `data` and `metadata` items hold the names of
+#' entries in the object's data and metadata
+#' slots, respectively. The `dataDerived`
+#' and `metadataDerived` items hold the names of things
+#' that can be inferred from the object's contents, e.g.
+#' `"SA"` is named in `dataDerived`, indicating that
+#' `argo[["SA"]]` is permitted (to compute Absolute Salinity).
 #'
 #' * If `i` is `"profile"` and `j` is an integer vector,
 #' then an argo object is returned, as specified by `j`. For example,
@@ -152,16 +152,17 @@ setMethod(f="[[",
               dots <- list(...)
               debug <- if ("debug" %in% names(dots)) dots$debug else 0
               oceDebug(debug, "[[,argo-method(\"", i, "\") {\n", sep="", style="bold", unindent=1)
-              iKnown <- sort(unique(c(names(x@data), names(x@metadata),
-                          "profile",
-                          "CT", "N2", "SA", "sigmaTheta", "theta", "sigma0",
-                          "sigma1", "sigma2", "sigma3", "sigma4",
-                          "z", "depth", "ID", "id", "cycleNumber", "cycle",
-                          "latitude", "longitude",
-                          "*Flag", "*Unit")))
-              if (i == "?") {
-                  return(iKnown)
-              } else if (i == "profile") {
+              metadataDerived <- c("ID", "cycle", "*Flag", "*Unit")
+              dataDerived <- c("profile", "CT", "N2", "SA", "sigmaTheta",
+                  "theta", "sigma0", "sigma1", "sigma2", "sigma3", "sigma4",
+                  "z", "depth", paste("Absolute", "Salinity"),
+                  paste("Conservative", "Temperature"))
+              if (i == "?")
+                  return(list(metadata=sort(names(x@metadata)),
+                          metadataDerived=sort(metadataDerived),
+                          data=sort(names(x@data)),
+                          dataDerived=sort(dataDerived)))
+              if (i == "profile") {
                   ## This assignment to profile is merely to prevent a NOTE from
                   ## the syntax checker. It is needed because of issues with non-standard
                   ## evaluation in subset() calls. This is a problem that many
@@ -175,7 +176,10 @@ setMethod(f="[[",
               }
               namesData <- names(x@data)
               ## handle some computed items
[…]
