### 2021-12-15 bf37ca7
first outline of adding transect function
