+++ R/section.R
@@ -1275,6 +1275,7 @@ setMethod(f="plot",
                               axes=TRUE, mgp, mar,
                               col, cex, pch,
                               labcex=1,
+                              transect=NULL,
                               debug, ...)
           {
               if (missing(debug))
@@ -1370,6 +1371,7 @@ setMethod(f="plot",
                                          legend=TRUE,
                                          debug=0,
                                          axes=TRUE,
+                                         transect=NULL,
                                          col=par("col"),
                                          ...)
               {
@@ -1584,9 +1586,14 @@ setMethod(f="plot",
                       }
 
 
+                      if (class (transect) != "NULL"){  ## FIX - MR -- need to dfine xxrange and yyrange from transect -- test
+                          xxrange <- range (transect@data$coordinates [,1], na.rm = TRUE)
+                          yyrange <- range (transect@data$coordinates [,2], na.rm = TRUE)
+                      }else{
                       ## FIXME: contours don't get to plot edges
                       xxrange <- range(xx, na.rm=TRUE)
                       yyrange <- range(yy, na.rm=TRUE)
+                      }
 
                       ylim <- if (!is.null(ylim)) sort(-abs(ylim)) else yyrange
                       par(xaxs="i", yaxs="i")
@@ -1940,6 +1947,7 @@ setMethod(f="plot",
               zz <- array(NA_real_, dim=c(numStations, num.depths))
               xx <- rep(NA, numStations)
               yy <- rep(NA, num.depths)
+              if (class (transect) == "NULL"){
               if (is.null(at)) {
                   lon0 <- if (missing(longitude0)) mean(firstStation[["longitude"]], na.rm=TRUE) else longitude0
                   lat0 <- if (missing(latitude0)) mean(firstStation[["latitude"]], na.rm=TRUE) else latitude0
@@ -1981,6 +1989,9 @@ setMethod(f="plot",
               } else {
                   xx <- at
               }
+              }else{
+                  ## map xx station locations onto transect distance
+              }
               ##> message("which.xtype: ", which.xtype)
               if (which.xtype == 5) {
                   xx <- numberAsPOSIXct(xx)
@@ -2010,9 +2021,13 @@ setMethod(f="plot",
                                                      }))
                   }
                   ## Map points back to the spine
+                  if (class (transect)!="NULL"){
+                      ## map xx onto distance along transect rather than geoDist
+                  }else{
                   longitudeRemapped <- lonfun(ss[closest])
                   latitudeRemapped <- latfun(ss[closest])
                   xx <- geodDist(longitudeRemapped, latitudeRemapped, alongPath=TRUE)
+                }
               }
               ## Grid is regular (so need only first station) unless which=="data"
               ## FIXME: why checking just first which[] value?
@@ -3303,8 +3318,8 @@ as.section <- function(salinity, temperature, pressure, longitude, latitude, sta
                 if ("waterDepth" %in% names(thelist[[i]]@metadata)) {
                     if (is.na(thelist[[i]]@metadata$waterDepth)) {
                         thelist[[i]]@metadata$waterDepth <- max(thelist[[i]]@data$pressure, na.rm=TRUE)
-                        badDepths <- c(badDepths, i)
-                    }
+                    badDepths <- c(badDepths, i)
+                }
                 } else {
                     thelist[[i]]@metadata$waterDepth <- NA
                 }
