### 2021-12-16 c92c72c
Added decreasing option to sectionSort
