### 2021-12-16 90b7b6e
Corresponding change to documentation, adding the transect argument to plot.section
