### 2021-12-17 060daaa
Revert to basics -- only added option to sectionSort now. Branch to implement transectTemplate.
