+++ R/tides.R
@@ -258,27 +258,41 @@ setMethod(f="summary",
 #'
 #' @section Details of the specialized `tidem` method:
 #'
-#' A vector of the frequencies of fitted constituents is recovered
-#' with e.g. `x[["frequency"]]`. Similarly, amplitude is
-#' recovered with e.g. `x[["amplitude"]]` and phase with
-#' e.g. `x[["phase"]]`. If any other string is specified, then
-#' the underlying accessor \code{\link{[[,oce-method}}) is used.
-#'
 #' @template sub_subTemplate
+#'
+#' @section Details of the specialized `tides` method:
+#'
+#' * If `i` is `"?"`, then the return value is a list
+#' containing four items, each of which is a character vector
+#' holding the names of things that can be accessed with `[[`.
+#' The `data` and `metadata` items hold the names of
+#' entries in the object's data and metadata
+#' slots, respectively. The 
+#' `metadataDerived` holds only `""`, because
+#' no derived metadata values are defined for `cm` objects.
+#'
+#' * If `i` is `"frequency"` or `"freq"`, then a vector of
+#' constituent frequencies (stored
+#' as `freq` in the data slot) is returned.
+#'
+#' * If `i` is `"amplitude"` then a vector of constituent amplitudes
+#' is returned.
+#'
+#' * If `i` is `"phase"` then a vector of constituent phases
+#' is returned.
 #'
 #' @family things related to tides
 setMethod(f="[[",
           signature(x="tidem", i="ANY", j="ANY"),
           definition=function(x, i, j, ...) {
-              if (i == "frequency") {
-                  x@data$freq
-              } else if (i == "amplitude") {
-                  x@data$amplitude
-              } else if (i == "phase") {
-                  x@data$phase
-              } else {
-                  callNextMethod()         # [[
-              }
+              if (i == "?")
+                  return(list(metadata=sort(names(x@metadata)),
+                          metadataDerived="",
+                          data=sort(names(x@data)),
+                          dataDerived="frequency"))
+              if (i == "frequency")
+                  return(x@data$freq)
+              callNextMethod()         # [[
           })
 
 #' Replace Parts of a Tidem Object
