### 2021-12-18 ca57ad1
reorder/improve [[,cm-method doc
