### 2021-12-18 e38662c
reorder/improve [[,tidem-method doc
