+++ man/sub-sub-windrose-method.Rd
@@ -16,7 +16,7 @@
 \item{...}{Optional additional information (ignored).}
 }
 \description{
-\code{[[} provides a way to look up information
+Generally, \code{[[} provides a way to look up information
 within \code{oce} objects, without knowing
 the details of how that information is stored
 within the object. For some object classes,
