### 2021-12-18 61f7edf
rearranging [[ docs
Left to do:
- [ ] amsr.R
- [ ] bremen.R
- [ ] echosounder.R
- [ ] g1sst.R
- [ ] ladp.R
- [ ] landsat.R
- [ ] lisst.R
- [ ] lobo.R
- [ ] met.R
- [ ] odf.R
- [ ] rsk.R
- [ ] satellite.R
- [ ] section.R
- [ ] topo.R
- [ ] windrose.R
- [ ] xbt.R
