### 2021-12-18 df03fac
more [[ doc updates (7 remaining)
Remaining:

- [ ] amsr.R
- [ ] bremen.R
- [ ] g1sst.R
- [ ] ladp.R
- [ ] landsat.R
- [ ] odf.R
- [ ] section.R
