+++ R/coastline.R
@@ -78,7 +78,7 @@ setMethod(f="initialize",
 #' The `data` and `metadata` items hold the names of
 #' entries in the object's data and metadata
 #' slots, respectively. The `dataDerived`
-#' and `metadataDerived` items are each `""`, because
+#' and `metadataDerived` items are each NULL, because
 #' no derived values are defined for [coastline-class] objects.
 #'
 #' * In many cases, the focus will be on the coastline trace
@@ -95,9 +95,9 @@ setMethod(f="[[",
           definition=function(x, i, j, ...) {
               if (i == "?")
                   return(list(metadata=sort(names(x@metadata)),
-                          metadataDerived="",
+                          metadataDerived=NULL,
                           data=sort(names(x@data)),
-                          dataDerived=""))
+                          dataDerived=NULL))
               callNextMethod()         # [[
           })
 
