+++ R/cm.R
@@ -59,8 +59,8 @@ NULL
 #' The `data` and `metadata` items hold the names of
 #' entries in the object's data and metadata
 #' slots, respectively. The `dataDerived`
-#' and `metadataDerived` items are each `""`, because
-#' no derived values are defined by `cm` objects.
+#' and `metadataDerived` items are each NULL, because
+#' no derived values are defined by [cm-class] objects.
 #'
 #' @template sub_subTemplate
 #'
@@ -72,9 +72,9 @@ setMethod(f="[[",
           definition=function(x, i, j, ...) {
               if (i == "?")
                   return(list(metadata=sort(names(x@metadata)),
-                          metadataDerived="",
+                          metadataDerived=NULL,
                           data=sort(names(x@data)),
-                          dataDerived=""))
+                          dataDerived=NULL))
               callNextMethod()         # [[
           })
 
