+++ R/sealevel.R
@@ -215,7 +215,7 @@ setMethod(f="subset",
 #' The `data` and `metadata` items hold the names of
 #' entries in the object's data and metadata
 #' slots, respectively. The `dataDerived`
-#' and `metadataDerived` items are each `""`, because
+#' and `metadataDerived` items are each NULL, because
 #' no derived values are defined by [sealevel-class] objects.
 #'
 #' * In many cases, the focus will be on variations of
@@ -235,9 +235,9 @@ setMethod(f="[[",
           definition=function(x, i, j, ...) {
               if (i == "?")
                   return(list(metadata=sort(names(x@metadata)),
-                          metadataDerived="",
+                          metadataDerived=NULL,
                           data=sort(names(x@data)),
-                          dataDerived=""))
+                          dataDerived=NULL))
               callNextMethod()         # [[
           })
 
