### 2021-12-18 0e842ba
[[ docs: make 'Details' section of uniform title
