+++ man-roxygen/sub_subTemplate.R
@@ -32,6 +32,10 @@
 #'
 #' @section Details of the General Method:
 #'
+#' Note: the text of this section is identical for all `oce` subclasses, and so
+#' some of what you read here may not be relevant to the class being described
+#' in this help page.
+#'
 #' If the specialized method produces no matches, the following generalized
 #' method is applied. As with the specialized method, the procedure hinges first
 #' on the values of `i` and, optionally, `j`. The work proceeds in steps, by
