+++ man/sub-sub-gps-method.Rd
@@ -42,7 +42,7 @@ holding the names of things that can be accessed with \code{[[}.
 The \code{data} and \code{metadata} items hold the names of
 entries in the object's data and metadata
 slots, respectively. The \code{dataDerived}
-and \code{metadataDerived} items are each \code{""}, because
+and \code{metadataDerived} items are each NULL, because
 no derived values are defined by \code{gps} objects.
 \item If \code{i} is \code{"longitude"} or \code{"latitude"}, then the corresponding
 vector is returned.
@@ -54,6 +54,10 @@ with \code{\link[=as.gps]{as.gps()}} with the \code{filename} argument specified
 
 \section{Details of the General Method}{
 
+
+Note: the text of this section is identical for all \code{oce} subclasses, and so
+some of what you read here may not be relevant to the class being described
+in this help page.
 
 If the specialized method produces no matches, the following generalized
 method is applied. As with the specialized method, the procedure hinges first
