+++ R/landsat.R
@@ -256,12 +256,16 @@ setMethod(f="summary",
 #'
 #' @templateVar class landsat
 #'
-#' @template sub_subTemplate
-#'
 #' @section Details of the Specialized Method:
 #'
-#' Users are isolated from the details of the two-byte storage system
-#' by using the `[[` operator.
+#' If `i` is `"?"`, then the return value is a list
+#' containing four items, each of which is a character vector
+#' holding the names of things that can be accessed with `[[`.
+#' The `data` and `metadata` items hold the names of
+#' entries in the object's data and metadata
+#' slots, respectively. The `data` entries are difficult
+#' to deal with directly, and so users are advised to
+#' use `dataDerived` instead.
 #'
 #' *Accessing band data.*  The data may be accessed with e.g.
 #' `landsat[["panchromatic"]]`, for the panchromatic band.  If a new
@@ -312,6 +316,8 @@ setMethod(f="summary",
 #' decimation.  An exception is the lat-lon box, which is altered by
 #' [landsatTrim()].
 #'
+#' @template sub_subTemplate
+#'
 #' @author Dan Kelley
 #'
 #' @family things related to landsat data
@@ -322,9 +328,16 @@ setMethod(f="[[",
               oceDebug(debug, "landsat [[ {\n", unindent=1)
               if (missing(i))
                   stop("Must name a landsat item to retrieve, e.g. '[[\"panchromatic\"]]'", call.=FALSE)
-              i <- i[1]                # drop extras if more than one given
+              if (length(i) > 1L)
+                  stop("[[,landsat-method requires length(i) to be of 1\n")
               if (!is.character(i))
                   stop("landsat item must be specified by name", call.=FALSE)
+              dataDerived <- c("longitude", "latitude", "temperature", "panchromatic")
+              if (i == "?")
+                  return(list(metadata=sort(names(x@metadata)),
+                          metadataDerived=NULL,
+                          data=sort(names(x@data)),
+                          dataDerived=sort(dataDerived)))
               ## Handle cases one by one, starting with simplest.
               if (!(is.na(pmatch(i, "longitude")))) {
                   ## FIXME: ignoring decimation (may be best, anyway)
