+++ R/odf.R
@@ -59,12 +59,29 @@ setMethod(f="initialize",
 #'
 #' @param x an [odf-class] object.
 #'
+#' @section Details of the Specialized Method:
+#'
+#' If `i` is `"?"`, then the return value is a list
+#' containing four items, each of which is a character vector
+#' holding the names of things that can be accessed with `[[`.
+#' The `data` and `metadata` items hold the names of
+#' entries in the object's data and metadata
+#' slots, respectively. The `dataDerived`
+#' and `metadataDerived` items are each NULL, because
+#' no derived values are defined by [odf-class] objects.
+#'
+#'
 #' @template sub_subTemplate
 #'
 #' @family things related to odf data
 setMethod(f="[[",
           signature(x="odf", i="ANY", j="ANY"),
           definition=function(x, i, j, ...) {
+              if (i == "?")
+                  return(list(metadata=sort(names(x@metadata)),
+                          metadataDerived=NULL,
+                          data=sort(names(x@data)),
+                          dataDerived=NULL))
               callNextMethod()         # [[
           })
 
