### 2021-12-21 c87a9f4
add computableWaterProperties()
