+++ R/sw.R
@@ -10,23 +10,28 @@
 #' @param x An [oce-class] object.
 #'
 #' @return A character vector listing the names of computable
-#' water properties.
+#' water properties, or NULL, if there are none.
 #'
 #' @author Dan Kelley
 #'
 #' @family functions that calculate seawater properties
 computableWaterProperties <- function(x)
 {
-    if (inherits(x, "oce")
-        && (3L == sum(c("salinity", "temperature", "pressure")
+    res <- NULL
+    if (inherits(x, "oce") && (3L == sum(c("salinity", "temperature", "pressure")
                 %in% c(names(x@metadata), names(x@data))))) {
-        c("SP", "SR", "Sstar", "N2", "density", "sigmaTheta", paste0("sigma",
-                0:4), "theta", paste("potential", "temperature"), "Rrho",
-            "spice", "spiciness", "SA", paste("Absolute", "Salinity"), "CT",
-            paste("Conservative", "Temperature"), "z", "depth")
-    } else {
-        vector("character", 0L)
+        dataNames <- names(x@data)
+        res <- c("SP", "SR", "Sstar", "N2", "density", "sigmaTheta",
+            paste0("sigma", 0:4), "theta", paste("potential", "temperature"),
+            "Rrho", "spice", "spiciness", "SA", paste("Absolute", "Salinity"),
+            "CT", paste("Conservative", "Temperature"), "z", "depth")
+        # It is possible to compute nitrate from NO2+NO3 and nitrite, if
+        # it's not stored in the object already.
+        if (!("nitrate" %in% dataNames) && ("NO2+NO3" %in% dataNames) && ("nitrite" %in% dataNames))
+            res <- c(res, "nitrate")
+        res <- sort(res) # makes it easier to scan results
     }
+    res
 }
 
 #' Convert from ITS-90 to IPTS-68 temperature
@@ -145,7 +150,9 @@ lookWithin <- function(list)
 #' If the default arguments are acceptable, `ctd[["Rrho"]]` may be used
 #' instead of `swRrho(ctd)`.
 #'
-#' @param ctd an object of class `ctd`
+#' @param ctd an [oce-class] object that holds `salinity`, `temperature`, and
+#' `pressure`.  If `eos` is `"gsw"`, then it must also hold `longitude` and
+#' `latitude`.
 #' @param sense an indication of the sense of double diffusion under study and
 #' therefore of the definition of Rrho; see \sQuote{Details}
 #' @param smoothingLength ignored if `df` supplied, but otherwise the
@@ -168,10 +175,10 @@ lookWithin <- function(list)
 #'
 #' @family functions that calculate seawater properties
 swRrho <- function(ctd, sense=c("diffusive", "finger"), smoothingLength=10, df,
-                   eos=getOption("oceEOS", default="gsw"))
+    eos=getOption("oceEOS", default="gsw"))
 {
-    if (!inherits(ctd, "ctd"))
-        stop("first argument must be of class \"ctd\"")
+    if (!inherits(ctd, "oce"))
+        stop("first argument must be of class \"oce\"")
     sense <- match.arg(sense)
     eos <- match.arg(eos, c("unesco", "gsw"))
     p <- ctd[['pressure']]
@@ -314,15 +321,21 @@ swN2 <- function(pressure, sigmaTheta=NULL, derivs, df,
     oceDebug(debug, "swN2(...) {\n", sep="", unindent=1)
     ##cat("swN2(..., df=", df, ")\n",sep="")
     ##useSmoothing <- !missing(df) && is.finite(df)
-    if (inherits(pressure, "ctd")) {
-        if (!is.null(sigmaTheta))
-            warning("sigmaTheta is ignored, since first argument was a ctd object\n")
[…]
