+++ R/AllClass.R
@@ -1,3 +1,5 @@
+# vim:textwidth=80:expandtab:shiftwidth=4:softtabstop=4
+
 #' Get the present time, in a stated timezone
 #'
 #' @param tz String indicating the desired timezone. The default is
@@ -400,24 +402,106 @@ setMethod(f="[[",
           signature(x="oce", i="ANY", j="ANY"),
           definition=function(x, i, j, ...) {
               ##DEBUG debug <- 1               # DEVELOPER: set manually to test code
-              if (i == "metadata") {
+              metadataNames <- sort(names(x@metadata))
+              dataNames <- sort(names(x@data))
+              if (i == "?")
+                  return(list(metadata=metadataNames,
+                          metadataDerived=NULL,
+                          data=dataNames,
+                          dataDerived=sort(computableWaterProperties(x))))
+              if (i == "metadata")
                   return(x@metadata)
-              } else if (i == "data") {
+              if (i == "data")
                   return(x@data)
-              } else if (i == "processingLog") {
+              if (i == "processingLog")
                   return(x@processingLog)
-              } else if (grepl("Unit$", i)) {
-                  ## returns a list
+              if (grepl("Unit$", i)) # returns a list
                   return(if ("units" %in% names(x@metadata)) x@metadata$units[[gsub("Unit$", "", i)]] else x@metadata[[i]])
-              } else if (grepl(" unit$", i)) {
-                  ## returns just the unit, an expression
+              if (grepl(" unit$", i)) # returns just the unit, an expression
                   return(if ("units" %in% names(x@metadata)) x@metadata$units[[gsub(" unit$", "", i)]][[1]] else "")
-              } else if (grepl(" scale$", i)) {
-                  ## returns just the scale, a character string
+              if (grepl(" scale$", i)) # returns just the scale, a character string
                   return(if ("units" %in% names(x@metadata)) as.character(x@metadata$units[[gsub(" scale$", "", i)]][[2]]) else "")
-              } else if (grepl("Flag$", i)) {
-                  ## returns a list
+              if (grepl("Flag$", i)) # returns a list
                   return(if ("flags" %in% names(x@metadata)) x@metadata$flags[[gsub("Flag$", "", i)]] else NULL)
+              # NOTE: we do not pass data and metadata through directly because e.g. we want
+              # [[ to convert to the modern temperature scale, if the data are
+              # in an old scale.
+              if (i == "conductivity") {
+                  C <- x@data$conductivity
+                  ##message("i=", i, ", j=", if (missing(j)) "(missing)" else j)
+                  if (!is.null(C) && !missing(j)) {
+                      if (!(j %in% c("", "ratio", "uS/cm", "mS/cm", "S/m")))
+                          stop("unknown conductivity unit \"", j, "\"; must be \"\", \"ratio\", \"uS/cm\", \"mS/cm\" or \"S/m\"")
+                      if (j == "")
+                          j <- "ratio" # lets us use switch()
+                      unit <- x@metadata$units$conductivity$unit
+                      if (is.null(unit) || !length(unit)) {
+                          ## FIXME: maybe should look at median value, to make a guess
+                          ## warning("ctd object lack conductivity units; assuming \"ratio\"")
+                          unit <- "ratio"
+                      }
+                      ##message("A")
+                      unit <- as.character(unit)
+                      ##message("next is unit:")
+                      ##print(dput(unit))
+                      C <- x@data$conductivity
+                      ##message("B")
+                      ## Rather than convert from 3 inputs to 3 outputs, express as ratio, then convert as desired
+                      if (!unit %in% c("ratio", "uS/cm", "mS/cm", "S/m"))
+                          stop("object has unknown conductivity unit \"", unit, "\"; must be \"ratio\", \"uS/cm\", \"mS/cm\" or \"S/m\"")
[…]
