+++ tests/testthat/test_hydrographic.R
@@ -4,41 +4,78 @@
 
 library(oce)
 data(ctd)
+
 S <- ctd[["salinity"]]
 T <- ctd[["temperature"]]
 p <- ctd[["pressure"]]
 lon <- ctd[["longitude"]]
 lat <- ctd[["latitude"]]
+nlevel <- length(S)
+
+# Object with no location data (for UNESCO testing)
+nu <- new("oce")
+nu <- oceSetData(nu, "salinity", S)
+nu <- oceSetData(nu, "temperature", T)
+nu <- oceSetData(nu, "pressure", p)
+
+# Object with location data (for GSW testing)
+ng <- new("oce")
+ng <- oceSetData(ng, "salinity", S)
+ng <- oceSetData(ng, "temperature", T)
+ng <- oceSetData(ng, "pressure", p)
+ng <- oceSetMetadata(ng, "longitude", lon)
+ng <- oceSetMetadata(ng, "latitude", lat)
+
 # The two theta values differ by max 4.819825e-6 degC
 thetaUNESCO <- swTheta(S, T, p, eos="unesco")
 thetaGSW <- swTheta(S, T, p, longitude=lon, latitude=lat, eos="gsw")
-CT <- ctd[["CT"]]                      # GSW
-SA <- ctd[["SA"]]                      # GSW
-Sstar <- ctd[["Sstar"]]                # GSW
 
-test_that("base-level [[\"?\"]] works on hydrographic things (UNESCO)", {
+test_that("base-level [[\"?\"]] gives expected names (UNESCO)", {
+    # using sort() because the order (for some reason I do not understand) is
+    # different when I do it interactively, versus in the test as set up by
+    # RStudio.  I guess RStudio is using a different default for sorting
+    # character values.
+    expect_equal(sort(nu[["?"]]$dataDerived), sort(c("density", "depth", "N2",
+                paste("potential", "temperature"), "Rrho", "sigmaTheta", "SP",
+                "spice", "spiciness", "theta", "z")))
+})
+
+test_that("base-level [[\"?\"]] works on some hydrographic things (UNESCO)", {
     options(oceEOS="unesco")
-    n <- new("oce")
-    n <- oceSetData(n, "salinity", S)
-    n <- oceSetData(n, "temperature", T)
-    n <- oceSetData(n, "pressure", p)
-    expect_equal(thetaUNESCO, n[["theta"]])
-    expect_error(n[["CT"]], "object lacks location information")
-    expect_error(n[["SA"]], "object lacks location information")
-    expect_error(n[["Sstar"]], "object lacks location information")
+    expect_equal(ctd[["theta"]], nu[["theta"]])
+    expect_error(nu[["CT"]], "object lacks location information")
+    expect_error(nu[["SA"]], "object lacks location information")
+    expect_error(nu[["Sstar"]], "object lacks location information")
 })
 
-test_that("base-level [[\"?\"]] works on hydrographic things (GSW)", {
-    options(oceEOS="gsw")
-    n <- new("oce")
-    n <- oceSetData(n, "salinity", S)
-    n <- oceSetData(n, "temperature", T)
-    n <- oceSetData(n, "pressure", p)
-    n <- oceSetMetadata(n, "longitude", lon)
-    n <- oceSetMetadata(n, "latitude", lat)
-    expect_equal(thetaGSW, n[["theta"]])
-    expect_equal(CT, n[["CT"]])
-    expect_equal(SA, n[["SA"]])
-    expect_equal(Sstar, n[["Sstar"]])
+
+test_that("base-level [[\"?\"]] gives expected names (GSW)", {
[…]
