### 2021-12-23 a1f7b23
as.ctd() obeys longitude and latitude [TESTING]
