### 2021-12-24 be8ca6b
as.ctd() handles location better if arg1 is list
See https://github.com/dankelley/oce/issues/1898
