+++ R/ctd.R
@@ -1027,69 +1027,69 @@ setMethod(f="[[<-",
 
 #' Coerce data into CTD object
 #'
-#' Assemble data into a [ctd-class] object.
+#' Assemble data into a [ctd-class] object.  This function is complicated
+#' (spanning approximately 500 lines of code) because it tries to handle many
+#' special cases, and tries to make sensible defaults for unspecified
+#' parameters.  If odd results are found, users might find it helpful to call
+#' this function with the first argument being a simple vector of Practical
+#' Salinity values, in which case the processing of the other arguments is
+#' relatively straightforward.
 #'
-#' @param salinity There are several distinct choices for `salinity`.
+#' @param salinity a numeric vector holding Practical Salinity, or something
+#' else from which it and other hydrographic variables can be inferred.  The
+#' choices are as follows.  (As noted in the \dQuote{Description}, the simplest
+#' case is the first, and users should try this, if other cases yield unexpected
+#' results.)
 #'
-#' * It can be a
-#' vector indicating the practical salinity through the water column. In that case,
-#' `as.ctd` employs the other arguments listed below.
+#' * It can be a vector holding values of Practical Salinity through the water
+#' column. In this case, `as.ctd()` uses the other function parameters
+#' (`temperature`, `pressure` etc) as specified.
+#'
+#' * It can be a [ctd-class] object, in which case `as.ctd()` simply returns
+#' it, unaltered.
 #'
 #' * It can be an [rsk-class] object (see \dQuote{Converting rsk objects} for details).
 #'
-#' * It can be something (a data frame, a list or an `oce` object)
-#' from which practical salinity, temperature, pressure, and conductivity
-#' can be inferred. In this case, the relevant information
-#' is extracted  and the other arguments to `as.ctd` are ignored, except for
-#' `pressureAtmospheric`. If the first argument has salinity, etc., in
-#' matrix form (as can happen with some objects of [argo-class]),
-#' then only the first column is used, and a warning to that effect is given,
-#' unless the `profile` argument is specified and then that specific
-#' profile is extracted.
+#' * It can be something (a data frame, a list or an `oce` object) from which
+#' Practical Salinity, in-situ temperature, sea pressure, and possibly
+#' conductivity) can be inferred. In this case, the relevant information is
+#' extracted  and the other arguments to `as.ctd` are ignored, except for
+#' `longitude`, `latitude` and `pressureAtmospheric`, each of which must be
+#' specified.  If the first argument has salinity, etc., in matrix form (as can
+#' happen with some objects of [argo-class]), then only the first column is
+#' used, and a warning to that effect is given, unless the `profile` argument is
+#' specified and then that specific profile is extracted.
 #'
 #' * It can be unspecified, in which case `conductivity` becomes a mandatory
-#' argument, because it will be needed for computing actual salinity,
-#' using [swSCTp()].
+#' argument, because it will be needed for computing actual salinity, using
+#' [swSCTp()].
 #'
-#' @param temperature *in-situ* temperature in \eqn{^\circ}{deg}C
-#' on the ITS-90 scale; see \dQuote{Temperature units} in the documentation for
-#' [swRho()].
+#' @param temperature a numeric vector containing *in-situ* temperature in
+#' \eqn{^\circ}{deg}C on the ITS-90 scale; see \dQuote{Temperature units} in the
+#' documentation for [swRho()].
 #'
-#' @param pressure Vector of pressure values, one for each `salinity` and
-#' `temperature` pair, or just a single pressure, which is repeated to match
[…]
