+++ tests/testthat/test_ctd.R
@@ -1,6 +1,21 @@
 # vim:textwidth=80:expandtab:shiftwidth=4:softtabstop=4
 library(oce)
 
+# https://github.com/dankelley/oce/issues/1898
+test_that("as.ctd() handles location well (issue 1898)", {
+    data(ctd)
+    l <- list(salinity=ctd[["salinity"]],
+        temperature=ctd[["temperature"]],
+        pressure=ctd[["pressure"]])
+    ctd1 <- as.ctd(l)
+    expect_false("longitude" %in% names(ctd1@metadata))
+    expect_false("latitude" %in% names(ctd1@metadata))
+    ctd2 <- as.ctd(l, longitude=-60, latitude=45)
+    expect_true("longitude" %in% names(ctd2@metadata))
+    expect_true("latitude" %in% names(ctd2@metadata))
+})
+
+
 test_that("ctdRepair() works as expected", {
     data(ctd)
     expect_warning(ctd <- ctdRepair(ctd), "^changed metadata\\$recoveryTime")
@@ -23,6 +38,7 @@ test_that("plotTS() handles differently EOSs correctly", {
 })
 
 test_that("as.ctd() with specified arguments, including salinity", {
+    data(ctd)
     ctd_ctd <- as.ctd(salinity=ctd[["salinity"]], temperature=ctd[["temperature"]], pressure=ctd[["pressure"]])
     expect_equal(ctd[["salinity"]], ctd_ctd[["salinity"]])
     expect_equal(ctd[["temperature"]], ctd_ctd[["temperature"]])
@@ -56,44 +72,41 @@ test_that("as.ctd() with a list of oce objects", {
     expect_equal(length(A[["longitude"]]), length(A[["pressure"]]))
 })
 
-
-test_that("ctd[[\"CT\"]] requires lon and lat", {
+test_that("ctd[[\"CT\"]], ctd[[\"SA\"]] and ctd[[\"Sstar\"]] require location", {
     a <- as.ctd(35,10,0)
     expect_error(a[["CT"]])
-})
-
-test_that("ctd[[\"Sstar\"]] requires lon and lat", {
-    a <- as.ctd(35,10,0)
-    expect_error(a[["Sstar"]])
-})
-
-test_that("ctd[[\"SA\"]] requires lon and lat", {
-    a <- as.ctd(35,10,0)
     expect_error(a[["SA"]])
+    expect_error(a[["Sstar"]])
+    b <- as.ctd(35,10,0,longitude=0, latitude=0)
+    expect_silent(b[["CT"]])
+    expect_silent(b[["SA"]])
+    expect_silent(b[["Sstar"]])
 })
+
 
 test_that("as.ctd() with specified arguments, not including salinity", {
+    data(ctd)
     salinity <- ctd[["salinity"]]
     temperature <- ctd[["temperature"]]
     pressure <- ctd[["pressure"]]
     conductivity <- swCSTp(salinity, temperature, pressure)
     options(oceEOS="unesco")
     ctdNew <- as.ctd(conductivity=conductivity, temperature=temperature, pressure=pressure)
-    ## Test that all fields were created accurately.
+    # Test that all fields were created accurately.
     expect_equal(salinity, ctdNew[["salinity"]])
     expect_equal(temperature, ctdNew[["temperature"]])
     expect_equal(pressure, ctdNew[["pressure"]])
     ##
     options(oceEOS="gsw")
     ctdNew <- as.ctd(conductivity=conductivity, temperature=temperature, pressure=pressure)
-    ## Test that all fields were created accurately.
+    # Test that all fields were created accurately.
     expect_equal(salinity, ctdNew[["salinity"]])
[…]

[…]
