### 2021-12-30 5d2f431
broken (checking in at end of day)
The units are not showing up.  They were very recently and I don't see
what I broke.  I put in comments that demonstrate that my method ought
to work.
