+++ R/section.R
@@ -1019,8 +1019,8 @@ sectionAddCtd <- sectionAddStation
 #' `which`.
 #'
 #' The type of plot is governed by `which`, as follows.
-#' * `which=0` or `"potential temperature"` for temperature contours
-#' * `which=1` or `"temperature"` for temperature contours (the default)
+#' * `which=0` or `"potential temperature"` for potential temperature contours
+#' * `which=1` or `"temperature"` for in-situ temperature contours (the default)
 #' * `which=2` or `"salinity"` for salinity contours
 #' * `which=3` or `"sigmaTheta"` for sigma-theta contours
 #' * `which=4` or `"nitrate"` for nitrate concentration contours
@@ -1137,6 +1137,10 @@ sectionAddCtd <- sectionAddStation
 #'
 #' @param legend.loc Location of legend, as supplied to [legend()], or
 #' set to the empty string to avoid plotting a legend.
+#'
+#' @param legend.text character value indicating the text for the legend.
+#' If this is NULL (the default) then the legend is automatically
+#' constructed, based on the value of `which`.
 #'
 #' @param showStations Logical indicating whether to draw station numbers on maps.
 #'
@@ -1305,6 +1309,7 @@ setMethod(f="plot",
                               xtype="distance", ytype="depth", ztype="contour",
                               longitude0, latitude0,
                               legend.loc="bottomright",
+                              legend.text=NULL,
                               showStations=FALSE,
                               showStart=TRUE,
                               stationTicks=TRUE,
@@ -1339,10 +1344,64 @@ setMethod(f="plot",
                   cex <- par("cex")
               if (missing(pch))
                   pch <- par("pch")
+              ## L and R are used much later, for constructing labels
+              L <- if (getOption("oceUnitBracket") == "[") " [" else " ("
+              R <- if (getOption("oceUnitBracket") == "[")  "]" else  ")"
 
-              ## Make 'which' be numeric, to simplify following code
+              # Internal function to create z label with units (if known) Ignore
+              # the value of 'unit' for some things, because modern convention
+              # dictates the unit.  Besides, a user can set legend.text if a
+              # custom unit is required.
+              # FIXME: do we need 'eos'?
+              vtitleCreate <- function(zname, eos=NULL, unit=NULL)
+              {
+                  UNIT<<-unit
+                  u <- as.character(unit$unit)
+                  print(unit)
+                  u <- unit
+                  L <- if (getOption("oceUnitBracket") == "[") " [" else " ("
+                  R <- if (getOption("oceUnitBracket") == "[")  "]" else  ")"
+                  # Note that the code is alphabetical in the first item of
+                  # equivalents. Please follow that convention if adding new
+                  # entries. Also, use paste() to combine words, to prevent
+                  # text-editor reflow operations from inserting line breaks.
+                  if (zname %in% c(paste("Absolute", "Salinity"), "SA")) {
+                      rval <- bquote(S[A]*.(L)*g/kg*.(R))
+                  } else if (zname %in% c(paste("Conservative", "Temperature"), "CT")) {
+                      rval <- bquote(Theta*.(L)*degree*C*.(R))
+                  } else if (zname %in% c("nitrate", "NO3")) {
+                      rval <- if (is.null(unit)) bquote(NO[3]) else bquote(NO[3]*.(L)*.(u)*.(R))
+                  } else if (zname %in% c("nitrite", "NO2")) {
+                      rval <- if (is.null(unit)) bquote(NO[2]) else bquote(NO[2]*.(L)*.(u)*.(R))
+                  } else if (zname == "NO2+NO3") {
[…]
