+++ NEWS.md
@@ -10,6 +10,7 @@
 * Change `as.ctd()` to drop `other` parameter, deprecated 4y ago.
 * Change `read.odf()` to handle new CODEs and UNITs.
 * Change `oce.plot.ts()` by adding `simplify` argument.
+* Change `plot,section-method()` temperature label to "T".
 * Change `pwelch()`, improving low-frequency results.
 * Fix `[[` to yield ITS90 temperature for all classes, not just `ctd`.
 * Fix `as.ctd()` handling of temperature scale of first argument.
