### 2021-12-31 4af327b
improve plot(section) for unspecified contours
Now, there is a single block of code for all variables.  I have left the
old code in commented-out form, for a while, until I do the case of
specified contours.
