+++ R/misc.R
@@ -223,6 +223,125 @@ argShow <- function(x, nshow=4, last=FALSE, sep="=")
     res
 }
 
+#' Create label with unit
+#'
+#' `labelWithUnit` creates a label with a unit, for graphical
+#' display, e.g. by \code{\link{plot,section-method}}.
+#' The unit is enclosed in square brackets, although setting
+#' `options(oceUnitBracket="(")` will cause parentheses to be
+#' used, instead.
+#'
+#' If `name` is in a standard list, then alterations are made as appropriate,
+#' e.g. `"SA"` or `"Absolute Salinity"` yields an S with subscript A; `"CT"` or
+#' `"Conservative Temperature"` yields an upper-case Theta, `sigmaTheta`
+#' yields a sigma with subscript theta, `sigma0` yields
+#' sigma with subscript 0 (with similar for 1 through 4), `"N2"` yields "N" with
+#' superscript 2, and `"pressure"` yields "p".
+#' These basic hydrographic quantities have default units that will
+#' be used if `unit` is not supplied (see \dQuote{Examples}).
+#'
+#' In addition to the above, several chemical names are recognized,
+#' but no unit is guessed for them, because the oceanographic
+#' community lacks agreed-upon standards.
+#'
+#' If `name` is not recognized, then it is simply repeated in the
+#' return value.
+#'
+#' @param name character value naming a quantity.
+#'
+#' @param unit a list containing items `unit` and (optionally) `scale`, only the
+#' first of which, an [expression()], is used.  If `unit` is not provided,
+#' then a default will be used (see \dQuote{Details}).
+#'
+#' @return `labelWithUnit` returns a language object, created with [bquote()],
+#' that that may supplied as a text string to [legend()], [mtext()], [text()],
+#' etc.
+#'
+#' @examples
+#' library(oce)
+#' # 1. temperature has a predefined unit, but this can be overruled
+#' labelWithUnit("temperature")
+#' labelWithUnit("temperature",
+#'     list(unit=expression(m/s), scale="erroneous"))
+#' # 2. phosphate lacks a predefined unit
+#' labelWithUnit("phosphate")
+#' data(section)
+#' labelWithUnit("phosphate",
+#'     section[["station",1]][["phosphateUnit"]])
+#'
+#' @family functions that create labels
+#'
+#' @author Dan Kelley
+labelWithUnit <- function(name, unit=NULL)
+{
+    u <- if (!is.null(unit) && length(unit$unit) > 0L) unit$unit[[1]] else "unitless"
+    L <- if (getOption("oceUnitBracket", "[") == "[") " [" else " ("
+    R <- if (getOption("oceUnitBracket", "[") == "[")  "]" else  ")"
+    # Note that the code is alphabetical in the first item of
+    # equivalents. Please follow that convention if adding new
+    # entries. Also, use paste() to combine words, to prevent
+    # text-editor reflow operations from inserting line breaks.
+    #
+    # There's no need to handle unitless quantities that don't
+    # need renaming, since they are handled by the final else.
+    if (name %in% c(paste("Absolute", "Salinity"), "SA")) {
[…]
