+++ NAMESPACE
@@ -155,6 +155,7 @@ export(abbreviateTimeLabels,
        is.ad2cp,
        julianDay,
        julianCenturyAnomaly,
+       labelWithUnit,
        landsatAdd,
        landsatTrim,
        latFormat,

[…]
