### 2022-01-02 bbaadfa
new density tests for issue 1904
