### 2022-01-02 74386a5
start work on 1906 (TO DO...)
1. eliminate special cases (start with 10)
2. recognize 'oxygen'
3. recognize 'potential temperature'

1.
