+++ R/misc.R
@@ -2676,29 +2676,31 @@ fullFilename <- function(filename)
 #' @family functions that create labels
 #'
 #' @author Dan Kelley
-resizableLabel <- function(item, axis="x", sep, unit=NULL, debug=getOption("oceDebug"))
+resizableLabel <- function(item, axis="x", sep, unit=NULL,
+    debug=getOption("oceDebug"))
 {
     oceDebug(debug, "resizableLabel(item=\"", item,
-             "\", axis=\"", axis,
-             "\", sep=\"", if (missing(sep)) "(missing)" else sep, "\", ...) {\n",
-            sep="", unindent=1, style="bold")
+        "\", axis=\"", axis,
+        "\", sep=\"", if (missing(sep)) "(missing)" else sep, "\", ...) {\n",
+        sep="", unindent=1, style="bold")
     if (missing(item))
         stop("must provide 'item'")
     if (axis != "x" && axis != "y")
         stop("axis must be \"x\" or \"y\"")
-    itemAllowed <- c("S", "SA", "C", "CT", "conductivity mS/cm", "conductivity S/m", "T",
-                     "theta", "sigmaTheta", "conservative temperature",
-                     "absolute salinity", "N2", "nitrate", "nitrite",
-                     "oxygen", "oxygen saturation", "oxygen mL/L", "oxygen umol/L", "oxygen umol/kg",
-                     "phosphate", "silicate", "tritium", "spice",
-                     "fluorescence", "p", "z", "distance", "distance km",
-                     "along-spine distance km",
-                     "along-track distance km",
-                     "heading", "pitch", "roll", "u",
-                     "v", "w", "speed", "direction", "eastward", "northward",
-                     "depth", "elevation", "latitude", "longitude", "frequency cph",
-                     "sound speed", "spectral density m2/cph",
-                     "sigma0", "sigma1", "sigma2", "sigma3", "sigma4")
+    itemAllowed <- c("S", "SA", "C", "CT", paste("conductivity", "mS/cm"),
+        paste("conductivity", "S/m"), "T", "theta", "sigmaTheta",
+        paste("Conservative", "Temperature"), paste("Absolute", "Salinity"),
+        "N2", "nitrate", "nitrite", "oxygen", paste("oxygen", "saturation"),
+        paste("oxygen", "mL/L"), paste("oxygen", "umol/L"), paste("oxygen",
+            "umol/kg"), "phosphate", "silicate", "tritium", "spice",
+        "fluorescence", "p", "z", "distance", "distance km",
+        paste("along-spine", "distance", "km"), paste("along-track", "distance",
+            "km"), "heading", "pitch", "roll", "u", "v", "w", "speed",
+        "direction", "eastward", "northward", "depth", "elevation", "latitude",
+        "longitude", paste("frequency", "cph"), paste("sound", "speed"),
+        paste("spectral", "density", "m2/cph"), "sigma0", "sigma1", "sigma2",
+        "sigma3", "sigma4", "Sstar", "SR")
+
     ## FIXME: if anything is added, run the next, and paste results into roxygen.
     ## > A<-paste0("'",paste(sort(itemAllowed), collapse="'`, `'"),"'");A
     ## NOTE: some hand-tweaking must be done to fix linebreaks and (preferably) to
@@ -2738,11 +2740,11 @@ resizableLabel <- function(item, axis="x", sep, unit=NULL, debug=getOption("oceD
[…]
