### 2022-01-02 42337d9
plot,ctd-method obeys xtype="salinity"
Previously, it had a complicated behaviour, switching to SA if
gsw=="eos".  But now it simply obeys users' intentions.
