### 2022-01-03 745175a
improve data(section) (1907 and 1908)
https://github.com/dankelley/oce/issues/1907
https://github.com/dankelley/oce/issues/1908
