### 2022-01-04 9fe86a1
large refactor of plot,ctd-method
This lets users plot profiles of anything stored in the object's data
slot, of things computable from such contents, and also from the
previous-documented possibilities.  (The docs do not yet discuss this
expansion.)

Quite a lot of copy/paste/edit was done for this so one might expect
problems.  Below is the (very limited) testing that I did as I was
refactoring.

library(oce)
data(ctd)
debug <- 0
plot(ctd, which="CT", debug=debug)
plot(ctd, which="CT", type="l", Tlim=c(8,14), debug=debug)
plot(ctd, which="index", type="l", debug=debug)
plot(ctd, which="text"
[…]
