+++ R/ctd.R
@@ -2923,13 +2923,13 @@ write.ctd <- function(object, file, metadata=TRUE, flags=TRUE, format="csv")
 #'
 #' * `which=21` or `which="SA"` gives an Absolute Salinity profile
 #'
-#' * `which=30` gives a time series of Salinity
+#' * `which=30` or `which="Sts"` gives a time series of Salinity
 #'
-#' * `which=31` gives a time series of Temperature
+#' * `which=31` or `which="Tts"` gives a time series of Temperature
 #'
-#' * `which=32` gives a time series of pressure
+#' * `which=32` or `which="pts"` gives a time series of pressure
 #'
-#' * `which=33` gives a time series of sigmaTheta
+#' * `which=33` or `which="rhots"` gives a time series of sigmaTheta
 #'
 #' @param col Color of lines or symbols.
 #'
@@ -3137,724 +3137,656 @@ write.ctd <- function(object, file, metadata=TRUE, flags=TRUE, format="csv")
 #'
 #' @aliases plot.ctd
 setMethod(f="plot",
-          signature=signature("ctd"),
-          definition=function(x, which,
-                              col=par("fg"),
-                              fill, # to catch old method
-                              borderCoastline=NA, colCoastline="lightgray",
-                              eos=getOption("oceEOS", default='gsw'),
-                              ref.lat=NaN, ref.lon=NaN,
-                              grid=TRUE, coastline="best",
-                              Slim, Clim, Tlim, plim, densitylim, N2lim, Rrholim,
-                              dpdtlim, timelim,
-                              lonlim, latlim, # FIXME: maybe should be deprecated 2014-01-07
-                              drawIsobaths=FALSE, clongitude, clatitude, span, showHemi=TRUE,
-                              lonlabels=TRUE, latlabels=TRUE,
-                              projection=NULL,
-                              latlon.pch=20, latlon.cex=1.5, latlon.col="red",
-                              cex=1, cex.axis=par('cex.axis'),
-                              pch=1,
-                              useSmoothScatter=FALSE,
-                              df,
-                              keepNA=FALSE,
-                              type,
-                              mgp=getOption("oceMgp"),
-                              mar=c(mgp[1]+1.5, mgp[1]+1.5, mgp[1]+1.5, mgp[1]+1),
-                              inset=FALSE,
-                              add=FALSE,
-                              debug=getOption("oceDebug"),
-                              ...)
-          {
-              if (!inherits(x, "ctd"))
-                  stop("method is only for objects of class 'ctd'")
-              eos <- match.arg(eos, c("unesco", "gsw"))
-              if (!missing(fill)) {
-                  ## permit call as documented before 2016-02-03
-                  ## Note: the code permitted fill=TRUE but this was never documented
-                  if (is.character(fill)) {
-                      colCoastline <- fill
-                  } else {
-                      if (is.logical(fill) && !fill) {
-                          colCoastline <- NULL
-                      }
-                  }
-                  warning("In plot,ctd-method() : 'fill' is outdated; use 'colCoastline' instead", call.=FALSE)
-              }
-              if (missing(which)) {
-                  oceDebug(debug, "plot,ctd-method(..., eos=\"", eos, "\", inset=", inset, ", ...) {\n", sep="", unindent=1)
-                  dt <- x@metadata$deploymentType
-                  if (is.null(dt)) {
-                      which <- c(1, 2, 3, 5)
-                      if (missing(type))
-                          type <- c("l", "l", "p", "p")
-                  } else {
-                      types <- c("profile", "moored", "thermosalinograph", "tsg", "towyo")
-                      itype <- pmatch(dt, types, nomatch=0)
-                      if (itype == 0) {
[…]
