### 2022-01-04 ad7e755
plot,ctd-method docs improved; remove defunct args
1. Combine some related arguments.
2. Clarify what happens if eos is switched.
3. Remove lonlim and latlim, which were marked defunct in 2014. This
   change is documeted in the History section.
