+++ R/ctd.R
@@ -962,34 +962,12 @@ setMethod(f="[[<-",
 #' Salinity values, in which case the processing of the other arguments is
 #' relatively straightforward.
 #'
-#' @param salinity a numeric vector holding Practical Salinity, or something
-#' else from which it and other hydrographic variables can be inferred.  The
-#' choices are as follows.  (As noted in the \dQuote{Description}, the simplest
-#' case is the first, and users should try this, if other cases yield unexpected
-#' results.)
-#'
-#' * It can be a vector holding values of Practical Salinity through the water
-#' column. In this case, `as.ctd()` uses the other function parameters
-#' (`temperature`, `pressure` etc) as specified.
-#'
-#' * It can be a [ctd-class] object, in which case `as.ctd()` simply returns
-#' it, unaltered.
-#'
-#' * It can be an [rsk-class] object (see \dQuote{Converting rsk objects} for details).
-#'
-#' * It can be something (a data frame, a list or an `oce` object) from which
-#' Practical Salinity, in-situ temperature, sea pressure, and possibly
-#' conductivity) can be inferred. In this case, the relevant information is
-#' extracted  and the other arguments to `as.ctd` are ignored, except for
-#' `longitude`, `latitude` and `pressureAtmospheric`, each of which must be
-#' specified.  If the first argument has salinity, etc., in matrix form (as can
-#' happen with some objects of [argo-class]), then only the first column is
-#' used, and a warning to that effect is given, unless the `profile` argument is
-#' specified and then that specific profile is extracted.
-#'
-#' * It can be unspecified, in which case `conductivity` becomes a mandatory
-#' argument, because it will be needed for computing actual salinity, using
-#' [swSCTp()].
+#' @param salinity may be (1) a numeric vector holding Practical Salinity,
+#' (2) a list or data frame holding `salinity` and other
+#' hydrographic variables or (3) an `oce-class` object that holds
+#' hydrographic information. If `salinity` is not provided,
+#' then `conductivity` must be provided, so that [swSCTp()]
+#' can be used to compute salinity.
 #'
 #' @param temperature a numeric vector containing *in-situ* temperature in
 #' \eqn{^\circ}{deg}C on the ITS-90 scale; see \dQuote{Temperature units} in the
@@ -2835,13 +2813,17 @@ write.ctd <- function(object, file, metadata=TRUE, flags=TRUE, format="csv")
 
 #' Plot a ctd Object
 #'
-#' Plot CTD data, by default in a four-panel display showing (a) profiles of
-#' salinity and temperature, (b) profiles of density and the square of buoyancy
-#' frequency, (c) a TS diagram and (d) a coastline diagram indicating the station
-#' location.
+#' Plot CTD data in any of many different ways. In many cases,
+#' the best choice is to use default values for all parameters
+#' other than the first.  This yields a 4-panel plot that
+#' displays a basic overview of the data, with a combined
+#' profile of salinity and temperature at the top
+#' left, a combined plot of density and the square of buoyancy
+#' frequency at top right, a TS diagram at bottom left, and
+#' a map at bottom right.
 #'
 #' @details
-#' Creates a multi-panel summary plot of data measured in a CTD cast. The
+#' The
 #' default values of `which` and other arguments are chosen to be useful
 #' for quick overviews of data. However, for detailed work it is common
 #' to call the present function with just a single value of `which`, e.g.
[…]
