+++ man/plot-ctd-method.Rd
@@ -26,8 +26,6 @@
   Rrholim,
   dpdtlim,
   timelim,
-  lonlim,
-  latlim,
   drawIsobaths = FALSE,
   clongitude,
   clatitude,
@@ -35,10 +33,10 @@
   showHemi = TRUE,
   lonlabels = TRUE,
   latlabels = TRUE,
-  projection = NULL,
   latlon.pch = 20,
   latlon.cex = 1.5,
   latlon.col = "red",
+  projection = NULL,
   cex = 1,
   cex.axis = par("cex.axis"),
   pch = 1,
@@ -58,21 +56,34 @@
 \item{x}{a \linkS4class{ctd} object.}
 
 \item{which}{List of desired plot types, as given below. If \code{which} is not
-supplied, a default will be used. This default will be \code{c(1,2,3,5)} if the
-CTD is in profiling mode (i.e. if \code{deploymentType} in the \code{metadata}
-slot equals \code{"profile"},
-or is missing) or
-\code{"moored"}/\code{"thermosalinograph"}, the default will be \code{c(30, 3, 31, 5)}.  If it
-is \code{"towyo"}, \code{c(30, 31, 32, 3)} will be used. Details are as
-follows.
+supplied, a default will be used. This default depends on \code{deploymentType}
+in the \code{metadata} slot of \code{x}.  If \code{deploymentType} is
+\code{"profile"} or missing, then \code{which} defaults to
+\code{c(1,2,3,5)}.  If \code{deploymentType} is
+\code{"moored"} or \code{"thermosalinograph"} then \code{which} defaults to
+\code{c(30, 3, 31, 5)}. Finally, if \code{deploymentType} is \code{towyo} then
+\code{which} defaults to \code{c(30, 31, 32, 3)}.
+
+The details of individual \code{which} values are as follows. Some of the entries
+refer to the EOS (equation of state for seawater), which may either \code{"gsw"}
+for the modern Gibbs Seawater system, or \code{"unesco"} for the older UNESCO
+system. The EOS may be set with the \code{eos} argument, or by using \code{\link[=options]{options()}},
+with \code{options(oceEOS="unesco")} or \code{options(oceEOS="unesco")}. The default EOS
+is \code{"gsw"}.
 \itemize{
-\item \code{which=1} or \code{which="salinity+temperature"} gives
-a combined profile of temperature and salinity
-\item \code{which=2} or \code{which="density+N2"} gives a combined
-profile of \eqn{\sigma_\theta}{sigma-theta} and \eqn{N^2}{N^2}
-\item \code{which=3} or \code{which="TS"} gives a TS plot
-\item \code{which=4} or \code{which="text"} gives a textual
-summary of some aspects of the data
+\item \code{which=1} or \code{which="salinity+temperature"} gives a combined profile of
+temperature and salinity.  If the EOS is \code{"gsw"} then Conservative
+Temperature and Absolute Salinity will be shown; otherwise in-situ
+temperature and practical salinity will be shown.
+\item \code{which=2} or \code{which="density+N2"} gives a combined profile of density
+anomaly, computed with \code{\link[=swSigma0]{swSigma0()}}, along with the square of the buoyancy
+frequency, computed with \code{\link[=swN2]{swN2()}}.  Note that the \code{eos} parameter is passed
+to each of these functions.
+\item \code{which=3} or \code{which="TS"} gives a TS plot. If the EOS is \code{"gsw"}, T is
+Conservative Temperature and S is Absolute Salinity; otherwise, they are
+in-situ temperature and practical salinity, respectively.
+\item \code{which=4} or \code{which="text"} gives a textual  summary of
+some aspects of the data.
 \item \code{which=5} or \code{which="map"} gives a map plotted
[…]
