+++ R/ctd.R
@@ -748,11 +748,6 @@ setMethod(f="[[",
               # https://github.com/dankelley/oce/issues/1891
               # Use paste() for two-word items so a text editor won't break strings
               metadataDerived <- c("time", "*Flag", "*Unit")
-              #>dataDerived <- c("SP", "SR", "Sstar", "N2", "density",
-              #>    "sigmaTheta", "sigma0", "sigma1", "sigma2", "sigma3",
-              #>    "sigma4", "theta", paste("potential", "temperature"), "Rrho",
-              #>    "spice", "spiciness", "SA", paste("Absolute", "Salinity"),
-              #>    "CT", paste("Conservative", "Temperature"), "z", "depth")
               if (i == "?")
                   return(list(metadata=sort(names(x@metadata)),
                           metadataDerived=sort(metadataDerived),
@@ -2841,31 +2836,30 @@ write.ctd <- function(object, file, metadata=TRUE, flags=TRUE, format="csv")
 #'
 #' @param x a [ctd-class] object.
 #'
-#' @param which List of desired plot types, as given below. If `which` is not
-#' supplied, a default will be used. This default depends on `deploymentType`
-#' in the `metadata` slot of `x`.  If `deploymentType` is
-#' `"profile"` or missing, then `which` defaults to
-#' `c(1,2,3,5)`.  If `deploymentType` is
-#' `"moored"` or `"thermosalinograph"` then `which` defaults to
-#' `c(30, 3, 31, 5)`. Finally, if `deploymentType` is `towyo` then
-#' `which` defaults to `c(30, 31, 32, 3)`.
+#' @param which a numeric or character vector specifying desired plot types. If
+#' `which` is not supplied, a default will be used. This default depends on
+#' `deploymentType` in the `metadata` slot of `x`.  If `deploymentType` is
+#' `"profile"` or missing, then `which` defaults to `c(1,2,3,5)`.  If
+#' `deploymentType` is `"moored"` or `"thermosalinograph"` then `which` defaults
+#' to `c(30, 3, 31, 5)`. Finally, if `deploymentType` is `towyo` then `which`
+#' defaults to `c(30, 31, 32, 3)`.
 #'
 #' The details of individual `which` values are as follows. Some of the entries
 #' refer to the EOS (equation of state for seawater), which may either `"gsw"`
 #' for the modern Gibbs Seawater system, or `"unesco"` for the older UNESCO
-#' system. The EOS may be set with the `eos` argument, or by using [options()],
-#' with `options(oceEOS="unesco")` or `options(oceEOS="unesco")`. The default EOS
-#' is `"gsw"`.
+#' system. The EOS may be set with the `eos` argument to `plot,ctd-method()` or
+#' by using [options()], with `options(oceEOS="unesco")` or
+#' `options(oceEOS="unesco")`. The default EOS is `"gsw"`.
 #'
 #' * `which=1` or `which="salinity+temperature"` gives a combined profile of
 #' temperature and salinity.  If the EOS is `"gsw"` then Conservative
-#' Temperature and Absolute Salinity will be shown; otherwise in-situ
-#' temperature and practical salinity will be shown.
+#' Temperature and Absolute Salinity are shown; otherwise in-situ temperature
+#' and practical salinity are shown.
 #'
 #' * `which=2` or `which="density+N2"` gives a combined profile of density
 #' anomaly, computed with [swSigma0()], along with the square of the buoyancy
-#' frequency, computed with [swN2()].  Note that the `eos` parameter is passed
-#' to each of these functions.
[…]
