+++ R/argo.R
@@ -1802,226 +1802,226 @@ as.argo <- function(time, longitude, latitude,
 #' @family functions that plot oce data
 #' @aliases plot.argo
 setMethod(f="plot",
-          signature=signature("argo"),
-          definition=function (x, which=1, level,
-                               coastline=c("best", "coastlineWorld", "coastlineWorldMedium",
-                                           "coastlineWorldFine", "none"),
-                               cex=1, pch=1, type='p', col, fill=FALSE,
-                               projection=NULL,
-                               mgp=getOption("oceMgp"), mar=c(mgp[1]+1.5, mgp[1]+1.5, 1.5, 1.5),
-                               tformat,
-                               debug=getOption("oceDebug"),
-                               ...)
-          {
-              if (!inherits(x, "argo"))
-                  stop("method is only for objects of class '", "argo", "'")
-              oceDebug(debug, "plot.argo(x, which=c(", paste(which, collapse=","), "),",
-                      argShow(mgp),
-                      argShow(mar),
-                      argShow(cex),
-                      " ...) {\n", sep="", unindent=1, style="bold")
-              coastline <- match.arg(coastline)
-              nw  <- length(which)
-              if (nw > 1)
-                  par(mfcol=c(1, nw))
-              par(mgp=mgp, mar=mar)
-              if (missing(level) || level == "all")
-                  level <- seq(1L, dim(x@data$temperature)[1])
-              longitude <- x[["longitude"]]
-              latitude <- x[["latitude"]]
-              dim <- dim(x@data$salinity)
-              if (length(longitude) < prod(dim)) {
-                  ## Copy across depths. This is inside a conditional because
-                  ## possibly argo[["longitude"]] should mimic section[["longitude"]],
-                  ## in doing the lengthing by itself unless the second argument is
-                  ## "byStation" (issue 1273 ... under consideration 2017jul12)
-                  longitude <- rep(x[["longitude"]], each=dim[1])
-                  latitude <- rep(x[["latitude"]], each=dim[1])
-              }
-              ctd <- as.ctd(x@data$salinity, x@data$temperature, x@data$pressure,
-                            longitude=longitude, latitude=latitude,
-                            units=list(temperature=list(unit=expression(degree*C), scale="ITS-90"),
-                                       conductivity=list(list=expression(), scale=""))) # guess on units
-              whichOrig <- which
-              which <- oce.pmatch(which,
-                                  list("trajectory"=1,
-                                       "map"=1,
-                                       "salinity ts"=2,
-                                       "temperature ts"=3,
-                                       "TS"=4,
-                                       "salinity profile"=5,
-                                       "temperature profile"=6))
-              if (any(is.na(which)))
-                  stop("In plot,argo-method() :\n  unrecognized value(s) of which: ", paste(whichOrig[is.na(which)], collapse=", "), call.=FALSE)
-              lw <- length(which)
-              par(mgp=mgp)
-              if (lw == 2) {
-                  par(mfcol=c(2, 1))
-              } else if (lw == 3) {
-                  par(mfcol=c(3, 1))
-              } else if (lw == 4) {
-                  par(mfrow=c(2, 2))
-              } else if (lw != 1) {
-                  nnn <- floor(sqrt(lw))
-                  par(mfcol=c(nnn, ceiling(lw/nnn)))
-                  rm(nnn)
-              }
-              for (w in 1:nw) {
-                  if (which[w] == 1) {
-                      oceDebug(debug, "which[", w, "] == 1, so plotting a map\n")
-                      oceDebug(debug, "note: par(\"mfrow\") = ", paste(par("mfrow"), collapse=","), "\n")
-                      ## map
[…]

[…]
