### 2022-01-05 b60d775
all T variants obey Tlim
