+++ R/ctd.R
@@ -3312,7 +3312,7 @@ setMethod(f="plot",
             oceDebug(debug, "which[", w, "]\"=", which[w], "\"\n", sep="")
             # Case 1: non-single-profile plots.
             if (which[w] == "salinity+temperature") {
-                oceDebug(debug, "handling which=\"", whichOrig[w], "\" as a special case\n", sep="")
+                oceDebug(debug, "handling which[", w, "]=\"", whichOrig[w], "\" as a special case\n", sep="")
                 plotProfile(x, xtype="salinity+temperature",
                     plim=plim, Slim=Slim, Tlim=Tlim,
                     eos=eos,
@@ -3324,7 +3324,7 @@ setMethod(f="plot",
                     debug=debug-1,
                     ...)
             } else if (which[w] == "density+N2") {
-                oceDebug(debug, "handling which=\"", whichOrig[w], "\" as a special case\n", sep="")
+                oceDebug(debug, "handling which[", w, "]=\"", whichOrig[w], "\" as a special case\n", sep="")
                 plotProfile(x, xtype="density+N2",
                     plim=plim, N2lim=N2lim, densitylim=densitylim,
                     eos=eos,
@@ -3337,7 +3337,7 @@ setMethod(f="plot",
                     debug=debug-1,
                     ...)
             } else if (which[w] == "density+dpdt") {
-                oceDebug(debug, "handling which=\"", whichOrig[w], "\" as a special case\n", sep="")
+                oceDebug(debug, "handling which[", w, "]=\"", whichOrig[w], "\" as a special case\n", sep="")
                 plotProfile(x, xtype="density+dpdt",
                     plim=plim, densitylim=densitylim, dpdtlim=dpdtlim,
                     col=col,
@@ -3350,7 +3350,7 @@ setMethod(f="plot",
                     debug=debug-1,
                     ...)
             } else if (which[w] == "density+time") {
-                oceDebug(debug, "handling which=\"", whichOrig[w], "\" as a special case\n", sep="")
+                oceDebug(debug, "handling which[", w, "]=\"", whichOrig[w], "\" as a special case\n", sep="")
                 plotProfile(x, xtype="density+time",
                     plim=plim, densitylim=densitylim, timelim=timelim,
                     useSmoothScatter=useSmoothScatter,
@@ -3360,25 +3360,14 @@ setMethod(f="plot",
                     keepNA=keepNA, inset=inset, add=add,
                     debug=debug-1,
                     ...)
-            #> } else if (which[w] == "salinity") { # a special case because can use Slim and plim
-            #>     plotProfile(x, xtype="salinity",
-            #>         plim=plim,
-            #>         Slim=Slim,
-            #>         col=col,
-            #>         eos=eos,
-            #>         useSmoothScatter=useSmoothScatter,
-            #>         grid=grid, col.grid="lightgray", lty.grid="dotted",
-            #>         cex=cex[w], pch=pch[w],
-            #>         type=if (!missing(type)) type[w],
-            #>         keepNA=keepNA, inset=inset, add=add,
-            #>         debug=debug-1,
-            #>         ...)
-            } else if (which[w] %in% c("salinity", "SP",
-                    "SA", paste("absolute", "salinity"), paste("Absolute", "Salinity"),
-                    "Sstar")) { # a special case because can use Slim and plim
-                oceDebug(debug, "handling whichOrig=\"", whichOrig[w], "\" (which=\"",which[w],"\") as a special case\n", sep="")
-                plotProfile(x, xtype=x[[which[w]]], xlab=resizableLabel(which, debug=debug-1),
-                    xlim=Slim, ylim=plim, eos="gsw",
+            } else if (which[w] %in% c("salinity",
[…]
