### 2022-01-07 2de6db3
[[ handles SR and Splus
including new handling of argo (and similar) as a special case.
