### 2022-01-07 e0da5d3
fix two problematic URLs
