### 2022-01-07 525bf02
fix a broken quote mark (found by spell check)
This is curious, since I was pretty sure I fixed that in the "develop"
branch before merging that into "master".  I'll keep an eye out for
recurrance after I merge "master" back into "develop" but I will be
submitting to CRAN from "master" and so this won't be for days.
