+++ man-roxygen/colourBlindnessTemplate.R
@@ -14,12 +14,12 @@
 ## \samp{https://www.ngdc.noaa.gov/mgg/bathymetry/arctic/ibcao_gebco_comp.html}.
 #'
 #' Stephenson, David B. "Comment on 'Color Schemes for Improved Data
-#' Graphics,' by A Light and P.J. Bartlein." *Eos, Transactions American
+#' Graphics', by A Light and P.J. Bartlein." *Eos, Transactions American
 #' Geophysical Union* 86, no.  20 (2005): 196.
 #' \url{https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2005EO200005}.
 #'
-#' Light, Adam, and Patrick J. Bartlein. "Reply \[to 'Comment on Color
-#' Schemes for Improved Data Graphics,’ by A. Light and P.J.  Bartlein'\]."
+#' Light, Adam, and Patrick J. Bartlein. "Reply to 'Comment on Color
+#' Schemes for Improved Data Graphics,' by A. Light and P.J.  Bartlein'."
 #' *Eos, Transactions American Geophysical Union* 86, no. 20 (2005): 196–196.
 #' \url{https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2005EO200006}
 
