+++ man/oceColorsCDOM.Rd
@@ -71,12 +71,12 @@ Union} 85, no. 40 (2004): 385.
 \url{https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2004EO400002}
 
 Stephenson, David B. "Comment on 'Color Schemes for Improved Data
-Graphics,' by A Light and P.J. Bartlein." \emph{Eos, Transactions American
+Graphics', by A Light and P.J. Bartlein." \emph{Eos, Transactions American
 Geophysical Union} 86, no.  20 (2005): 196.
 \url{https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2005EO200005}.
 
-Light, Adam, and Patrick J. Bartlein. "Reply [to 'Comment on Color
-Schemes for Improved Data Graphics,’ by A. Light and P.J.  Bartlein']."
+Light, Adam, and Patrick J. Bartlein. "Reply to 'Comment on Color
+Schemes for Improved Data Graphics,' by A. Light and P.J.  Bartlein'."
 \emph{Eos, Transactions American Geophysical Union} 86, no. 20 (2005): 196–196.
 \url{https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2005EO200006}
 }
