+++ .github/workflows/pkgdown.yaml
@@ -1,7 +1,7 @@
 on:
   push:
     branches:
-      - master
+      - main
 
 name: pkgdown
 
