### 2022-01-10 d7c964f
rename vignettes to get proper in order on CRAN
