+++ vignettes/06_subclassing.Rmd
@@ -7,8 +7,8 @@ output:
     toc: true
     number_sections: true
     fig_caption: yes
-    fig_width: 3.5
-    fig_height: 3.5
+    fig_width: 4
+    fig_height: 4
     dpi: 72
     dev.args: list(pointsize=11)
 vignette: >
@@ -32,7 +32,7 @@ objects, in terms of operators such as `[[` and `[[<-`, functions such as
 `subset()`  and `summary()`, schemes for handling units, etc.
 
 
-## Initializing oce objects
+# Initializing oce objects
 
 Oce handles a wide variety of data types, creating sub-classes the inherit from
 the base class that is named `oce`. However, even if a dataset does not fit
@@ -46,14 +46,18 @@ first create some fake data:
 ```{r}
 lon <- -60
 lat <- -30
-t <- seq(as.POSIXct("2017-08-30"), as.POSIXct("2017-09-01"), by="15 min")
+t <- seq(as.POSIXct("2022-01-09"), as.POSIXct("2022-01-10"), by="15 min")
 tsec <- as.numeric(t)
-u <- sin(tsec * 2 * pi / (3600 * 12.4206))
-v <- 0.5 * cos(tsec * 2 * pi / (3600 * 12.4206))
+u <- 0.5 * sin(tsec * 2 * pi / (3600 * 12.4206))
+v <- 1.0 * cos(tsec * 2 * pi / (3600 * 12.4206) - 3600)
 ```
-```{r echo=FALSE, eval=FALSE}
-plot(t, u, type="l")
-lines(t, v, col=2)
+
+Before proceeding to the creation of an `oce` object, it makes sense to get a
+quick view of the variation.
+
+```{r fig.width=4, fig.height=4, dpi=72, fig.cap="**Figure 1.** Variation of $u=u(t)$ and $v=v(t)$."}
+matplot(t, cbind(u,v), ylab="u (black) and v (red)",
+    type="l", col=1:2, lty=1)
 ```
 
 These may be inserted into an `oce` object as follows. First, create the object.
@@ -70,14 +74,13 @@ As noted previously, it's possible to insert `t`, `u`, and `v` directly into
 the `data` slot, but it's better to use `oceSetData` to do that, because it
 can set up units, and it leaves a trail in the processing log.
 ```{R}
-o <- oceSetData(o, "t", t, list(unit=expression(s), scale=""))
+o <- oceSetData(o, "time", t, list(unit=expression(s), scale=""))
 o <- oceSetData(o, "u", u, list(unit=expression(m/s), scale=""))
 o <- oceSetData(o, "v", v, list(unit=expression(m/s), scale=""))
 ```
 
-Astute readers will realize that these data are akin to tidal velocities, and
-so might agree with the author that `lon` and `lat` belong in the `metadata`
-slot, not the `data` slot.
+These data are akin to tidal velocities, and so might make sense to store the
+sampling location in the `metadata` slot, as opposed to the `data` slot, using
 ```{R}
 o <- oceSetMetadata(o, "longitude", lon)
 o <- oceSetMetadata(o, "latitude", lat)
@@ -90,13 +93,13 @@ summary(o)
 and individual data components may be extracted with the `[[` notation
 ```{r}
 o[["latitude"]]                        # from metadata
-head(o[["t"]])                         # from data
+head(o[["time"]])                      # from data
 head(o[["u"]])
 head(o[["v"]])
 ```
 
 A crude plot can be done with
-```{r fig.cap="**Figure 11.** Crude plot of constructed `oce` object.", fig.width=4, fig.height=4, dpi=72}
+```{r fig.cap="**Figure 2.** Crude plot of constructed `oce` object.", fig.width=4, fig.height=4, dpi=72}
 plot(o)
 ```
[…]
