### 2022-01-10 2d3afc0
update sub-classing vignette & reorder vignettes
