+++ vignettes/F_subclassing.Rmd
@@ -1,0 +1,282 @@
+f--
+title: "6. Subclassing oce objects"
+author: "Dan Kelley (https://orcid.org/0000-0001-7808-5911)"
+date: "`r Sys.Date()`"
+output:
+  rmarkdown::html_vignette:
+    toc: true
+    number_sections: true
+    fig_caption: yes
+    fig_width: 4
+    fig_height: 4
+    dpi: 72
+    dev.args: list(pointsize=9)
+vignette: >
+  %\VignetteIndexEntry{6. Subclassing oce objects}
+  %\VignetteEngine{knitr::rmarkdown}
+  \usepackage[utf8]{inputenc}
+---
+
+<!-- HOW TO BUILD THE VIGNETTE. -->
+<!-- Look in oce.Rmd for instructions. -->
+
+
+```{r, echo=FALSE}
+knitr::opts_chunk$set(collapse=TRUE, comment="#>")
+```
+
+**Abstract.** This vignette explains how new classes of objects can be created,
+using oce objects as a base class. The benefit of doing this is that the
+newly-formed objects will automatically have important properties of oce
+objects, in terms of operators such as `[[` and `[[<-`, functions such as
+`subset()`  and `summary()`, schemes for handling units, etc. The procedures
+for working with sub-classes, mainly through the use of the `setMethod()`
+function, are outlined here in a practical way, using data from a drifting buoy
+as an example.
+
+
+# A drifting buoy
+
+Data from buoy number 4201703 were downloaded from
+http://www.coriolis.eu.org/Data-Products/Data-Delivery/Data-selection on
+2022-01-09, by deselecting all data types except for "Drifting buoy", selecting
+a data window south of Nova Scotia and of area similar to that province,
+expanding the tarball created by the website and then focussing on the file
+named "`drifting-buoys-4201703.csv"`.
+
+There are many data in the source data file, but only a subset is used here:
+starting with 300 samples of just five of the data fields: time, longitude,
+latitude, temperature and atmospheric pressure. This was further trimmed by
+removing data ensembles in which temperature was not reported.  The result was
+146 observations of 5 variables.  Note that they are *not* spaced uniformly in
+time, a feature that is of importance only at the end of this document.
+
+The data were combined into a data frame and packaged in an rda file, as
+illustrated with
+```{r}
+library(oce)
+load(system.file("extdata", "drifter.rda", package="oce"))
+str(db)
+```
+
+
+This dataset shows some intriguing patterns, so it seems a suitable choice for
+an illustration of sub-classing. But first, it is worth showing how to work
+with the base class.
+
+# Using base oce classes
+
+The base class is named `"oce"`, and so a new object of that type can be created with
+```{r}
+o <- new("oce")
+```
+At this point, the `summary()` method does not reveal much
+```{r}
+summary(o)
+```
+
+The next step is to populate the object with some data and metadata.
+```{r}
+o <- oceSetData(o, "time", db$t)
+o <- oceSetData(o, "longitude", db$lon)
+o <- oceSetData(o, "latitude", db$lat)
+o <- oceSetMetadata(o, "ID", 4201703)
+```
+
+Now the summary is more informative
+```{r}
+summary(o)
+```
+However, note that the `ID` we added is not listed, because metadata tend to be
+so varied and expansive that `summary,oce-method` does not bother our entry.
+Getting this listed will require specializing constructing a new object, and
+telling that object's `summary()` method that `"ID"` is something we want to
+know.
+
+Objects of `oce-class` automatically have other methods, besides
[…]
