+++ NEWS.md
@@ -1,5 +1,9 @@
 # oce
 
+## 1.6.0
+
+* TBA
+
 ## 1.5.0
 
 * Add `[["?"]]` facility to all objects.
