+++ vignettes/06_subclassing.Rmd
@@ -1,456 +1,0 @@
----
-title: "6. Subclassing oce objects"
-author: "Dan Kelley (https://orcid.org/0000-0001-7808-5911)"
-date: "`r Sys.Date()`"
-output:
-  rmarkdown::html_vignette:
-    toc: true
-    number_sections: true
-    fig_caption: yes
-    fig_width: 4
-    fig_height: 4
-    dpi: 72
-    dev.args: list(pointsize=11)
-vignette: >
-  %\VignetteIndexEntry{6. Subclassing oce objects}
-  %\VignetteEngine{knitr::rmarkdown}
-  \usepackage[utf8]{inputenc}
----
-
-<!-- HOW TO BUILD THE VIGNETTE. -->
-<!-- Look in oce.Rmd for instructions. -->
-
-
-```{r, echo = FALSE}
-knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
-```
-
-**Abstract.** This vignette explains how new classes of objects can be created,
-using oce objects as a base class. The advantage of this is that the
-newly-formed objects will automatically have important properties of oce
-objects, in terms of operators such as `[[` and `[[<-`, functions such as
-`subset()`  and `summary()`, schemes for handling units, etc.
-
-
-# Initializing oce objects
-
-Oce handles a wide variety of data types, creating sub-classes the inherit from
-the base class that is named `oce`. However, even if a dataset does not fit
-within these sub-classes, it is still possible to use some of the features of
-`oce`, such as accessors, plotting, and summaries.
-
-For example, suppose that there is a dataset that contains two variables, `mm`
-and `ss`, that have been measured at times `t`. Suppose that the data are
-recorded at a fixed location. To illustrate how to deal with such data, we
-first create some fake data:
-```{r}
-lon <- -60
-lat <- -30
-t <- seq(as.POSIXct("2022-01-09"), as.POSIXct("2022-01-10"), by="15 min")
-tsec <- as.numeric(t)
-u <- 0.5 * sin(tsec * 2 * pi / (3600 * 12.4206))
-v <- 1.0 * cos(tsec * 2 * pi / (3600 * 12.4206) - 3600)
-```
-
-Before proceeding to the creation of an `oce` object, it makes sense to get a
-quick view of the variation.
-
-```{r fig.width=4, fig.height=4, dpi=72, fig.cap="**Figure 1.** Variation of $u=u(t)$ and $v=v(t)$."}
-matplot(t, cbind(u,v), ylab="u (black) and v (red)",
-    type="l", col=1:2, lty=1)
-```
-
-These may be inserted into an `oce` object as follows. First, create the object.
-```{R}
-library(oce)
-o <- new("oce")
-```
-This contains slots without data:
-```{R}
-str(o)
-```
-
-As noted previously, it's possible to insert `t`, `u`, and `v` directly into
-the `data` slot, but it's better to use `oceSetData` to do that, because it
-can set up units, and it leaves a trail in the processing log.
-```{R}
-o <- oceSetData(o, "time", t, list(unit=expression(s), scale=""))
-o <- oceSetData(o, "u", u, list(unit=expression(m/s), scale=""))
-o <- oceSetData(o, "v", v, list(unit=expression(m/s), scale=""))
-```
-
-These data are akin to tidal velocities, and so might make sense to store the
-sampling location in the `metadata` slot, as opposed to the `data` slot, using
-```{R}
-o <- oceSetMetadata(o, "longitude", lon)
-o <- oceSetMetadata(o, "latitude", lat)
-```
-
-At this stage, `o` is a full-fledged `oce` object. It may be summarized with
-```{r}
-summary(o)
-```
[…]
