+++ vignettes/F_subclassing.Rmd
@@ -1,4 +1,4 @@
-f--
+---
 title: "6. Subclassing oce objects"
 author: "Dan Kelley (https://orcid.org/0000-0001-7808-5911)"
 date: "`r Sys.Date()`"
@@ -10,9 +10,9 @@ output:
     fig_width: 4
     fig_height: 4
     dpi: 72
-    dev.args: list(pointsize=9)
+    dev.args: list(pointsize=11)
 vignette: >
-  %\VignetteIndexEntry{6. Subclassing oce objects}
+  %\VignetteIndexEntry{6. Creating new classes in the oce system}
   %\VignetteEngine{knitr::rmarkdown}
   \usepackage[utf8]{inputenc}
 ---
@@ -25,14 +25,13 @@ vignette: >
 knitr::opts_chunk$set(collapse=TRUE, comment="#>")
 ```
 
-**Abstract.** This vignette explains how new classes of objects can be created,
-using oce objects as a base class. The benefit of doing this is that the
-newly-formed objects will automatically have important properties of oce
-objects, in terms of operators such as `[[` and `[[<-`, functions such as
-`subset()`  and `summary()`, schemes for handling units, etc. The procedures
-for working with sub-classes, mainly through the use of the `setMethod()`
-function, are outlined here in a practical way, using data from a drifting buoy
-as an example.
+**Abstract.** This vignette explains how to create new object classes, using
+the low-level `oce` class as a foundation. This scheme is beneficial, because
+the newly-formed objects automatically gain the key properties of oce objects,
+in terms of operators such as `[[` and `[[<-`, functions such as `subset()` and
+`summary()`, schemes for handling units, etc. The `setMethod()` function is
+central to the process of creating new classes, as will be illustrated here in
+a practical way, using drifting-buoy data for example.
 
 
 # A drifting buoy
@@ -41,89 +40,92 @@ Data from buoy number 4201703 were downloaded from
 http://www.coriolis.eu.org/Data-Products/Data-Delivery/Data-selection on
 2022-01-09, by deselecting all data types except for "Drifting buoy", selecting
 a data window south of Nova Scotia and of area similar to that province,
-expanding the tarball created by the website and then focussing on the file
-named "`drifting-buoys-4201703.csv"`.
+expanding the tarball created by the website and then focussing on the
+downloaded file named `"drifting-buoys-4201703.csv`".
 
 There are many data in the source data file, but only a subset is used here:
-starting with 300 samples of just five of the data fields: time, longitude,
-latitude, temperature and atmospheric pressure. This was further trimmed by
-removing data ensembles in which temperature was not reported.  The result was
-146 observations of 5 variables.  Note that they are *not* spaced uniformly in
-time, a feature that is of importance only at the end of this document.
+starting with 300 samples of just five of the data fields: time (called `t` for
+brevity), longitude (`lon`), latitude (`lat`), temperature (`Tatm`) and
+atmospheric pressure (`Patm`). This set was further trimmed by removing data
+ensembles in which temperature was not reported.  The result was 146
+observations of 5 variables. (Note that they are *not* spaced uniformly in
+time, a feature that is of importance only at the end of this document.)
 
-The data were combined into a data frame and packaged in an rda file, as
-illustrated with
+The data were combined into a data frame named `db` and packaged in an RDA
+file, an overview of which is provided as follows.
 ```{r}
 library(oce)
 load(system.file("extdata", "drifter.rda", package="oce"))
 str(db)
 ```
 
+Before showing how to create a new sub-class, it will first be shown that such
+data could also be stored in the foundational class of the package.
 
-This dataset shows some intriguing patterns, so it seems a suitable choice for
-an illustration of sub-classing. But first, it is worth showing how to work
-with the base class.
-
[…]
