### 2022-01-11 2fb481a
See https://github.com/dankelley/oce/issues/1914
1. The is.raw() check is correct now, so issue 1914 has been addressed.
2. I made the debugging output be more informative, yet less verbose.
3. The diagram made by example('handleFlags,adp-method') now has an arrow
   pointing out a spot that changes (subtly) when flags are applied.
4. There is now a test-suite component on this, as below (from
   tests/testthat/test_flags.R)

```R
test_that("adp handleFlag gives error for raw data (issue 1914)", {
    data(adp)
    v <- adp[["v"]]
    i2 <- array(FALSE, dim=dim(v))
    g <- adp[["g", "numeric"
[…]
