+++ R/adp.R
@@ -310,7 +310,7 @@ setMethod(f="initialize",
 #' v <- adp[["v"]]
 #' i2 <- array(FALSE, dim=dim(v))
 #' g <- adp[["g", "numeric"]]
-#' # Thresholds on percent "goodness" and error "velocity"
+#' # Set thresholds on percent "goodness" and error "velocity".
 #' G <- 25
 #' V4 <- 0.45
 #' for (k in 1:3)
@@ -320,34 +320,39 @@ setMethod(f="initialize",
 #' adpClean <- handleFlags(adpQC, flags=list(3), actions=list("NA"))
 #' # Demonstrate (subtle) change graphically.
 #' par(mfcol=c(2, 1))
-#' plot(adp, which="u1")
-#' plot(adpClean, which="u1")
+#' plot(adp, which="u1", drawTimeRange=FALSE)
+#' plot(adpClean, which="u1", drawTimeRange=FALSE)
+#' t0 <- 1214510000 # from locator()
+#' arrows(t0, 20, t0, 35, length=0.1, lwd=3, col="magenta")
+#' mtext("Slight change above arrow", col="magenta", font=2)
 #'
 #' @family things related to adp data
 setMethod("handleFlags", signature=c(object="adp", flags="ANY", actions="ANY", where="ANY", debug="ANY"),
-          definition=function(object, flags=NULL, actions=NULL, where=NULL, debug=getOption("oceDebug")) {
-              ## DEVELOPER 1: alter the next comment to explain your setup
-              ## Flag=1 means bad velocity; 0 means good
-              names <- names(object[["flags"]])
-              for (name in names) {
-                  for (j in 1:length(object[[name]])) {
-                      if (any(class(object[[name]][j]) == "raw"))
-                          stop("use adpConvertRawToNumeric() first to convert raw values to numeric")
-                  }
-              }
-              if (is.null(flags)) {
-                  flags <- defaultFlags(object)
-                  if (is.null(flags))
-                      stop("must supply 'flags', or use initializeFlagScheme() on the adp object first")
-              }
-              if (is.null(actions)) {
-                  actions <- list("NA") # DEVELOPER 3: alter this line to suit a new data class
-                  names(actions) <- names(flags)
-              }
-              if (any(names(actions)!=names(flags)))
-                  stop("names of flags and actions must match")
-              handleFlagsInternal(object=object, flags=flags, actions=actions, where=where, debug=debug)
-          })
+    definition=function(object, flags=NULL, actions=NULL, where=NULL, debug=getOption("oceDebug")) {
+        oceDebug(debug, "handleFlags,adp-method {\n", sep="", style="bold", unindent=1)
+        ## DEVELOPER 1: alter the next comment to explain your setup
+        ## Flag=1 means bad velocity; 0 means good
+        names <- names(object[["flags"]])
+        for (name in names) {
+            if (is.raw(object@data[[name]])) {
+                stop("use adpConvertRawToNumeric() first to convert raw values to numeric")
+            }
+        }
+        if (is.null(flags)) {
+            flags <- defaultFlags(object)
+            if (is.null(flags))
+                stop("must supply 'flags', or use initializeFlagScheme() on the adp object first")
+        }
+        if (is.null(actions)) {
+            actions <- list("NA") # DEVELOPER 3: alter this line to suit a new data class
+            names(actions) <- names(flags)
+        }
+        if (any(names(actions)!=names(flags)))
+            stop("names of flags and actions must match")
+        res <- handleFlagsInternal(object=object, flags=flags, actions=actions, where=where, debug=debug-1L)
+        oceDebug(debug, "} # handleFlags,adp-method()\n", sep="", style="bold", unindent=1)
[…]
