+++ man/handleFlags-adp-method.Rd
@@ -97,7 +97,7 @@ data(adp)
 v <- adp[["v"]]
 i2 <- array(FALSE, dim=dim(v))
 g <- adp[["g", "numeric"]]
-# Thresholds on percent "goodness" and error "velocity"
+# Set thresholds on percent "goodness" and error "velocity".
 G <- 25
 V4 <- 0.45
 for (k in 1:3)
@@ -107,8 +107,11 @@ adpQC <- setFlags(adpQC, "v", i2, 3)
 adpClean <- handleFlags(adpQC, flags=list(3), actions=list("NA"))
 # Demonstrate (subtle) change graphically.
 par(mfcol=c(2, 1))
-plot(adp, which="u1")
-plot(adpClean, which="u1")
+plot(adp, which="u1", drawTimeRange=FALSE)
+plot(adpClean, which="u1", drawTimeRange=FALSE)
+t0 <- 1214510000 # from locator()
+arrows(t0, 20, t0, 35, length=0.1, lwd=3, col="magenta")
+mtext("Slight change above arrow", col="magenta", font=2)
 
 }
 \seealso{
