### 2022-01-15 86fcc9a
vectorShow([empty]) ends with single newline
It was giving two newlines.
