### 2022-01-29 da0b046
vectorShow() can handle lists
