### 2022-01-29 9c9d901
vectorShow() improved for lists and named vectors
Now, it provides names for the items within lists and named-item
vectors.  See https://github.com/dankelley/oce/issues/1915 for details.
