### 2022-01-30 ec25b81
read.met() handles new EnvCanada format
See https://github.com/dankelley/oce/issues/1916
