+++ R/met.R
@@ -550,10 +550,8 @@ metNames2oceNames <- function(names, scheme)
 #'
 #' Reads some meteorological file formats used by the Environment
 #' Canada (see reference 1).  Since the agency does not publish the
-#' data formats, this function had to be based on some sample files, and it
-#' is likely to fail if Environment Canada changes their file format. For
-#' example, a change was required in October 2019, to accommodate changes
-#' to the file format noticed at that time (see the notes on the `type` argument).
+#' data formats, this function has to be adjusted every few years,
+#' when a user finds that the format has changed.
 #'
 #' @param file a character string naming a file that holds met data.
 #'
@@ -564,16 +562,27 @@ metNames2oceNames <- function(names, scheme)
 #' old CSV format no longer provided as of October 2019,
 #' (b) `"csv2"` for a CSV format noticed on the Environment Canada
 #' website in October 2019 (but note that the paired metadata file
-#' is ignored), and (c) `"xml2"` for an XML format that was noticed
+#' is ignored), (c) `"csv3"` for a CSV format noticed on the
+#' Environment Canada website in January 2020,
+#' and (d) `"xml2"` for an XML format that was noticed
 #' on the Environment Canada website in October 2019.
 #'
 #' @param skip number of lines of header that occur before the actual
 #' data.  This is ignored unless `type` is `"csv"` or `"csv1"`, in which case
 #' a non-`NULL` value of `skip` is taken as the number of lines preceding
 #' the columnar data ... and this is only needed if [read.met()] cannot
-#' find a line starting with `"Date/Time"`.
+#' find a line starting with `"Date/Time"` (for csv2 format)
+#' or `"Date/Time (LTC)"` (for csv3 format).
 #'
-#' @param tz timezone assumed for time data
+#' @param tz timezone assumed for time data.  This defaults to
+#' `getOption("oceTz")`, which is very likely to be wrong.  In
+#' a scientific context, where UTC is typically used for oceanographic
+#' measurement, it makes sense to set `tz="UTC"`.  Note that these
+#' data files do not contain timezone information, instead giving
+#' data in Local Standard Time (LST).  Since LST differs from city
+#' to city, users must make corrections to the time, as
+#' shown in the \dQuote{Examples}, which use data for
+#' Halifax Nova Scotia, where LST is UTC-4.
 #'
 #' @param debug a flag that turns on debugging.  Set to 1 to get a moderate
 #' amount of debugging information, or to 2 to get more.
@@ -583,23 +592,33 @@ metNames2oceNames <- function(names, scheme)
 #' @author Dan Kelley
 #'
 #' @examples
-#' # The old csv format (no longer supplied by Environment Canada as of Oct 2019)
+#' # Example 1: "csv1" Environment Canada format (found to be obsolete as of Oct 2019)
 #' csv1 <- read.met(system.file("extdata", "test_met_vsn1.csv", package="oce"))
+#' csv1 <- oceSetData(csv1, "time", csv1[["time"]]+4*3600,
+#'     note="add 4h to local time to get UTC time")
 #'
-#' # The new "csv2" format (provided by Environment Canada as of Oct 2019)
+#' # Example 2: "csv2" Environment Canada format (found to be obsolete as of Jan 2022)
 #' csv2 <- read.met(system.file("extdata", "test_met_vsn2.csv", package="oce"))
[…]
