### 2022-01-30 1cdb843
improve read.met(); fix broken links in vignettes
