### 2022-01-30 80800e4
drop the met test on all platforms
