### 2022-01-30 c2906d3
met test again on non-windows machines
