+++ tests/testthat/test_met.R
@@ -69,17 +69,18 @@ test_that("read.met() handles type=\"csv1\" files", {
 # users to tell me if they see a problem ... that is, users who will
 # be willing to do some tests that will be faster than my waiting 30min
 # for each win_builder test.
-#>if (.Platform$OS.type != "windows") {
-if (FALSE) {
+if (.Platform$OS.type != "windows") {
+#>if (FALSE) {
     test_that("read.met() handles type=\"csv3\" files", {
-        d <- read.met(csv3)
+        expect_silent(d <- read.met(csv3))
         # Sort both because the ordering is different when done interactively
         # and in the test (for reasons I don't understand).
-        expect_equal(sort(names(d@data)), sort(c(paste("Date/Time", "(LST)"), "dewPoint",
-                    "direction", "humidex", "humidity", paste("Precip.", "Amount",
-                        "(mm)"), paste("Precip.", "Amount", "Flag"), "pressure",
-                    "speed", "temperature", "time", paste("Time", "(LST)"), "u", "v",
-                    "visibility", "weather", "windChill")))
+        expect_equal(sort(names(d@data)), sort(c(paste("Date/Time", "(LST)"),
+                    "dewPoint", "direction", "humidex", "humidity",
+                    "precipitation",  "pressure", "speed", "temperature",
+                    "time", paste("Time", "(LST)"), "u", "v", "visibility",
+                    "weather", "windChill")))
+
         expect_equal(d[["latitude"]], 44.88)
         expect_equal(d[["longitude"]], -63.51)
         expect_equal(d[["station"]], "HALIFAX STANFIELD INT'L A")
