### 2022-02-02 6ecc229
plot,ctd-method now obeys col (again)
See https://github.com/dankelley/oce/issues/1917 for details.  This
error crept in when I revamped plotting a few weeks ago to permit
plotting pretty much *anything* in the CTD object or computable from the
contents. I've checked this for which="CT" but not too many other
things, but the diff should reveal that it ought to work in quite a lot
of cases.
