+++ R/ctd.R
@@ -3238,6 +3238,8 @@ setMethod(f="plot",
         if (lw > 1) on.exit(par(opar))
         if (missing(type))
             type <- rep("p", lw)
+        if (length(col) < lw)
+            col <- rep(col, lw)  # FIXME: recycle more sensibly
         if (length(type) < lw)
             type <- rep(type, lw)
         if (length(pch) < lw)
@@ -3374,7 +3376,7 @@ setMethod(f="plot",
                     xlim=Slim, ylim=plim,
                     useSmoothScatter=useSmoothScatter,
                     grid=grid, col.grid="lightgray", lty.grid="dotted",
-                    cex=cex[w], pch=pch[w],
+                    cex=cex[w], pch=pch[w], col=col[w],
                     type=if (!missing(type)) type[w],
                     keepNA=keepNA, inset=inset, add=add,
                     debug=debug-1,
@@ -3390,7 +3392,7 @@ setMethod(f="plot",
                     xlim=Tlim, ylim=plim,
                     useSmoothScatter=useSmoothScatter,
                     grid=grid, col.grid="lightgray", lty.grid="dotted",
-                    cex=cex[w], pch=pch[w],
+                    cex=cex[w], pch=pch[w], col=col[w],
                     type=if (!missing(type)) type[w],
                     keepNA=keepNA, inset=inset, add=add,
                     debug=debug-1,
@@ -3405,7 +3407,7 @@ setMethod(f="plot",
                     eos=eos,
                     useSmoothScatter=useSmoothScatter,
                     col.grid="lightgray", lty.grid="dotted",
-                    cex=cex[w], pch=pch[w],
+                    cex=cex[w], pch=pch[w], col=col[w],
                     type=if (!missing(type)) type[w],
                     keepNA=keepNA, inset=inset, add=add,
                              debug=debug-1,
@@ -3431,12 +3433,11 @@ setMethod(f="plot",
                     plim=plim,
                     N2lim=N2lim,
                     grid=grid,
-                    col=col,
                     eos=eos,
                     df=df,
                     useSmoothScatter=useSmoothScatter,
                     col.grid="lightgray", lty.grid="dotted",
-                    cex=cex[w], pch=pch[w],
+                    cex=cex[w], pch=pch[w], col=col[w],
                     type=if (!missing(type)) type[w],
                     keepNA=keepNA, inset=inset, add=add,
                     debug=debug-1,
@@ -3448,7 +3449,7 @@ setMethod(f="plot",
                     eos=eos,
                     useSmoothScatter=useSmoothScatter,
                     grid=grid, col.grid="lightgray", lty.grid="dotted",
-                    cex=cex[w], pch=pch[w],
+                    cex=cex[w], pch=pch[w], col=col[w],
                     type=if (!missing(type)) type[w],
                     keepNA=keepNA, inset=inset, add=add,
                     debug=debug-1,
@@ -4850,7 +4851,7 @@ plotProfile <- function(x,
     eos=getOption("oceEOS", default="gsw"),
     lty=1,
     xlab=NULL, ylab=NULL,
-    col='black',
+    col="black",
     col.salinity="darkgreen",
     col.temperature="red",
     col.rho="blue",
@@ -4886,6 +4887,7 @@ plotProfile <- function(x,
              ", plim=", if (missing(plim)) "MISSING" else paste("c(",paste(plim, collapse=","),")",sep=""),
              ", xlim=", if (missing(xlim)) "MISSING" else paste("c(",paste(xlim, collapse=","),")",sep=""),
              ", ylim=", if (missing(ylim)) "MISSING" else paste("c(",paste(ylim, collapse=","),")",sep=""),
[…]
