+++ R/ctd.R
@@ -3403,7 +3403,6 @@ setMethod(f="plot",
                     plim=plim,
                     densitylim=densitylim,
                     grid=grid,
-                    col=col,
                     eos=eos,
                     useSmoothScatter=useSmoothScatter,
                     col.grid="lightgray", lty.grid="dotted",
@@ -3445,7 +3444,6 @@ setMethod(f="plot",
             } else if (which[w] == "conductivity") { # a special case because can use Clim and plim
                 oceDebug(debug, "handling which[", w, "]=\"", whichOrig[w], "\" as a special case\n", sep="")
                 plotProfile(x, xtype="conductivity", Clim=Clim, plim=plim,
-                    col=col,
                     eos=eos,
                     useSmoothScatter=useSmoothScatter,
                     grid=grid, col.grid="lightgray", lty.grid="dotted",
