### 2022-02-17 e286818
fix doc error with image width (issue 1918)
As you can see at https://github.com/dankelley/oce/issues/1918, a new
CRAN release is required to address this.
