### 2022-02-17 3b79bb6
document the image-size declarations (issue 1918)
