### 2022-02-17 884998f
cran notes name the 4 test setups for R-CMD-check
