+++ R/landsat.R
@@ -1068,9 +1068,8 @@ read.landsatmeta <- function(file, debug=getOption("oceDebug"))
 #' @references
 #'
 #' 1. Konda, M. Imasato N., Nishi, K., and T. Toda, 1994.  Measurement of the Sea
-#' Surface Emissivity.  *Journal of Oceanography*, 50, 17:30.  Available at
-#' \url{http://www.terrapub.co.jp/journals/JO/pdf/5001/50010017.pdf} as of
-#' February 2015.
+#' Surface Emissivity.  *Journal of Oceanography*, 50, 17:30.
+#' \code{http://www.terrapub.co.jp/journals/JO/pdf/5001/50010017.pdf}
 #'
 #' @author Dan Kelley
 #'
