### 2022-02-17 3f9ec2b
bump version number (for CRAN submission)
