+++ man/read.landsat.Rd
@@ -99,7 +99,7 @@ early in an analysis.
 \enumerate{
 \item Konda, M. Imasato N., Nishi, K., and T. Toda, 1994.  Measurement of the Sea
 Surface Emissivity.  \emph{Journal of Oceanography}, 50, 17:30.
-\code{http://www.terrapub.co.jp/journals/JO/pdf/5001/50010017.pdf}
+\doi{10.1007/BF02233853}
 }
 }
 \seealso{
