### 2022-02-18 c7a219d
address "length_1 error"
I think I have there are 2 problems (see
https://www.stats.ox.ac.uk/pub/bdr/LENGTH1_self/oce.out) and search for
FAILURE_REPORT and "value of length:") and this commit addresses both.

I don't know how to turn on this flag in testing so I think I ought to
scour through that error file in more detail, to gain a bit more
confidence that I've fixed this.  In the meantime, this commit should be
small enough to be clear.
