### 2022-02-22 c5b9797
demo to 'test' drawPalette
It's not really a test, because it relies on a viewer's eyes to verify
that things are as stated.  (I don't know how to build formal tests
based on graphical output.  I suppose I could save to PNG and then load
that PNG and check pixels, but that's a lot of work and I think my time
is better spent otherwise.)
