### 2022-02-22 33f3398
add an 'at' and 'label' test to @dankelley drawPalette test code
