### 2022-02-23 4a88604
drawPalette() obeys 'at' and 'labels' (issue 1921)
See the test file in sandbox/dk/drawPalette
