+++ R/imagep.R
@@ -300,17 +300,17 @@ paletteCalculations <- function(separation=par('cin')[2]/2,
 #' contour(volcano, add=TRUE)
 #' drawPalette(c(90, 200), fullpage=TRUE, col=oce.colorsViridis)
 drawPalette <- function(zlim, zlab="",
-                        breaks, col, colormap,
-                        mai, cex=par("cex"),
-                        pos=4, las=0,
-                        labels=NULL, at=NULL,
-                        levels, drawContours=FALSE,
-                        plot=TRUE, fullpage=FALSE, drawTriangles=FALSE,
-                        axisPalette, tformat,
-                        debug=getOption("oceDebug"), ...)
+    breaks, col, colormap,
+    mai, cex=par("cex"),
+    pos=4, las=0,
+    labels=NULL, at=NULL,
+    levels, drawContours=FALSE,
+    plot=TRUE, fullpage=FALSE, drawTriangles=FALSE,
+    axisPalette, tformat,
+    debug=getOption("oceDebug"), ...)
 {
     debug <- max(0, min(debug, 2))
-    oceDebug(debug, gsub(" = [^,)]*", "", deparse(expr=match.call())), " {\n", style="bold", sep="", unindent=1)
+    #oceDebug(debug, gsub(" = [^,)]*", "", deparse(expr=match.call())), " {\n", style="bold", sep="", unindent=1)
 
     oceDebug(debug, "drawPalette(",
              argShow(zlab),
@@ -318,8 +318,12 @@ drawPalette <- function(zlim, zlab="",
              argShow(cex),
              argShow(mai),
              argShow(breaks),
+             argShow(at),
+             argShow(labels),
              argShow(fullpage),
              "...) {\n", sep="", unindent=1, style="bold")
+    atGiven <- !is.null(at)
+    labelsGiven <- !missing(labels)
     zlimGiven <- !missing(zlim)
     if (zlimGiven) {
         if (!is.numeric(zlim))
@@ -384,7 +388,7 @@ drawPalette <- function(zlim, zlab="",
             if (breaksGiven) {
                 breaksOrig <- breaks
                 contours <- breaks
-                oceDebug(debug, "user gave", length(breaks),"breaks\n")
+                oceDebug(debug, "user gave ", length(breaks)," breaks\n")
             } else {
                 zrange <- zlim
                 if (missing(col)) {
@@ -447,7 +451,7 @@ drawPalette <- function(zlim, zlab="",
                 stop("pos must be 1, 2, 3 or 4") # cannot be reached
             }
         } else {
-            oceDebug(debug, "'breaks' was given, starting with", vectorShow(breaks), "\n")
+            oceDebug(debug, "'breaks' was given, starting with ", vectorShow(breaks))
             palette <- seq(zlim[1], zlim[2], length.out=PLEN)
             oceDebug(debug, "drawing palette image, with", vectorShow(par('mai')))
             oceDebug(debug, "palette image width =",
@@ -527,11 +531,14 @@ drawPalette <- function(zlim, zlab="",
             if (missing(axisPalette)) {
                 ## NB. in next line, the '10' matches a call to pretty() in imagep().
                 at <- if (!is.null(contours) & is.null(at)) prettyLocal(contours, 10) else prettyLocal(palette, 10)
+                oceDebug(debug, "with no axisPalette, created 'at' as c(", paste(at,collapse=", "), ")\n")
             } else {
                 ## Guess that the axis labels may need more space
                 at <- if (!is.null(contours) & is.null(at)) prettyLocal(contours, 6) else prettyLocal(palette, 6)
[…]
