+++ sandbox/dk/draw_palette/Makefile
@@ -1,0 +1,5 @@
+test_drawPalette.out: test_drawPalette.R
+	Rscript $< > $@
+clean: force
+	-rm -rf *.out *.pdf *~
+force:
