### 2022-02-23 529cb3b
try to infer zlim from 'breaks' or 'at', if given
We check 'breaks' first, and then 'at'.  An error message results if
none of 'zlim', 'breaks' and 'at' is given.
