+++ R/imagep.R
@@ -1,4 +1,4 @@
-## vim: tw=120 shiftwidth=4 softtabstop=4 expandtab:
+## vim: tw=80 shiftwidth=4 softtabstop=4 expandtab:
 
 PLEN <- 300 # palette has this many default levels
 
@@ -99,8 +99,8 @@ paletteCalculations <- function(separation=par('cin')[2]/2,
     ## } 2014-04-02
 
     lineHeight <- par("cin")[2]  # character height in inches
-    oceDebug(debug, "lineHeight:", lineHeight, "from cin\n")
-    oceDebug(debug, "par('csi'):", par('csi'), "\n")
+    oceDebug(debug, "lineHeight: ", lineHeight, " from cin\n")
+    oceDebug(debug, "par('csi'): ", par('csi'), "\n")
     tickSpace <- abs(par("tcl")) * lineHeight # inches (not sure on this)
     textSpace <- 1.25 * (lineHeight + if (haveZlab) lineHeight else 0)
     figureWidth <- par("fin")[1]
@@ -324,6 +324,7 @@ drawPalette <- function(zlim, zlab="",
              "...) {\n", sep="", unindent=1, style="bold")
     atGiven <- !is.null(at)
     labelsGiven <- !missing(labels)
+    breaksGiven <- !missing(breaks)
     zlimGiven <- !missing(zlim)
     if (zlimGiven) {
         if (!is.numeric(zlim))
@@ -335,17 +336,30 @@ drawPalette <- function(zlim, zlab="",
         if (zlim[2] < zlim[1])
             stop("zlim must be ordered")
     }
+    # If 'zlim' was not given, try to infer it from 'breaks' or from 'at' if
+    # 'breaks' was not given.  If none of the three is given, we cannot draw the
+    # palette.
+    if (!zlimGiven) {
+        if (breaksGiven) {
+            zlim <- range(breaks, na.rm=TRUE)
+            oceDebug(debug, "setting 'zlim' to c(", paste(zlim, collapse=", "), ") based on 'breaks'\n")
+            zlimGiven <- TRUE # fake it
+        } else if (atGiven) {
+            zlim <- range(at, na.rm=TRUE)
+            oceDebug(debug, "setting 'zlim' to c(", paste(zlim, collapse=", "), ") based on 'at'\n")
+            zlimGiven <- TRUE # fake it
+        }
+    }
     colGiven <- !missing(col)
     oceDebug(debug, "colGiven=", colGiven, "\n", sep="")
     colormapGiven <- !missing(colormap)
     oceDebug(debug, "colormapGiven=", colormapGiven, "\n", sep="")
     ##message("missing(col) ", missing(col))
-    if (!zlimGiven && !colormapGiven)
-        plot <- FALSE
+    if (!zlimGiven && !colormapGiven && !breaksGiven && !atGiven)
+        stop("No scale can be computed without zlim, breaks, or colormap")
     levelsGiven <- !missing(levels)
     if (zlimGiven)
         zlim <- range(zlim, na.rm=TRUE)
-    breaksGiven <- !missing(breaks)
     pos <- as.integer(pos)
     if (!(pos %in% 1:4))
         stop("'pos' must be 1, 2, 3 or 4")
@@ -453,10 +467,10 @@ drawPalette <- function(zlim, zlab="",
         } else {
             oceDebug(debug, "'breaks' was given, starting with ", vectorShow(breaks))
             palette <- seq(zlim[1], zlim[2], length.out=PLEN)
-            oceDebug(debug, "drawing palette image, with", vectorShow(par('mai')))
-            oceDebug(debug, "palette image width =",
+            oceDebug(debug, "drawing palette image, with ", vectorShow(par('mai')))
+            oceDebug(debug, "palette image width: ",
[…]
