### 2022-02-23 25aedb0
undo half-hour-old change (do not stop if no zlim)
