### 2022-02-26 adab4b0
demo drawDirectionField
