### 2022-02-28 #3594 @danielskatz
Commands need to start a comment, and the bot name has changed...
