### 2022-03-01 #3594 @patrickcgray
Thanks @danielskatz!

Okay I've basically finished up my run through. I have to admit I'm not super familiar with the oceanography tool landscape in R as a python user primarily but this package seems to cleanly solve many needs in oceanographic analysis, has a robust open-source community surrounding it, and the documentation seems to be up to date and comprehensive. I'm quite impressed by the ability to handle so many common datastreams gracefully. The companion textbook is an excellent resource and will serve many beginning users well.

I installed the develop branch of oce from git in a clean Windows 64 R Studio install running R 4.1.2 and it functioned properly and ran all examples. The only bit not stated in the install instructions is that in order to run the `remotes` install line from a clean install you need to install `Rtools` and then `remotes`. Probably not relevant for most users since they'll have done this, so not sure this is useful information to include. I didn't benchmark it and it might've been my system (which had a lot of analysis products currently in memory) but it took a long time to install, maybe 15 minutes?

I plotted some of my own recent CTD which worked quite well. I did notice that in the plot function when specifying the argument `which=13` it threw an error, but `which=spice` did work in contrast to the documentation. It looks correct here https://github.com/dankelley/oce/blob/adab4b0732d1d33949bd66285678f7c6c1fb2fc6/R/ctd.R#L3300 but maybe elsewhere the cases are mixed up? not sure what's up. Happy to file a bug report if needed. Anyway quite minor and otherwise things worked very well and I explored a couple datasets.

The only other bit that may be worth adding to the paper is that I'm not familiar with other tools, this is one that I am excited to dabble with more even though I'm primarily a python user, but it would be helpful in the paper to add a sentence about other tools attempting to fill a similar gap, or if there aren't other similar packages just a quick line about that and what they do instead.

Overall seems like an excellent package, functions without major issue, and is extremely well documented.
