### 2022-03-01 #3594 @dankelley
@patrickcgray - thanks for the detailed and very quickly-done review.  I'll look into spots (likely in the main README) where I can mention other tools.  Most R users won't even see this website, actually, since the normal case is just to type `install.packages("oce")`, to get the version on CRAN, the Comprehensive R Archive Network.  That's fast, taking perhaps 20 seconds or so, but of CRAN policies dictate against updating more than about twice per year (because all previous versions are permanently available).

If you don't mind, it would help a lot if you could post an issue for that `which=13` bug.  Just a sentence would be enough.  We like to have reports even for things that will involve just a line of code change, because users often search reports to see whether they have encountered a problem that's already been addressed in the "develop" branch.

Again, many thanks for the review.

Dan.
