### 2022-03-01 #3594 @patrickcgray
@dankelley that sounds good! If you'll just tag me when you've added to the readme/paper and I'll give it another quick review and can finalize or add more feedback based on that.
