### 2022-03-01 81f4817
Things in response to JOSS review 2.
In response to the helpful and well-informed review by @patrickcgray, we have
updated the "develop" branch of oce, as well as our manuscript. An overview of
the changes is as follows.

1. `oce` now handles the `plot(ctd,which=13)` case properly. Inspired by this
   report, and to avoid other `plot(...,which)` problems, we have expanded the
   test suite (run by `R CMD check`, or by clicking an icon in Rstudio) to
   ensure that `plot()` handles all the documented textual and numeric `which`
   values for the `ctd` class. (We have also added a check that thi
[…]
