### 2022-03-01 2fb3daf
thank editor and reviewers of JOSS manuscript
