+++ inst/joss/paper.md
@@ -192,7 +192,9 @@ Our other goal is to encourage the development of new R packages, such as
 
 We thank the many users who have helped with `oce` development over the years,
 by finding bugs, suggesting new features, and providing us with manufacturers'
-documentation about data formats for various oceanographic instruments.
+documentation about data formats for various oceanographic instruments. In
+addition, we thank the editor and reviewers of this manuscript, for their
+patience during the submission process.
 
 # References
 
