### 2022-03-01 #3594 @dankelley
@patrickcgray We've updated the manuscript, and also fixed that bug about plot(ctd,which=12) that you so astutely found.  More notes are below.

PS. please look at the second article proof.  In the first one, we had forgotten to thank you, the other reviewer, and the editor.  The bot seems to take just a few minutes to create the article proof so likely there will be a note to that effect, by the time I'm finished typing this.

In response to the helpful and well-informed review by @patrickcgray, we have
updated the "develop" branch of oce, as well as our manuscript. An overview of
the changes is as follows.

1. `oce` now handles the `plot(ctd,which=13)` case properly. Inspired by this
   report, and to avoid other `plot(...,which)` problems, we have expanded the
   test suite (run by `R CMD check`, or by clicking an icon in Rstudio) to
   ensure that `plot()` handles all the documented textual and numeric `which`
   values for the `ctd` class. (We have also added a check that this function
   reports a error for attempts to plot something that is neither contained in
   the object nor computable from the contents.)
2. The newly added Footnote 1 on page 1 of the manuscript now addresses the
   status of other tools that are analogous to oce. In a nutshell, we are not
   aware of any single tool (in any language) that covers the ground that oce
   covers. Even so, we hope that the link now given in this Footnote will be
   a useful starting point for Python users.
3. The newly added Footnote 3 on page 1 of the manuscript now touches on the
   issue of setting up a system to build oce. We do this by providing a link to
   an oce wiki page that details the steps required for one particular OS type.
   Please note that the procedure is not particular to oce; any R package that
   uses Fortran, C or C++ will require that the user install compilers for
   those languages. There are online resources to help with this work in
   general, and we hope that our discussion of one particular case will be
   a useful addition to these.  The good thing is that setting up a system to
   build one source-based R package will usually cover other packages as well,
   so this work is not something that has to be repeated unless there are major
   system changes.
