### 2022-03-01 #3594 @dankelley
@patrickcgray -- the bot supposed built up the PDF of the manuscript, but the links it provided do not work for me.  They might work for you, though.  I don't know how to help with this bot work, unfortunately.
