### 2022-03-01 ff2c118
change some AGU links to DOI links.
The links seem to work one moment, and fail the next.  But I've been
given the impression lately that \doi{} is more forgiving than \url{}.
