### 2022-03-02 #3594 @kthyng
@dankelley We're ready to wrap up! I read through your paper and didn't see any changes to make. After a couple more steps, I plan to accept this as finished.

1. please confirm the version of the software that you would like associated with your JOSS paper — feel free to make a new release if you want to.
2. Archive the current snapshot of your software at a place like Zenodo, then report back here the DOI. Please update the metadata to match the title and author list of your JOSS paper for consistency.
