### 2022-03-03 #3594 @dankelley
@kthyng Thanks.  In response to your comment:
1. The version is 1.6-1, that is, the most recent CRAN version as of this date.  We do not refer to the version in the text.  Did you want me to do that?
2. I have archived version 1.6-1 on zenodo (the link is https://zenodo.org/record/6325185#.YiCc1N9E3tc) i.e. the DOI is 10.5281/zenodo.6325185. I did not understand your instruction to 'update the metadata'.  Is that something that is separate from the paper itself?

As you can see, I do not quite understand what my next steps are.  I apologize, if those steps relate to the procedures I used when initially submitting the text.   Maybe there's a URL I should be reading, to avoid bothering you?

I apologize for being a big dense here.  Dan.
