### 2022-03-03 #3594 @kthyng
@dankelley Sorry I was confusing in my statements! I'll try to clarify on these steps which are always at the end of the review.

1. At the top of this issue you can see there is a version associated with your software — often this changes as part of the JOSS review as changes are made in the software though in this case it maybe have just progressed since so much time has passed since the review began. I'll update the version from `1.5-0` to `1.6-1` on my end. I now realize I'm not sure where this is presented but it is meant to indicate which version of the software exactly was reviewed in the JOSS paper.
2. I see your zenodo archive and it looks good. The metadata I mentioned is in fact for the Zenodo archive — we have a small preference for the title and author list in the Zenodo archive matching the JOSS paper for consistency between the two products. In your case, the title and author list almost but don't exactly match. If you prefer them as they are, you can leave them, otherwise you can go into Zenodo and your archive itself and update the title and author list there to match.

Please let me know if you have more questions!
