### 2022-03-03 #3594 @dankelley
It's unhappy about the formats of my issn fields in my .bib file.  I'll look at the error log and my .bib file and make corrections.  Is there any way I can make the system do further tests, without bothering you?  (This might take a few tries.)
