### 2022-03-03 #3594 @kthyng
You can compile separately from this issue (see docs [here](https://joss.readthedocs.io/en/latest/submitting.html#checking-that-your-paper-compiles)) but also feel free to just do it here — I'll ignore the messages for a bit :)
