### 2022-03-03 #3594 @arfon
> @dankelley There have been some changes in the tech at JOSS just recently and I see some evidence of that here! Can you look through the error messages and see if you can determine the problem? If it still won't work, I will try and/or get help from someone who knows more about it.

It looks like the Crossref validation is failing on this ISSN: https://github.com/openjournals/joss-papers/pull/3010/commits/b08033650d5451dce4a64f8b2db92fd61b66117e#diff-0cf1feada8e37b446552b475411a1595fb35c56e42863f627b956f39902c338bR118
