### 2022-03-03 #3594 @dankelley
I think I see two problems in my .bib file
*   The value 1939-5108, 1939-0068 is not accepted by the pattern d{4}-?d{3}[dX].
*  The value 2503636306 is not accepted by the pattern d{4}-?d{3}[dX].

and will fix them once I do some searching of the sources.  (If I cannot find proper values, I'll delete the fields.)
