### 2022-03-03 #3594 @dankelley
@kthyng actually, the `docker` method fails for me (maybe something to do with the fact I'm on osx).  Maybe you can try `@editorialbot recommend-accept` again, since it (sensibly) won't let me do that.   (I don't think these checks are done when I do `@editorialbot generate pdf`, based on the fact that I hadn't seen these errors before).
