### 2022-03-03 #3594 @kthyng
@dankelley yes apparently things that are warnings when generating the PDFs regularly become errors at this stage.
